Objective function

calculates likelihood function. Used to assess convergence of fitting algorithm. This corresponds to the Q(theta) function in the paper

Q_theta(R, nobs, lambda, alpha, we, wj, wje, betaE, theta_list, gamma)

Arguments

R
residual

nobs
number of observations

lambda
a user supplied lambda sequence. Typically, by leaving this option unspecified users can have the program compute its own lambda sequence based on nlambda and lambda.factor. Supplying a value of lambda overrides this. It is better to supply a decreasing sequence of lambda values than a single (small) value, if not, the program will sort user-defined lambda sequence in decreasing order automatically. Default: NULL.

alpha
the mixing tuning parameter, with \(0<\alpha<1\). It controls the penalization strength between the main effects and the interactions. The penalty is defined as $$\lambda(1-\alpha)(w_e|\beta_e|+ \sum w_j ||\beta_j||_2) + \lambda\alpha(\sum w_{je} |\gamma_j|)$$Larger values of alpha will favor selection of main effects over interactions. Smaller values of alpha will allow more interactions to enter the final model. Default: 0.5

we
penalty factor for exposure variable

wj
penalty factor for main effects

wje
penalty factor for interactions

betaE
estimate of exposure effect

theta_list
estimates of main effects

gamma
estimates of gamma parameter

Value

value of the objective function