OASIS Brain Data
Source: R/data.R

A dataset containing a small subset of the OASIS Brain Data project. Noise variables have been added to increase the number of predictors and show the utility of the sail package.

oasis

Format

A list with 3 elements:

x

a matrix of dimension 136 x 30 with the following colums:

EDUC

Education

MMSE

Mini-Mental State Exam

eTIV

Estimated Total Intracranial Volume

nWBV

Normalized Whole Brain Volume

ASF

Atlas scaling factor

noise[1-24]

24 independent standard normal noise variables

y

a numeric vector of length 136 representing the right Hippocampal volume for each patient

e

a binary 0/1 vector of length 136, representing Dementia status. 0: Non-demented, 1: Demented

Source

https://github.com/stnava/RMI/tree/master/tomfletcher

http://www.oasis-brains.org/

Examples

oasis

#> $x
#>     Age EDUC MMSE eTIV  nWBV   ASF       noise1      noise2      noise3
#> 1    87   14   27 1987 0.696 0.883  1.659721103 -0.80300723  0.19820238
#> 3    75   12   23 1678 0.736 1.046 -0.429188258  0.28505948  0.86951506
#> 6    88   18   28 1215 0.710 1.444 -0.328186864 -0.35685894  0.17189256
#> 8    80   12   28 1689 0.712 1.039 -0.269670911 -1.38667014  2.05062705
#> 11   71   16   28 1357 0.748 1.293 -1.602176254  1.95649725  0.57158276
#> 14   93   14   30 1272 0.698 1.380  1.315575035 -0.99226448 -0.32321572
#> 16   68   12   27 1457 0.806 1.205 -0.086499035 -2.12932051 -0.04040397
#> 18   66   12   30 1447 0.769 1.213 -1.097944174  1.77390766  0.26549403
#> 20   78   16   29 1333 0.748 1.316  0.274700580 -1.60375435  0.21905683
#> 23   81   12   30 1230 0.715 1.427 -0.736763184  0.41367662  1.09736151
#> 26   76   16   21 1602 0.697 1.096  0.232147861 -0.49093239  0.32894005
#> 28   88    8   25 1651 0.660 1.063  0.299104357  0.53725782  0.17715911
#> 30   80   12   29 1783 0.752 0.985  0.161733526 -1.54676218 -0.35395967
#> 40   72   20   26 1911 0.719 0.919 -1.402983126  0.99544931  0.69759443
#> 42   61   16   30 1313 0.805 1.337 -0.002795899 -0.68886634  0.89567641
#> 44   86   12   21 1247 0.662 1.407  1.130584855  1.53113924 -0.47930859
#> 46   82   12   27 1420 0.713 1.236  1.657761487 -0.29605353 -0.42533284
#> 48   69   12   29 1365 0.783 1.286  0.163694443  0.80204079 -1.36981971
#> 52   64   18   22 1547 0.737 1.134 -1.351138415 -1.45368769 -1.78147206
#> 54   77   12   29 1377 0.734 1.275  0.743324183 -0.85951267  0.92159646
#> 56   60   18   30 1402 0.822 1.252 -0.234398808 -0.51162919 -1.05582729
#> 61   90   12   21 1307 0.679 1.342 -1.412092756 -1.52916222 -0.79751412
#> 63   79   16   29 1466 0.703 1.197  1.758745573  0.73662732 -0.06004689
#> 67   88   12   30 1336 0.738 1.313  2.006764724 -0.42410486 -0.40761329
#> 69   69   13   30 1359 0.789 1.291 -0.670662310 -0.40317842  1.07502532
#> 73   82   12   27 1477 0.729 1.188  0.160494634 -0.77680285 -0.72707412
#> 77   81   18   26 1174 0.742 1.495  0.979964121 -2.02671776 -0.31519972
#> 79   84    6   25 1310 0.727 1.339  0.226687099  0.58580483  2.44379837
#> 85   70   17   29 1640 0.766 1.070 -0.851673607  0.86805016  0.08679848
#> 87   72   12   26 1453 0.777 1.208  0.276217486  0.29786703 -0.37145393
#> 89   68   14   21 1333 0.685 1.317  0.275274942 -1.55001623 -0.48113003
#> 92   75   18   30 1317 0.737 1.332 -0.832706752  0.37754365 -0.28133264
#> 94   83   15   20 1476 0.750 1.189 -1.383037263  0.07683758 -0.73185712
#>            noise4      noise5      noise6       noise7       noise8      noise9
#> 1   -1.643948e-01 -0.47888129 -0.57397013 -1.257505219  0.446462842 -0.20482942
#> 3    7.538452e-01 -0.13120827  0.82618664  1.289151724  2.141352426  1.77339406
#> 6    2.944371e-01 -0.45905835  0.67353305 -0.404978549 -0.085534492  0.61598494
#> 8    1.378013e-01  0.16270516  0.19111560  0.046912489 -2.119964982  0.17023909
#> 11  -7.924753e-01 -0.67575400 -1.85146383  0.577407955  0.292067184 -0.29420729
#> 14  -4.527023e-01 -0.73205498 -0.03908070 -0.562208117  0.621890081 -0.09775946
#> 16  -2.395789e-01 -0.31996361 -1.01936303  0.119224432 -0.441372128  1.52663495
#> 18  -9.831256e-01 -1.15519743  1.09768436 -0.809542262 -0.751625465  0.34084498
#> 20  -5.671716e-03 -0.29588956  0.88650670  0.600706233 -0.302078120  0.26779255
#> 23  -8.965084e-01  0.84345247  0.75988418  1.383775757  1.840941794 -0.12336360
#> 26  -2.426388e-01 -1.67163847 -0.98025265  1.145856527 -0.857533795 -1.24036335
#> 28  -1.332207e+00  2.00322645 -2.50001171 -0.377743204 -0.815532742 -1.71531161
#> 30  -5.722194e-01 -0.57100075  0.06032381  0.157795821  0.770578343  0.75047183
#> 40   9.711795e-01  0.53541136  0.41525248 -0.552619872  0.905852670  0.72456731
#> 42  -1.859243e+00 -0.47733630  0.24458129 -0.912154322  1.993150800  0.17487200
#> 44   8.124252e-01 -0.93276551 -0.39111120 -0.072036217  1.198339972 -1.33368038
#> 46   6.595638e-01 -0.90671516  0.10919314 -0.560519419  0.026652458 -0.43060353
#> 48  -5.623774e-01 -2.71307223 -1.09212678  0.100194195 -0.826864513 -0.74769932
#> 52  -6.161201e-01  0.33472581  0.15627031  1.009830507  1.759893254 -0.44687368
#> 54  -1.093387e+00 -1.17636487 -0.33312354  0.480062046 -1.560067760  0.81397472
#> 56  -1.676798e+00  1.58890682  0.02653110  0.298855429  0.394770116 -1.36579717
#> 61   5.670389e-01  0.61175289 -1.32759082  1.092570304  1.340702045  0.22813295
#> 63   4.426423e-01  0.01086962 -0.09844410 -1.503062786  1.985961013 -0.72609235
#> 67   9.033405e-01  0.36818113  0.43905910 -1.869835220 -1.532304959  0.19685991
#> 69   1.078220e+00 -0.67088509 -0.12897487 -0.006920486  0.132607072  1.51208523
#> 73  -2.775533e-01  0.08189777  1.15864880 -0.868513838  0.023324324 -0.26792175
#> 77  -1.534731e+00  0.60891834 -0.09251995 -0.393581293 -0.447829785  0.89303412
#> 79  -1.007185e+00 -1.09744141 -0.61278832 -0.366034235 -0.147892840 -0.02109460
#> 85   1.846665e+00 -1.11074408 -1.65527245 -1.181848286  0.655585719 -0.52451102
#> 87   4.598973e-01  1.47304659 -0.88896457 -1.362698426  1.025836345  0.70440278
#> 89   1.047147e+00 -1.47398368 -0.27563686 -0.120697773  1.196974085  0.35248102
#> 92  -1.565506e+00  1.95646236 -0.24793620 -2.245493048 -1.106082051 -0.47945838
#> 94   8.108519e-01  0.27630069  0.21370914 -0.696030686  2.093449812  1.46044862
#>          noise10      noise11      noise12      noise13     noise14
#> 1    0.110836941  0.069794765  1.141868017 -0.628950231  1.59549225
#> 3    1.546244999  1.369734352  1.043564287  0.712438298  1.34198727
#> 6   -0.266836595  0.071530294 -0.002871695  0.011248481  0.23311297
#> 8    0.908769413 -0.066696202  0.103219017  1.306491128 -1.42788864
#> 11  -0.291521773  0.967207988 -1.756918370  0.519734284 -0.87211896
#> 14   0.523988192  1.677614419 -0.245058433 -1.981957276  0.58446194
#> 16   0.091112980  0.751333853 -1.292763585  0.240766137  0.80417641
#> 18   0.308028017  0.458894713  1.171088501  1.803097432 -0.66287516
#> 20  -0.338961778 -0.779511319  0.176291599  1.144960351  0.22807951
#> 23   1.400302478  0.523869924 -1.267753547  0.236213371  0.14250649
#> 26   0.091998309  0.701910569 -0.350556952  0.888737579  0.23883594
#> 28   0.365747717 -2.506178454  0.083126924  0.251632073 -0.60299581
#> 30   0.301098963  1.332048763 -0.629836176 -0.937862631  1.21162174
#> 40   0.892942051 -0.055617330  1.047946173 -0.395511413  0.77050384
#> 42   0.534336476  0.427220860 -1.196936680  0.158314634 -0.96260183
#> 44   0.764313712  0.966229986  0.294688358  1.872121606  0.01083961
#> 46   0.451787082 -1.347777040 -2.598610201 -1.178017479 -0.74641105
#> 48  -0.814884145 -0.333664604 -1.717617294 -0.600453445 -0.81791430
#> 52  -0.675740147  0.223509238 -1.644301458 -1.121835542 -0.24880743
#> 54   0.402241772  0.616617435  0.598503160 -0.120278103  1.00445856
#> 56   1.285441477  0.304144104 -0.103973589  1.160073708  0.62512879
#> 61  -0.502661325 -0.045570009 -0.338952078  1.992538590 -0.57069164
#> 63  -0.473624669  0.898979261 -0.312514441  0.237787107 -1.98540563
#> 67  -0.490846066 -1.114349190  2.426673256 -1.054857833 -1.31683488
#> 69   0.015992089  0.427349861  0.133730940 -0.007546664  0.29202159
#> 73   0.449308018  0.578116418  0.051829463 -1.353609195 -0.87944016
#> 77   0.651785678  1.132731481 -0.949794219  1.910832047  0.85592262
#> 79   0.430735128 -0.558183805  0.743987850  0.862818241  0.15511766
#> 85   0.380075803 -0.993274914  1.444477435 -0.072674304  1.03523524
#> 87   1.115973263 -0.904530585 -0.557344174 -0.668841920 -0.32722114
#> 89  -0.570189125 -0.690328296 -0.071939320 -0.023989497 -0.12280744
#> 92   0.395635765  0.601494416  0.620199966  0.668102166  0.78975596
#> 94   0.424286187 -1.233839094 -1.973504335  0.688594697  0.82400068
#>          noise15     noise16      noise17      noise18     noise19      noise20
#> 1   -0.570703615 -1.17056381  1.129461470  0.682116609  0.31032572  1.429982096
#> 3    0.553086573  0.99954557 -1.351246110  1.101574366  0.07188301 -1.049804293
#> 6   -0.645209695  0.65746872 -1.270510949 -1.097235674 -0.31987215  0.378252279
#> 8   -0.808761611 -1.26661457 -1.359607779  0.887061177 -0.06124014  0.329707674
#> 11  -0.812110066  0.24850455  1.057725975 -1.005959850  0.75767564 -0.767050180
#> 14   1.215473190  0.56108570 -0.917608448 -3.108079392 -0.61364929 -1.873793341
#> 16   1.796541921  0.19290155 -1.301371505  0.012554394  1.52426622  2.071637835
#> 18   1.148022183  0.20359625  1.605637557 -1.065045540  0.46802555  0.987454683
#> 20  -0.441435689 -0.05586193  0.645039697 -0.536420549  2.29702540 -1.069218203
#> 23   0.838842102  0.06782346  0.416627622 -1.283790488  0.22787234  0.169165305
#> 26  -0.868935106  0.79097726  0.389199085 -0.583207541  0.01319329  0.141352962
#> 28   2.417692695 -0.42227824  0.061394475  1.154246802 -1.10778401 -1.190852988
#> 30  -1.259120417  0.63853498  0.473049165 -0.269664288 -0.14875025  1.334608566
#> 40   0.123051924 -1.65601873 -0.799722114  0.665232324  0.41784099 -0.653372978
#> 42  -0.444457540 -0.91270481  0.424385705 -2.014794718 -0.43251053 -0.031555238
#> 44  -0.836689789 -1.02297944  1.790372621  0.253932038 -0.04180797 -1.538291348
#> 46   0.768438046 -0.19625546 -0.253394856 -0.641196445  0.68692932  0.389470711
#> 48  -0.618475645  1.17868515  0.222226727  0.079340913  0.10497143  1.914111283
#> 52   1.437916116  0.21619775 -0.860561344 -0.928754207 -0.52800564  0.667501920
#> 54   0.131843112  0.30747754 -0.485895232  0.600266237  0.85394982 -0.809557921
#> 56   0.284812763  0.29781518  0.389305330  0.486861121  1.24889472 -2.278069307
#> 61   0.116968874 -1.10485933  0.534289839  0.685394404 -0.54926605 -1.133651562
#> 63  -1.190794602 -1.34842537  0.279540460 -0.357576047 -0.62976028 -0.527334574
#> 67  -0.069898972  1.29588785 -1.846975875  0.555804063  0.64907020  0.665664694
#> 69  -0.734681730 -0.28301643  1.113171061 -0.050200148  0.33230482 -0.692925640
#> 73  -1.215704994 -0.77045139 -1.242712265 -0.984781605  0.62458527 -2.296949897
#> 77   0.394139487  1.31153162 -0.637352057  0.846168744 -0.75568766 -0.131830335
#> 79  -0.252234166 -0.48652470  0.831181918  1.464865890  0.43236601 -0.773169689
#> 85   1.114841461  1.12599272  0.771224199  0.298811864 -0.52862473  0.638365243
#> 87   0.575861328 -0.13623860 -2.151834070 -0.850988125 -0.20857139  0.950931230
#> 89   0.718138309  0.52891741 -0.795314573 -0.161416404 -1.48025409  0.722499848
#> 92   0.313424194  1.54658024 -1.217562250 -0.169408364 -1.80934636 -1.548219173
#> 94  -0.365883000 -0.34816350 -2.077013458 -0.210683404  0.45626645 -1.182562642
#>          noise21      noise22     noise23     noise24
#> 1   -0.671651890  0.494057245  1.24207389  1.99527361
#> 3   -0.738097925  0.129259922 -0.79877492 -0.49303206
#> 6   -0.238506693  0.044139279 -0.31183205  0.54580510
#> 8    0.859752906  1.415246963  0.51701961 -1.03235289
#> 11   0.714827734  0.433269203 -2.08773586  1.59735424
#> 14   0.438555959  0.308892284 -0.79131692 -0.40996481
#> 16  -0.431420861 -0.361493084  1.78237898 -1.33010515
#> 18  -0.644453202 -0.991180287  0.13268446  0.29092970
#> 20  -0.744880291  1.295647789 -0.78142739 -0.59896410
#> 23  -2.807382126 -0.487490810 -1.89973483 -2.23885436
#> 26   1.313328787  0.082391676 -0.85663046 -1.42941234
#> 28  -0.467317050 -0.565662219  0.84847096 -0.18025489
#> 30   0.871810093  0.677482342 -0.67590533  0.52590750
#> 40  -1.736872575 -0.938735042 -0.89772034  0.28365671
#> 42   0.048032523  0.929491786  0.61171892 -1.07207121
#> 44  -1.374057751 -0.335316329  0.86996320  0.62369692
#> 46  -2.159456238 -1.613533398  0.61129822  0.90107236
#> 48   1.123301972 -0.753998390 -1.34101055 -0.77664559
#> 52   0.836225842 -0.086515355 -2.19214762 -0.89374521
#> 54   0.745488729 -0.503820436  0.32565191  0.53519995
#> 56   0.823333763  1.787385329 -1.86519373  0.08004127
#> 61  -0.387097060 -0.119924815 -2.48654213 -1.02308589
#> 63   0.001666858  0.413756068  0.70605926  0.83226464
#> 67  -0.352921312  1.219064987  0.16404369  0.16613672
#> 69  -2.248476175  0.619594265  0.90859350  0.74698119
#> 73  -0.972549787 -0.325790543 -2.60990508  1.04627985
#> 77   0.178489104 -0.652157781  1.20300767 -1.17685528
#> 79  -1.404799822  1.037034512 -1.27788048  1.11718483
#> 85   0.090049641  0.828993737 -0.14331353 -0.16648553
#> 87   0.228188884  0.427864960 -0.14995991 -1.15003232
#> 89   0.870308610  0.271534717  1.41305347 -1.03716255
#> 92  -1.516309000 -0.863942269 -1.55432206 -0.30744689
#> 94   1.786752676 -1.104073144 -0.36405356 -1.44489902
#>  [ reached getOption("max.print") -- omitted 103 rows ]
#> 
#> $y
#>   [1] 3022.5 2786.2 2746.0 3662.9 2273.5 2623.6 3608.7 3626.5 3175.9 2750.2
#>  [11] 2618.1 2668.3 3661.4 2947.4 3418.9 1811.3 2783.6 3539.7 3714.5 3623.4
#>  [21] 3641.8 2624.7 2718.4 3305.0 3072.8 3223.1 1985.7 2376.2 4590.9 2538.8
#>  [31] 1701.1 3069.0 3602.7 3168.0 4123.0 3842.5 3349.6 2371.1 3020.4 2817.2
#>  [41] 3151.9 3142.3 3291.1 2169.9 3860.6 3883.6 3275.5 2645.1 2883.2 4879.5
#>  [51] 2907.3 2778.3 3070.9 3828.8 2770.4 3547.1 2315.3 4085.2 4061.8 3017.1
#>  [61] 2948.6 3355.7 2897.2 3115.7 4446.8 2282.9 2134.7 1618.2 3850.9 3153.4
#>  [71] 3245.9 3904.1 3236.9 3080.2 4256.5 2424.9 3925.5 3519.6 2648.8 1541.7
#>  [81] 3011.2 1710.2 3035.1 2978.0 1958.9 2228.6 2876.1 2483.4 2947.8 4420.0
#>  [91] 3252.1 1690.0 3231.7 3188.4 3181.2 3471.4 3401.4 2990.4 3470.7 2421.0
#> [101] 3662.8 3299.8 3324.9 3033.8 3751.1 3488.6 3277.1 2153.5 3924.9 2208.6
#> [111] 2304.7 3151.8 3113.4 3013.7 2760.8 3063.1 3163.4 2267.1 3803.9 2042.6
#> [121] 3580.1 3257.6 2854.4 4150.1 2821.7 3788.7 3891.3 4143.1 3810.5 1357.8
#> [131] 1995.2 2504.0 3520.0 2993.0 3326.3 3616.6
#> 
#> $e
#>   [1] 1 0 1 1 0 1 0 0 1 1 0 0 1 0 1 0 0 1 0 1 1 0 1 1 1 0 0 0 1 0 0 1 0 1 0 1 0
#>  [38] 1 1 1 1 1 1 0 0 1 1 0 0 0 1 1 1 1 0 1 0 1 1 1 0 0 0 1 1 0 0 0 1 1 1 1 1 1
#>  [75] 0 0 1 1 0 0 1 0 0 1 0 0 0 0 0 1 1 0 1 1 0 1 1 1 0 1 0 1 0 0 1 1 1 0 1 1 0
#> [112] 1 1 1 0 1 0 0 1 0 0 0 1 1 0 1 0 1 1 0 0 0 1 0 0 1
#>