Plot Method for sail object

Produces a coefficient profile plot of the coefficient paths for a fitted sail object. Both main effects and interactions (if present) are plotted.

# S3 method for sail
plot(x, type = c("both", "main", "interaction"), ...)

Arguments

x
fitted sail object

type
which type of predictors should be plotted. type="both" will plot the solution path for main effects and interactions, type="main" will only plot solution path of main effects (this also includes the exposure variable) and type="interaction" will only plot solution path for interaction effects Default: c("both", "main", "interaction"). Default: type="both".

...
other graphical paramters passed to plot

Value

A plot is produced and nothing is returned

Details

A coefficient profile plot is produced

See also

Examples

if (FALSE) {
if(interactive()){
data("sailsim")
f.basis <- function(i) splines::bs(i, degree = 3)
fit <- sail(x = sailsim$x[,1:10,drop=F], y = sailsim$y, e = sailsim$e,
            basis = f.basis)
plot(fit)
 }
}