Make predictions from a cv.sail object
Source: R/methods.R

This function makes predictions from a cross-validated sail model, using the stored "sail.fit" object, and the optimal value chosen for lambda.

# S3 method for cv.sail
predict(object, newx, newe, s = c("lambda.1se",
  "lambda.min"), ...)

# S3 method for cv.sail
coef(object, s = c("lambda.1se", "lambda.min"), ...)

Arguments

object
fitted cv.sail object

newx
matrix of new values for x at which predictions are to be made. Do not include the intercept (this function takes care of that). Must be a matrix. This argument is not used for type=c("coefficients","nonzero"). This matrix will be passed to design_sail to create the design matrix necessary for predictions. This matrix must have the same number of columns originally supplied to the sail fitting function.

newe
vector of new values for the exposure variable e. This is passed to the design_sail function.

s
Value(s) of the penalty parameter lambda at which predictions are required. Default is the value s="lambda.1se" stored on the CV object. Alternatively s="lambda.min" can be used. If s is numeric, it is taken as the value(s) of lambda to be used.

...
other arguments passed to predict.sail

Value

The object returned depends the ... argument which is passed on to the predict method for sail objects.

Details

This function makes it easier to use the results of cross-validation to make a prediction.

See also

Examples

if(interactive()){
data("sailsim")
f.basis <- function(i) splines::bs(i, degree = 3)
cvfit <- cv.sail(x = sailsim$x[,1:20,drop=F], y = sailsim$y, e = sailsim$e,
                  basis = f.basis, nfolds = 10)
predict(cvfit) # predict at "lambda.1se"
predict(cvfit, s = "lambda.min") # predict at "lambda.min"
predict(cvfit, s = 0.5) # predict at specific value of lambda
predict(cvfit, type = "nonzero") # non-zero coefficients at lambda.1se
 }

#>                        1
#> (Intercept)   5.45327020
#> X1_1         -0.32795686
#> X1_2          0.83586426
#> X1_3          2.40801559
#> X3_1          3.22773055
#> X3_2         -2.74925398
#> X3_3         -2.40840181
#> X4_1         -0.28482193
#> X4_2         -9.06472928
#> X4_3         -1.70457282
#> X8_1          0.32709320
#> X8_2         -0.06081891
#> X8_3         -0.52497808
#> X11_1         0.15836459
#> X11_2        -0.47121929
#> X11_3        -0.85764458
#> E             2.01866186
#> X3_1:E        1.86460172
#> X3_2:E       -1.58819444
#> X3_3:E       -1.39129028
#> X4_1:E       -0.58207855
#> X4_2:E      -18.52520426
#> X4_3:E       -3.48356346