Make predictions from a sail object
Source: R/methods.R

Similar to other predict methods, this functions predicts fitted values, logits, coefficients and more from a fitted sail object.

# S3 method for sail
predict(object, newx, newe, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), ...)

# S3 method for sail
coef(object, s = NULL, ...)

Arguments

object
Fitted sail model object

newx
matrix of new values for x at which predictions are to be made. Do not include the intercept (this function takes care of that). Must be a matrix. This argument is not used for type=c("coefficients","nonzero"). This matrix will be passed to design_sail to create the design matrix necessary for predictions. This matrix must have the same number of columns originally supplied to the sail fitting function.

newe
vector of new values for the exposure variable e. This is passed to the design_sail function.

s
Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire sequence used to create the model.

type
Type of prediction required. Type "link" gives the linear predictors for "binomial" (not implemented yet); for "gaussian" models it gives the fitted values. Type "response" gives the fitted probabilities for "binomial" (not implemented yet), for "gaussian" type "response" is equivalent to type "link". Type "coefficients" computes the coefficients at the requested values for s. Note that for "binomial" models, results are returned only for the class corresponding to the second level of the factor response (not implemented yet). Type "class" applies only to "binomial" models, and produces the class label corresponding to the maximum probability (not implemented yet). Type "nonzero" returns a list of the the nonzero coefficients for each value of s. Default: "link"

...
currently ignored

Value

The object returned depends on type.

Details

R Source code file for predict, coef, plot and print methods for the sail package Author: Sahir Bhatnagar Created: 2016 Updated: April 6, 2018

s is the new vector at which predictions are requested. If s is not in the lambda sequence used for fitting the model, the predict function will use linear interpolation to make predictions. The new values are interpolated using a fraction of predicted values from both left and right lambda indices. coef(...) is equivalent to predict(sail.object, type="coefficients",...)

See also

Examples

f.basis <- function(i) splines::bs(i, degree = 5)
fit <- sail(x = sailsim$x, y = sailsim$y, e = sailsim$e,
            basis = f.basis, dfmax = 10)
predict(fit) # predicted response for whole solution path

#>              s1       s2       s3       s4       s5       s6       s7       s8
#>   [1,] 5.218785 5.201803 5.185049 5.169320 5.154306 5.149931 5.204377 5.282594
#>   [2,] 5.218785 5.207461 5.196289 5.185801 5.175789 5.190517 5.193948 5.144461
#>   [3,] 5.218785 5.244160 5.269194 5.292697 5.315132 5.295644 5.120276 4.907047
#>   [4,] 5.218785 5.256039 5.292792 5.327297 5.360234 5.440834 5.619962 5.780230
#>   [5,] 5.218785 5.207406 5.196180 5.185640 5.175580 5.120285 4.888033 4.608463
#>   [6,] 5.218785 5.224509 5.230157 5.235459 5.240520 5.199148 5.013001 4.803700
#>   [7,] 5.218785 5.250692 5.282171 5.311724 5.339934 5.415259 5.594493 5.758650
#>   [8,] 5.218785 5.209743 5.200822 5.192446 5.184452 5.150848 5.110724 5.113171
#>   [9,] 5.218785 5.248808 5.278428 5.306236 5.332780 5.368862 5.458062 5.564281
#>  [10,] 5.218785 5.208675 5.198702 5.189338 5.180400 5.167279 5.192684 5.253771
#>  [11,] 5.218785 5.222890 5.226940 5.230742 5.234372 5.200024 5.120578 5.077424
#>  [12,] 5.218785 5.228900 5.238880 5.248249 5.257193 5.322951 5.557704 5.812157
#>  [13,] 5.218785 5.231433 5.243911 5.255626 5.266808 5.322983 5.533187 5.772852
#>  [14,] 5.218785 5.203753 5.188923 5.175001 5.161711 5.103654 4.997090 4.939628
#>  [15,] 5.218785 5.247622 5.276071 5.302781 5.328276 5.409655 5.677672 5.978932
#>  [16,] 5.218785 5.202570 5.186573 5.171554 5.157218 5.104981 5.018334 4.981894
#>  [17,] 5.218785 5.253090 5.286935 5.318709 5.349040 5.339933 5.215902 5.074429
#>  [18,] 5.218785 5.197336 5.176175 5.156308 5.137345 5.071699 4.951849 4.881892
#>  [19,] 5.218785 5.216473 5.214193 5.212052 5.210008 5.202088 5.172967 5.135985
#>  [20,] 5.218785 5.210864 5.203050 5.195714 5.188711 5.224860 5.318000 5.371221
#>  [21,] 5.218785 5.232119 5.245274 5.257624 5.269413 5.336095 5.516753 5.677351
#>  [22,] 5.218785 5.208491 5.198334 5.188800 5.179698 5.209755 5.293826 5.345834
#>  [23,] 5.218785 5.209036 5.199417 5.190387 5.181767 5.152084 5.059654 4.964621
#>  [24,] 5.218785 5.219100 5.219411 5.219702 5.219981 5.263675 5.454718 5.674071
#>  [25,] 5.218785 5.239838 5.260609 5.280109 5.298722 5.294748 5.185381 5.034590
#>  [26,] 5.218785 5.222151 5.225471 5.228589 5.231565 5.290209 5.508587 5.741410
#>  [27,] 5.218785 5.210782 5.202886 5.195473 5.188397 5.235982 5.464634 5.727149
#>              s9      s10      s11      s12      s13      s14      s15      s16
#>   [1,] 5.353189 5.416743 5.473834 5.525005 5.570785 5.611622 5.647960 5.680188
#>   [2,] 5.100780 5.062251 5.028254 4.998242 4.971665 4.948178 4.927331 4.908773
#>   [3,] 4.707889 4.521970 4.348429 4.186431 4.035161 3.893820 3.761681 3.638096
#>   [4,] 5.936180 6.087550 6.234130 6.375775 6.512323 6.643880 6.770376 6.891804
#>   [5,] 4.338460 4.078078 3.827311 3.586124 3.354462 3.132105 2.918919 2.714778
#>   [6,] 4.599656 4.401143 4.208372 4.021508 3.840695 3.665887 3.497135 3.334474
#>   [7,] 5.914587 6.062448 6.202419 6.334731 6.459592 6.577399 6.688393 6.792842
#>   [8,] 5.110054 5.102237 5.090544 5.075703 5.058440 5.039207 5.018576 4.997000
#>   [9,] 5.660760 5.748232 5.827428 5.899043 5.963758 6.022166 6.074860 6.122366
#>  [10,] 5.306875 5.352798 5.392328 5.426186 5.455091 5.479579 5.500248 5.517601
#>  [11,] 5.031412 4.983301 4.933789 4.883480 4.832965 4.782549 4.732664 4.683643
#>  [12,] 6.058612 6.296982 6.527234 6.749378 6.963414 7.169576 7.367926 7.558546
#>  [13,] 6.002107 6.221242 6.430590 6.630492 6.821279 7.003399 7.177184 7.342944
#>  [14,] 4.881902 4.824516 4.768012 4.712836 4.659435 4.607965 4.558743 4.511971
#>  [15,] 6.269975 6.550880 6.821782 7.082846 7.334228 7.576289 7.809237 8.033269
#>  [16,] 4.953078 4.931301 4.915961 4.906437 4.902162 4.902516 4.907013 4.915160
#>  [17,] 4.932734 4.791453 4.651160 4.512368 4.375564 4.241007 4.109048 3.979994
#>  [18,] 4.814274 4.749343 4.687389 4.628615 4.573231 4.521194 4.472621 4.427534
#>  [19,] 5.093971 5.047573 4.997417 4.944098 4.888186 4.830129 4.770395 4.709417
#>  [20,] 5.420409 5.465548 5.506660 5.543827 5.577104 5.606740 5.632825 5.655514
#>  [21,] 5.831600 5.979412 6.120752 6.255652 6.384123 6.506433 6.622661 6.732942
#>  [22,] 5.398519 5.451399 5.504033 5.556059 5.607100 5.657036 5.705593 5.752590
#>  [23,] 4.884149 4.816468 4.759847 4.712656 4.673356 4.640630 4.613228 4.590074
#>  [24,] 5.883599 6.083584 6.274350 6.456231 6.629547 6.794728 6.952095 7.101945
#>  [25,] 4.883532 4.732742 4.582710 4.433900 4.286738 4.141501 3.998503 3.858046
#>  [26,] 5.965558 6.181073 6.388055 6.586645 6.776976 6.959396 7.134098 7.301280
#>  [27,] 5.980072 6.223521 6.457672 6.682724 6.898876 7.106506 7.305854 7.497143
#>              s17       s18       s19       s20        s21        s22        s23
#>   [1,]  5.198749  5.322727  5.440045  5.531475  5.5781494  5.6384620  5.6901930
#>   [2,]  3.501649  3.397069  3.303104  3.254042  3.2167312  3.2080322  3.1750985
#>   [3,]  3.557023  3.448360  3.343849  3.139720  2.8709368  2.6836370  2.5146455
#>   [4,] 12.835144 12.924555 13.002752 12.949046 12.9652777 12.9081293 12.9656631
#>   [5,]  3.596400  3.450420  3.303898  3.163248  3.0896458  2.9816179  2.8992619
#>   [6,]  3.085516  2.977242  2.871627  2.745382  2.6160296  2.5083671  2.4085997
#>   [7,] 11.314855 11.462122 11.597469 11.638524 11.7235597 11.7566084 11.8637339
#>   [8,]  5.668513  5.797563  5.915629  6.008497  6.0871354  6.1434303  6.2078103
#>   [9,]  6.836918  6.817749  6.797766  6.836988  6.9235585  6.9584584  6.9881155
#>  [10,]  5.644275  5.774065  5.895453  5.993547  6.0648336  6.1294882  6.1944034
#>  [11,]  4.244790  4.206135  4.166567  4.170678  4.1892840  4.1799212  4.1586199
#>  [12,]  8.998614  9.112722  9.224097  9.355359  9.5204107  9.6419261  9.7725651
#>  [13,]  8.831096  8.981415  9.125211  9.309111  9.5353851  9.7029551  9.8704367
#>  [14,]  5.388103  5.384068  5.374573  5.318681  5.2623705  5.1779335  5.1293983
#>  [15,] 12.566140 12.721037 12.864783 13.059425 13.3933521 13.5786863 13.8111529
#>  [16,]  5.125111  5.154502  5.182072  5.178645  5.1529660  5.1291498  5.1179612
#>  [17,]  1.655912  1.479811  1.313298  1.134932  0.8806431  0.7080484  0.5040244
#>  [18,]  5.782526  5.810727  5.831634  5.781835  5.7258096  5.6455690  5.6115941
#>  [19,]  4.476929  4.423289  4.370937  4.312582  4.2610212  4.2095378  4.1658817
#>  [20,]  4.222395  4.138264  4.062828  4.004546  3.9397800  3.9065866  3.8582021
#>  [21,]  8.513646  8.531922  8.549009  8.509214  8.4826571  8.4453314  8.4479259
#>  [22,]  3.940597  3.860346  3.790632  3.751675  3.6980370  3.6855117  3.6457416
#>  [23,]  4.142784  4.066453  3.994331  3.929237  3.8646065  3.8150201  3.7592953
#>  [24,]  7.317005  7.526619  7.728443  7.941808  8.1313906  8.3125478  8.4720506
#>  [25,]  3.618043  3.608827  3.597089  3.514320  3.3681352  3.2886510  3.2054958
#>  [26,]  7.496046  7.583681  7.672287  7.767612  7.8603552  7.9477680  8.0337448
#>  [27,]  6.063420  6.214676  6.365842  6.528989  6.6306127  6.7645538  6.8621092
#>               s24        s25        s26         s27          s28        s29
#>   [1,]  5.7503553  5.8193022  5.9073594  5.99117827  6.070822149  6.1059749
#>   [2,]  3.1669653  3.1350845  3.1166430  3.10023642  3.086271073  3.0645263
#>   [3,]  2.3758248  2.2304322  2.1172549  2.01011999  1.908461918  1.7733377
#>   [4,] 12.9339871 12.9596749 12.9506092 12.94384017 12.937768777 12.9337193
#>   [5,]  2.7873887  2.7033006  2.6202368  2.53730988  2.454276741  2.4045525
#>   [6,]  2.3160905  2.2117017  2.1110248  2.01530242  1.924395442  1.8344265
#>   [7,] 11.9078072 12.0083089 12.1269958 12.24170107 12.351628582 12.5637377
#>   [8,]  6.2526751  6.2606328  6.2711399  6.28040814  6.288398444  6.2850360
#>   [9,]  6.9961959  6.9947518  6.9724437  6.94941092  6.925927402  6.8697472
#>  [10,]  6.2518111  6.2665376  6.2476530  6.22947517  6.211948913  6.1763591
#>  [11,]  4.1335669  4.1209467  4.1023928  4.08310757  4.063297526  4.0325077
#>  [12,]  9.8787300 10.0131213 10.1602794 10.30157132 10.437075765 10.5841275
#>  [13,] 10.0081312 10.1652214 10.3344248 10.49555488 10.648896076 10.8335459
#>  [14,]  5.0549722  4.9936007  4.9189475  4.84852106  4.781647965  4.7444109
#>  [15,] 13.9468959 14.1115377 14.2814134 14.44233491 14.594221814 14.7516727
#>  [16,]  5.1069603  5.1291596  5.1793034  5.22997425  5.280851761  5.3355269
#>  [17,]  0.3606455  0.2449394  0.1612178  0.08118105  0.004951366 -0.1509227
#>  [18,]  5.5480788  5.4862791  5.4081817  5.33544653  5.267214993  5.2234939
#>  [19,]  4.1195331  4.0615773  3.9726150  3.88798296  3.807517726  3.7329971
#>  [20,]  3.8343823  3.8234608  3.7884370  3.75521378  3.723975417  3.6872602
#>  [21,]  8.4267070  8.4473455  8.4830253  8.51801682  8.551779745  8.5879408
#>  [22,]  3.6416379  3.6629783  3.7028701  3.74191851  3.780475989  3.8090575
#>  [23,]  3.7105350  3.6095355  3.5009220  3.39521968  3.292488217  3.1959983
#>  [24,]  8.6297886  8.7442136  8.8760619  9.00196555  9.122528686  9.2422612
#>  [25,]  3.1544877  3.1289360  3.0841888  3.04062584  2.998222218  3.0133530
#>  [26,]  8.1192621  8.1993811  8.2400341  8.28134097  8.323311393  8.3704682
#>  [27,]  6.9921315  7.1065186  7.2072708  7.30532682  7.401056807  7.4651209
#>               s30        s31        s32         s33         s34        s35
#>   [1,]  6.1238456  6.1301475  6.1165502  6.07333544  6.04652270  6.0030458
#>   [2,]  3.0777571  3.0674863  3.0980656  3.11208146  3.14981842  3.1668924
#>   [3,]  1.6553541  1.5220679  1.4623071  1.39736867  1.36184072  1.3029565
#>   [4,] 12.7842243 12.7694493 12.6493309 12.65368298 12.55153352 12.5606422
#>   [5,]  2.3222455  2.2808607  2.1642397  2.07767216  1.95357571  1.8731787
#>   [6,]  1.7536974  1.6711380  1.6108778  1.55415691  1.50757353  1.4598795
#>   [7,] 12.7227901 12.9368167 13.0150269 13.14762283 13.19491869 13.3141616
#>   [8,]  6.2564468  6.2321199  6.2391769  6.26653386  6.28790704  6.3128598
#>   [9,]  6.7766148  6.7150825  6.6920447  6.70825027  6.70645169  6.7192555
#>  [10,]  6.1297623  6.0785796  6.0295773  5.97339182  5.92766475  5.8763436
#>  [11,]  3.9950975  3.9648286  3.9060736  3.83983002  3.77674870  3.7179295
#>  [12,] 10.7022374 10.8454516 10.9386829 11.04863545 11.12630989 11.2261751
#>  [13,] 10.9947933 11.1709183 11.3069011 11.46101707 11.57813293 11.7146070
#>  [14,]  4.6883219  4.6638751  4.5738041  4.49910274  4.40124827  4.3340197
#>  [15,] 14.7783634 14.9183263 14.9111697 14.99255211 14.97335634 15.0521956
#>  [16,]  5.3874372  5.4402946  5.5229040  5.61847252  5.71050120  5.8007556
#>  [17,] -0.2704241 -0.4311151 -0.5251083 -0.67386783 -0.75370196 -0.8874382
#>  [18,]  5.1527466  5.1147360  5.0853576  5.09821756  5.07877688  5.0838771
#>  [19,]  3.6591274  3.5938202  3.5065971  3.41761437  3.32916391  3.2479498
#>  [20,]  3.6761504  3.6516326  3.6862531  3.72218903  3.76802254  3.7983904
#>  [21,]  8.5804975  8.6164029  8.5884644  8.58904907  8.55478981  8.5564199
#>  [22,]  3.8759428  3.9141406  3.9994618  4.06231303  4.14587359  4.2005423
#>  [23,]  3.1270513  3.0450231  2.9473392  2.82486902  2.72033686  2.6062287
#>  [24,]  9.3700532  9.4677392  9.6015535  9.72474077  9.85596583  9.9667599
#>  [25,]  3.0850401  3.0997057  3.1124752  3.08034028  3.06895234  3.0276400
#>  [26,]  8.4225357  8.4753825  8.5491891  8.63436405  8.71732311  8.8016214
#>  [27,]  7.5633487  7.6149735  7.6745364  7.67391873  7.71692875  7.7153537
#>               s36
#>   [1,]  5.9498492
#>   [2,]  3.1972582
#>   [3,]  1.2511431
#>   [4,] 12.5585634
#>   [5,]  1.7953403
#>   [6,]  1.4302303
#>   [7,] 13.3883126
#>   [8,]  6.3555899
#>   [9,]  6.6938121
#>  [10,]  5.8088474
#>  [11,]  3.6812668
#>  [12,] 11.3280322
#>  [13,] 11.8383918
#>  [14,]  4.2376031
#>  [15,] 15.0723274
#>  [16,]  5.8856629
#>  [17,] -0.9765147
#>  [18,]  5.0556721
#>  [19,]  3.1961960
#>  [20,]  3.8171681
#>  [21,]  8.5356952
#>  [22,]  4.2887330
#>  [23,]  2.4882981
#>  [24,] 10.0866145
#>  [25,]  2.9800825
#>  [26,]  8.8466300
#>  [27,]  7.7604713
#>  [ reached getOption("max.print") -- omitted 73 rows ]

predict(fit, s = 0.45) # predicted response for a single lambda value

#>                 1
#>   [1,]  5.9673300
#>   [2,]  3.1049044
#>   [3,]  2.0406021
#>   [4,] 12.9457661
#>   [5,]  2.5609043
#>   [6,]  2.0425374
#>   [7,] 12.2090651
#>   [8,]  6.2777711
#>   [9,]  6.9559642
#>  [10,]  6.2346471
#>  [11,]  4.0885946
#>  [12,] 10.2613709
#>  [13,] 10.4497100
#>  [14,]  4.8685588
#>  [15,] 14.3965494
#>  [16,]  5.2155573
#>  [17,]  0.1039532
#>  [18,]  5.3561412
#>  [19,]  3.9120625
#>  [20,]  3.7646665
#>  [21,]  8.5080610
#>  [22,]  3.7308084
#>  [23,]  3.4252942
#>  [24,]  8.9661433
#>  [25,]  3.0530204
#>  [26,]  8.2695883
#>  [27,]  7.2774279
#>  [28,]  0.6069351
#>  [29,] 10.9422715
#>  [30,]  3.1215920
#>  [31,]  5.5472522
#>  [32,]  7.2761852
#>  [33,]  7.2225982
#>  [34,]  2.8208385
#>  [35,] 15.3952408
#>  [36,]  3.8305986
#>  [37,]  5.5692761
#>  [38,]  3.2742106
#>  [39,]  4.7834286
#>  [40,]  5.9606372
#>  [41,]  3.6930744
#>  [42,]  1.9975336
#>  [43,]  5.1962678
#>  [44,] 14.1181509
#>  [45,]  2.9703610
#>  [46,]  4.9564300
#>  [47,]  5.5903895
#>  [48,] 14.3625987
#>  [49,]  3.2300613
#>  [50,]  6.6860514
#>  [51,]  5.2866828
#>  [52,]  2.8915051
#>  [53,]  7.2923735
#>  [54,]  7.0052339
#>  [55,]  2.4800233
#>  [56,]  1.6810683
#>  [57,]  4.3981568
#>  [58,]  7.2815994
#>  [59,]  7.2839413
#>  [60,]  3.8163750
#>  [61,]  2.4338552
#>  [62,]  1.7739604
#>  [63,]  8.8411001
#>  [64,]  3.0398590
#>  [65,]  4.1841537
#>  [66,]  4.7443959
#>  [67,]  4.6778632
#>  [68,]  5.8730607
#>  [69,]  7.9891439
#>  [70,]  7.0381020
#>  [71,]  3.9652579
#>  [72,]  2.3176206
#>  [73,]  6.5322168
#>  [74,]  5.0524299
#>  [75,]  3.1110195
#>  [76,]  3.3361716
#>  [77,]  2.9802143
#>  [78,] 10.3236917
#>  [79,]  6.1715220
#>  [80,]  2.3401506
#>  [81,]  4.8713422
#>  [82,]  4.1001745
#>  [83,]  3.5676135
#>  [84,]  5.5466088
#>  [85,]  5.4093081
#>  [86,]  1.9181951
#>  [87,]  3.7973995
#>  [88,]  3.1304777
#>  [89,]  4.3321605
#>  [90,]  4.9917762
#>  [91,]  4.5832944
#>  [92,]  8.9661251
#>  [93,]  5.0035550
#>  [94,]  6.4896429
#>  [95,]  5.6725367
#>  [96,]  3.3864410
#>  [97,]  6.7593511
#>  [98,]  3.7087109
#>  [99,]  4.4293729
#> [100,]  5.2384188

predict(fit, s = c(2.15, 0.32, 0.40), type="nonzero") # nonzero coefficients

#> [[1]]
#>                    1
#> (Intercept) 5.218785
#> 
#> [[2]]
#>                        2
#> (Intercept)   5.43273370
#> X1_1         -1.13416799
#> X1_2          0.35963042
#> X1_3          0.75400582
#> X1_4          1.03258813
#> X1_5          2.48105938
#> X2_1         -0.11829414
#> X2_2         -0.48895335
#> X2_3         -0.38891201
#> X2_4         -0.08634539
#> X2_5         -0.02014874
#> X3_1          4.15082483
#> X3_2          2.06155043
#> X3_3         -1.35860284
#> X3_4         -2.51997086
#> X3_5         -1.12096624
#> X4_1          5.08503756
#> X4_2         -3.23311149
#> X4_3         -7.63198370
#> X4_4         -3.53921628
#> X4_5         -0.51628844
#> X8_1          0.38999053
#> X8_2          0.26635045
#> X8_3          0.05017505
#> X8_4         -0.22972707
#> X8_5         -0.54498817
#> X11_1         0.30715276
#> X11_2         0.05511663
#> X11_3        -0.17773070
#> X11_4        -0.55016268
#> X11_5        -0.56668625
#> E             2.04386954
#> X1_1:E       -0.40553811
#> X1_2:E        0.12863171
#> X1_3:E        0.26963378
#> X1_4:E        0.36920969
#> X1_5:E        0.88712555
#> X3_1:E        2.24125165
#> X3_2:E        1.11314703
#> X3_3:E       -0.73357602
#> X3_4:E       -1.36065903
#> X3_5:E       -0.60525723
#> X4_1:E        8.66144615
#> X4_2:E       -5.50702960
#> X4_3:E      -12.99971148
#> X4_4:E       -6.02844273
#> X4_5:E       -0.87940572
#> 
#> [[3]]
#>                        3
#> (Intercept)   5.42476289
#> X1_1         -0.86999981
#> X1_2          0.17589811
#> X1_3          0.50737406
#> X1_4          0.85766171
#> X1_5          2.06927332
#> X3_1          3.61153102
#> X3_2          1.67474215
#> X3_3         -1.29748645
#> X3_4         -2.34271683
#> X3_5         -1.22519001
#> X4_1          4.58763480
#> X4_2         -2.95668934
#> X4_3         -6.85307834
#> X4_4         -3.47331529
#> X4_5         -0.48292524
#> X8_1          0.22356411
#> X8_2          0.14896473
#> X8_3          0.02374265
#> X8_4         -0.13302099
#> X8_5         -0.31359951
#> X11_1         0.18381572
#> X11_2         0.03622327
#> X11_3        -0.10082960
#> X11_4        -0.28964176
#> X11_5        -0.30994433
#> E             2.01541346
#> X1_1:E       -0.08211345
#> X1_2:E        0.01691943
#> X1_3:E        0.04811668
#> X1_4:E        0.08042830
#> X1_5:E        0.19394096
#> X3_1:E        1.92976124
#> X3_2:E        0.89487007
#> X3_3:E       -0.69329058
#> X3_4:E       -1.25179209
#> X3_5:E       -0.65466064
#> X4_1:E        8.44340927
#> X4_2:E       -5.44170725
#> X4_3:E      -12.61275662
#> X4_4:E       -6.39295065
#> X4_5:E       -0.88884338
#>