Print Method for sail object
Source: R/methods.R

Print a summary of the sail path at each step along the path.

# S3 method for sail
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

x
fitted sail object

digits
significant digits in printout. Default: max(3, getOption("digits") - 3)

...
additional print arguments

Value

OUTPUT_DESCRIPTION

Details

The call that produced the object x is printed, followed by a five-column matrix with columns df_main, df_interaction, df_environment, %Dev and Lambda. The df_ columns are the corresponding number of nonzero coefficients for main effects, interactions and exposure, respectively. %dev is the percent deviance explained (relative to the null deviance). For type="gaussian" this is the r-squared.

See also

Examples

if(interactive()){
data("sailsim")
f.basis <- function(i) splines::bs(i, degree = 3)
fit <- sail(x = sailsim$x, y = sailsim$y, e = sailsim$e,
            basis = f.basis)
fit
 }

#> 
#> Call:  sail(x = sailsim$x, y = sailsim$y, e = sailsim$e, basis = f.basis) 
#> 
#>      df_main df_interaction df_environment     %Dev  Lambda
#> s1         0              0              0 0.000000 1.41800
#> s2         0              0              1 0.001371 1.35400
#> s3         0              0              1 0.002711 1.29200
#> s4         0              0              1 0.003955 1.23400
#> s5         0              0              1 0.005129 1.17800
#> s6         0              0              1 0.006240 1.12400
#> s7         1              0              1 0.009825 1.07300
#> s8         2              0              1 0.029570 1.02400
#> s9         2              0              1 0.054140 0.97760
#> s10        2              0              1 0.077360 0.93320
#> s11        2              0              1 0.099290 0.89080
#> s12        2              0              1 0.120000 0.85030
#> s13        3              0              1 0.141100 0.81170
#> s14        3              0              1 0.162800 0.77480
#> s15        4              0              1 0.184900 0.73950
#> s16        4              0              1 0.206700 0.70590
#> s17        5              0              1 0.228600 0.67380
#> s18        5              0              1 0.249700 0.64320
#> s19        4              1              1 0.400900 0.61400
#> s20        5              1              1 0.416900 0.58610
#> s21        5              2              1 0.432300 0.55940
#> s22        5              2              1 0.447500 0.53400
#> s23        5              2              1 0.461900 0.50970
#> s24        5              2              1 0.472500 0.48660
#> s25        5              2              1 0.482700 0.46450
#> s26        5              2              1 0.494200 0.44330
#> s27        5              2              1 0.502400 0.42320
#> s28        5              2              1 0.511600 0.40400
#> s29        5              2              1 0.518900 0.38560
#> s30        5              2              1 0.527000 0.36810
#> s31        7              2              1 0.534300 0.35130
#> s32        7              2              1 0.543600 0.33540
#> s33        7              2              1 0.551300 0.32010
#> s34        7              3              1 0.559300 0.30560
#> s35        9              3              1 0.567200 0.29170
#> s36        9              4              1 0.578500 0.27840
#> s37        9              4              1 0.590000 0.26580
#> s38        8              4              1 0.631700 0.25370
#> s39        8              5              1 0.638600 0.24220
#> s40        8              5              1 0.646700 0.23120
#> s41        8              5              1 0.653100 0.22070
#> s42        9              5              1 0.659900 0.21060
#> s43        9              5              1 0.665700 0.20110
#> s44        9              7              1 0.687400 0.19190
#> s45        9              7              1 0.692600 0.18320
#> s46        9              7              1 0.697500 0.17490
#> s47       10              7              1 0.702100 0.16690
#> s48       11              7              1 0.707100 0.15930
#> s49       11              7              1 0.712200 0.15210
#> s50       12              7              1 0.717400 0.14520
#> s51       12              9              1 0.737400 0.13860
#> s52       12              9              1 0.742800 0.13230
#> s53       12              9              1 0.748200 0.12630
#> s54       13              9              1 0.757600 0.12050
#> s55       13              9              1 0.761600 0.11500
#> s56       13              9              1 0.765000 0.10980
#> s57       13              9              1 0.768200 0.10480
#> s58       14              9              1 0.771100 0.10010
#> s59       14              9              1 0.774000 0.09552
#> s60       16              9              1 0.777400 0.09117
#> s61       17              9              1 0.780900 0.08703
#> s62       17              9              1 0.784400 0.08308
#> s63       17              9              1 0.787600 0.07930
#> s64       16             13              1 0.815000 0.07569
#> s65       16             13              1 0.817400 0.07225
#> s66       16             13              1 0.819800 0.06897
#> s67       16             13              1 0.822300 0.06584
#> s68       16             14              1 0.831800 0.06284
#> s69       17             14              1 0.834200 0.05999
#> s70       18             14              1 0.836700 0.05726
#> s71       18             14              1 0.839300 0.05466
#> s72       20             17              1 0.893300 0.05217
#> s73       20             17              1 0.896500 0.04980
#> s74       20             19              1 0.906800 0.04754
#> s75       20             19              1 0.908300 0.04538
#> s76       20             19              1 0.910200 0.04332
#> s77       20             19              1 0.912300 0.04135
#> s78       20             19              1 0.914300 0.03947
#> s79       20             19              1 0.916300 0.03767
#> s80       20             19              1 0.918100 0.03596
#> s81       20             19              1 0.919900 0.03433
#> s82       20             19              1 0.921600 0.03277
#> s83       20             20              1 0.923100 0.03128
#> s84       20             20              1 0.924600 0.02986
#> s85       20             20              1 0.926100 0.02850
#> s86       20             20              1 0.927400 0.02720
#> s87       20             20              1 0.928700 0.02597
#> s88       20             20              1 0.929900 0.02479
#> s89       20             20              1 0.930800 0.02366
#> s90       20             20              1 0.932600 0.02258
#> s91       20             20              1 0.933700 0.02156
#> s92       20             20              1 0.934900 0.02058
#> s93       20             20              1 0.936000 0.01964
#> s94       20             20              1 0.936800 0.01875
#> s95       20             20              1 0.937700 0.01790
#> s96       20             20              1 0.938400 0.01708
#> s97       20             20              1 0.939200 0.01631
#> s98       20             20              1 0.939900 0.01557
#> s99       20             20              1 0.940600 0.01486
#> s100      20             20              1 0.941200 0.01418