OASIS Brain Data

A dataset containing a small subset of the OASIS Brain Data project. Noise variables have been added to increase the number of predictors and show the utility of the sail package.

oasis

Format

A list with 3 elements:

x

a matrix of dimension 136 x 30 with the following colums:

EDUC

Education

MMSE

Mini-Mental State Exam

eTIV

Estimated Total Intracranial Volume

nWBV

Normalized Whole Brain Volume

ASF

Atlas scaling factor

noise[1-24]

24 independent standard normal noise variables

y

a numeric vector of length 136 representing the right Hippocampal volume for each patient

e

a binary 0/1 vector of length 136, representing Dementia status. 0: Non-demented, 1: Demented

Source

https://github.com/stnava/RMI/tree/master/tomfletcher

http://www.oasis-brains.org/

Examples

oasis

#> $x
#>     Age EDUC MMSE eTIV  nWBV   ASF       noise1      noise2      noise3
#> 1    87   14   27 1987 0.696 0.883  1.659721103 -0.80300723  0.19820238
#> 3    75   12   23 1678 0.736 1.046 -0.429188258  0.28505948  0.86951506
#> 6    88   18   28 1215 0.710 1.444 -0.328186864 -0.35685894  0.17189256
#> 8    80   12   28 1689 0.712 1.039 -0.269670911 -1.38667014  2.05062705
#> 11   71   16   28 1357 0.748 1.293 -1.602176254  1.95649725  0.57158276
#> 14   93   14   30 1272 0.698 1.380  1.315575035 -0.99226448 -0.32321572
#> 16   68   12   27 1457 0.806 1.205 -0.086499035 -2.12932051 -0.04040397
#> 18   66   12   30 1447 0.769 1.213 -1.097944174  1.77390766  0.26549403
#> 20   78   16   29 1333 0.748 1.316  0.274700580 -1.60375435  0.21905683
#> 23   81   12   30 1230 0.715 1.427 -0.736763184  0.41367662  1.09736151
#> 26   76   16   21 1602 0.697 1.096  0.232147861 -0.49093239  0.32894005
#> 28   88    8   25 1651 0.660 1.063  0.299104357  0.53725782  0.17715911
#> 30   80   12   29 1783 0.752 0.985  0.161733526 -1.54676218 -0.35395967
#> 40   72   20   26 1911 0.719 0.919 -1.402983126  0.99544931  0.69759443
#> 42   61   16   30 1313 0.805 1.337 -0.002795899 -0.68886634  0.89567641
#> 44   86   12   21 1247 0.662 1.407  1.130584855  1.53113924 -0.47930859
#> 46   82   12   27 1420 0.713 1.236  1.657761487 -0.29605353 -0.42533284
#> 48   69   12   29 1365 0.783 1.286  0.163694443  0.80204079 -1.36981971
#> 52   64   18   22 1547 0.737 1.134 -1.351138415 -1.45368769 -1.78147206
#> 54   77   12   29 1377 0.734 1.275  0.743324183 -0.85951267  0.92159646
#> 56   60   18   30 1402 0.822 1.252 -0.234398808 -0.51162919 -1.05582729
#> 61   90   12   21 1307 0.679 1.342 -1.412092756 -1.52916222 -0.79751412
#> 63   79   16   29 1466 0.703 1.197  1.758745573  0.73662732 -0.06004689
#> 67   88   12   30 1336 0.738 1.313  2.006764724 -0.42410486 -0.40761329
#> 69   69   13   30 1359 0.789 1.291 -0.670662310 -0.40317842  1.07502532
#> 73   82   12   27 1477 0.729 1.188  0.160494634 -0.77680285 -0.72707412
#> 77   81   18   26 1174 0.742 1.495  0.979964121 -2.02671776 -0.31519972
#> 79   84    6   25 1310 0.727 1.339  0.226687099  0.58580483  2.44379837
#> 85   70   17   29 1640 0.766 1.070 -0.851673607  0.86805016  0.08679848
#> 87   72   12   26 1453 0.777 1.208  0.276217486  0.29786703 -0.37145393
#> 89   68   14   21 1333 0.685 1.317  0.275274942 -1.55001623 -0.48113003
#> 92   75   18   30 1317 0.737 1.332 -0.832706752  0.37754365 -0.28133264
#> 94   83   15   20 1476 0.750 1.189 -1.383037263  0.07683758 -0.73185712
#> 96   77   16   29 1433 0.723 1.225 -0.628327790  0.61214993 -1.21771497
#> 98   66   16   19 1695 0.711 1.036 -2.187398045  1.40591059 -0.97551436
#> 103  69   16   30 1491 0.794 1.177 -1.292937428 -1.61480702 -0.28240664
#> 106  71   12   20 1461 0.724 1.202  1.097231419  0.86110243 -2.37062078
#> 108  92   23   29 1454 0.701 1.207 -0.134718676  0.49397851  1.28912056
#> 111  74   18   29 1463 0.737 1.199  1.368475721  1.44846537  0.09830724
#> 113  82   16   29 1484 0.760 1.183  0.365978076  1.97968273 -0.95385709
#> 117  65   13   29 1362 0.837 1.289  0.411091110  0.44311764 -0.82548503
#> 119  71   14   28 1461 0.756 1.202  0.032809775 -0.65514662  0.72476858
#> 121  81   12   30 1599 0.755 1.098 -1.616143884  0.46579891 -1.43173841
#> 124  78   14   30 1315 0.707 1.335 -0.763100466 -1.07864057 -0.57849855
#> 127  75   13   29 1416 0.766 1.239 -1.551619473  0.78876527  0.80455856
#> 129  68   18   30 1654 0.747 1.061 -1.916510710  0.31550755 -0.29119824
#> 132  79   18   29 1641 0.695 1.069  1.206585519 -0.78481501 -0.46707426
#> 135  80   12   30 1430 0.737 1.228  1.206920957 -0.81318556  1.22899396
#> 137  78    8   23 1462 0.691 1.200  1.322245290 -0.75434030 -0.11072700
#> 140  61   18   30 1957 0.734 0.897  0.484401023 -0.90190631  1.07528928
#> 142  67   12   30 1440 0.727 1.219 -0.409209864 -0.42661856 -0.66039520
#> 146  88   12   30 1428 0.700 1.229 -1.210731112  0.24024169  0.82676590
#> 148  81   18   29 1470 0.687 1.194 -1.503439810 -0.27643535 -1.02491833
#> 150  80   17   28 1660 0.728 1.057  1.202734125 -0.80118466  1.31817954
#> 155  83   13   27 1391 0.705 1.262 -0.592785873  0.94058655 -0.89495944
#> 157  70   14   29 1524 0.787 1.151  0.544782558 -0.47451122 -0.89887218
#> 162  73    8   25 1151 0.743 1.525 -0.452540212 -0.05886764  1.44745925
#> 164  66   18   30 1504 0.725 1.167 -0.249368935 -1.15013632  1.72806431
#> 167  69   16   28 1848 0.737 0.950 -1.288674902 -1.02451828  1.84881974
#> 169  89   16   28 1631 0.674 1.076 -0.745500308 -2.52879522  1.58451445
#> 172  69   12   23 1447 0.759 1.213  0.530804344  1.84250352  1.21319119
#> 175  66   15   25 1548 0.727 1.134  0.944921624 -1.11778822 -0.81617689
#> 178  82   12   26 1271 0.695 1.381 -0.397290704 -1.82656934 -1.80972586
#> 180  78    8   29 1383 0.756 1.269 -1.411116579 -0.34528965  0.02302540
#> 182  63   15   28 1544 0.805 1.136 -0.494058240  0.91942661  1.54231849
#> 184  96   17   26 1465 0.683 1.198  0.074013695  1.21329950  1.48512380
#> 186  78   12   21 1477 0.672 1.188  0.292365742  0.41607194 -0.17413164
#> 188  70   12   29 1432 0.692 1.225 -0.560072607 -0.47887899 -0.30824360
#> 190  73   18   29 1548 0.773 1.134 -0.723742993 -0.15272148 -0.07582063
#> 193  75   12   28 1511 0.739 1.162  1.034915607  0.46588507 -0.39645277
#> 197  61   16   30 1513 0.771 1.160  0.792977523 -0.65565354  0.40212370
#> 199  71   18   30 1769 0.699 0.992  0.973897917  0.06565860 -0.86180278
#> 202  89   13   29 1154 0.750 1.521  0.399732537  1.60731839 -0.44713833
#> 204  74   16   30 1611 0.729 1.089 -0.163027001 -0.40643425 -1.20336927
#> 206  66   12   30 1446 0.780 1.214  0.002003821 -0.87649437 -0.71570619
#> 208  80   12   27 1475 0.762 1.190  0.262932096 -0.84763318  1.16337463
#> 210  77   11   29 1583 0.777 1.108 -0.944253571  1.92423824  0.53662015
#> 213  71   18   30 1371 0.769 1.280 -0.020809622  0.78642601 -0.06017020
#> 216  82   15   29 1499 0.689 1.171 -1.624355298  0.76883406  0.25297757
#> 222  70   16   25 1568 0.696 1.119 -0.075385947  0.71191139 -0.31122358
#> 224  86   12   29 1783 0.703 0.984  0.651170630  1.71327696  0.65621828
#> 226  70   11   22 1445 0.722 1.214 -2.762208181  0.60072500 -0.09606253
#> 228  77   18   25 1604 0.781 1.094  0.631795682  1.02302651  0.50323637
#> 230  81   11   28 1750 0.670 1.003  0.507783210  1.99487093 -0.33663323
#> 232  62   12   17 1525 0.732 1.151  0.031029071  2.07998241 -1.05074583
#> 234  76   12   27 1315 0.698 1.335 -1.659223597  0.85951656  0.33184843
#> 236  73   13   23 1536 0.725 1.142 -0.397954790  2.25298251  1.39354000
#> 238  76   12   27 1316 0.727 1.333  0.008250213  1.05133851  0.82945677
#> 240  73   12   27 1425 0.769 1.232  0.615868046 -0.35752066  2.49980144
#> 242  73   20   30 1842 0.758 0.953 -1.442635270  0.24940482  0.95671528
#> 248  81   15   28 1486 0.754 1.181 -0.445603350  0.36038853  0.22866263
#> 251  76   14   25 1409 0.715 1.246 -0.539899621 -0.15888750 -0.89557471
#> 253  73   11   30 1475 0.726 1.190  0.747266650 -0.84893730 -0.58237541
#> 255  86   16   30 1293 0.747 1.357  1.022910682 -0.30817209  0.62670270
#> 257  70   16   29 1463 0.749 1.200 -0.530684462 -0.84313322  1.27734887
#> 259  74   12   29 1344 0.739 1.306  0.998620112 -1.08495084 -0.02557450
#> 267  76   16   28 1346 0.762 1.304 -0.180464932  0.56012267  0.96547678
#> 269  78   18   30 1440 0.666 1.219 -1.306218417  0.94372813  1.28988221
#> 276  70   11   29 1295 0.748 1.355 -0.737478244 -0.59026814  0.79049777
#> 278  74   18   30 1636 0.680 1.073 -1.695797722  0.17357572 -0.57861193
#> 280  74   18   28 1659 0.739 1.058  0.412736356 -0.71689252  0.32160994
#> 282  73   16   29 1123 0.786 1.563  0.068217518  1.07460841 -0.04268300
#> 284  67   16   29 1337 0.766 1.312  0.425091601 -0.83503902 -0.55432646
#> 286  76   16   26 1391 0.705 1.262  0.281900424 -1.15529347  0.98674242
#> 289  65   18   30 1277 0.812 1.374  0.465594866 -0.02596412  0.98546501
#> 291  69   16   29 1380 0.819 1.272  0.904694126 -0.71802408 -1.38204715
#> 293  89   18   30 1715 0.746 1.023 -0.192109974  1.86513687  1.38911339
#> 300  80   15   20 1732 0.685 1.013  0.096913951 -0.08342729  1.16138288
#> 302  77   13   29 1351 0.769 1.299 -0.567463495 -0.59814861 -0.03260197
#> 306  81   13   29 1345 0.737 1.305  1.570499514 -1.02134121 -0.12984615
#> 308  73   12   30 1343 0.720 1.306 -1.174398132 -2.36361047  0.10507738
#> 310  66   18   29 1191 0.785 1.474 -0.231298727 -0.35068555  0.06071841
#> 313  75   18   29 1436 0.750 1.222  0.498473548 -1.51527296 -0.69984085
#> 315  78   18   30 1243 0.748 1.412  0.048050410  0.92770379 -3.13520638
#> 317  73   12   19 1274 0.728 1.377  0.763080018  0.33144079 -0.83900366
#> 319  73   15   29 1272 0.697 1.380  0.166429321 -0.79810912  1.52004766
#> 321  73   14   29 1238 0.757 1.418  0.704934813  0.20410796  0.34738247
#> 323  76   12   27 1557 0.705 1.127  1.857320747 -0.42324428  1.87063000
#> 325  77   16   29 1818 0.734 0.965  0.063782103  0.36319367  0.65337814
#> 328  82   14   23 1514 0.678 1.159  0.537783016 -0.99894980 -1.06462471
#> 330  77   20   23 1713 0.756 1.024 -0.408141186  0.41512223 -0.61422530
#> 332  78   12   23 1491 0.710 1.177 -0.023367060  0.72231282  0.18380996
#> 334  71   18   30 1426 0.731 1.231  0.086373650  1.30506834  0.09834321
#> 336  76   16   30 1832 0.769 0.958  0.130168628  1.00292741 -0.18242960
#> 339  75   16   30 1891 0.709 0.928  0.233308093  0.63137175  0.17457960
#> 341  60   12   30 1379 0.806 1.273  0.376785889  0.39113702 -1.46809627
#> 344  70   16   26 1796 0.742 0.977  1.090750381  0.51585923 -1.08154516
#> 350  68   14   26 1444 0.778 1.216 -0.542512508  1.24634382  0.72513005
#> 352  89   14   29 1509 0.756 1.163  0.166018758  0.89443563  0.24918263
#> 355  79   20   26 1548 0.711 1.134 -0.022229629  0.21828277  1.82277888
#> 357  74   12   26 1171 0.733 1.499  0.905942233  1.74015644  1.98425043
#> 360  73   12   23 1661 0.698 1.056 -0.246907052 -0.51274790  0.51235067
#> 362  66   13   30 1495 0.746 1.174  0.829722760 -0.14877431  0.62737133
#> 366  72   16   24 1354 0.733 1.296  0.467622252  0.29693638  0.23845949
#> 368  80   16   28 1704 0.711 1.030 -0.281557875 -0.04292601  0.66403914
#> 371  61   13   30 1319 0.801 1.331 -1.491816459 -0.18128380  0.47972270
#>            noise4      noise5      noise6       noise7       noise8      noise9
#> 1   -1.643948e-01 -0.47888129 -0.57397013 -1.257505219  0.446462842 -0.20482942
#> 3    7.538452e-01 -0.13120827  0.82618664  1.289151724  2.141352426  1.77339406
#> 6    2.944371e-01 -0.45905835  0.67353305 -0.404978549 -0.085534492  0.61598494
#> 8    1.378013e-01  0.16270516  0.19111560  0.046912489 -2.119964982  0.17023909
#> 11  -7.924753e-01 -0.67575400 -1.85146383  0.577407955  0.292067184 -0.29420729
#> 14  -4.527023e-01 -0.73205498 -0.03908070 -0.562208117  0.621890081 -0.09775946
#> 16  -2.395789e-01 -0.31996361 -1.01936303  0.119224432 -0.441372128  1.52663495
#> 18  -9.831256e-01 -1.15519743  1.09768436 -0.809542262 -0.751625465  0.34084498
#> 20  -5.671716e-03 -0.29588956  0.88650670  0.600706233 -0.302078120  0.26779255
#> 23  -8.965084e-01  0.84345247  0.75988418  1.383775757  1.840941794 -0.12336360
#> 26  -2.426388e-01 -1.67163847 -0.98025265  1.145856527 -0.857533795 -1.24036335
#> 28  -1.332207e+00  2.00322645 -2.50001171 -0.377743204 -0.815532742 -1.71531161
#> 30  -5.722194e-01 -0.57100075  0.06032381  0.157795821  0.770578343  0.75047183
#> 40   9.711795e-01  0.53541136  0.41525248 -0.552619872  0.905852670  0.72456731
#> 42  -1.859243e+00 -0.47733630  0.24458129 -0.912154322  1.993150800  0.17487200
#> 44   8.124252e-01 -0.93276551 -0.39111120 -0.072036217  1.198339972 -1.33368038
#> 46   6.595638e-01 -0.90671516  0.10919314 -0.560519419  0.026652458 -0.43060353
#> 48  -5.623774e-01 -2.71307223 -1.09212678  0.100194195 -0.826864513 -0.74769932
#> 52  -6.161201e-01  0.33472581  0.15627031  1.009830507  1.759893254 -0.44687368
#> 54  -1.093387e+00 -1.17636487 -0.33312354  0.480062046 -1.560067760  0.81397472
#> 56  -1.676798e+00  1.58890682  0.02653110  0.298855429  0.394770116 -1.36579717
#> 61   5.670389e-01  0.61175289 -1.32759082  1.092570304  1.340702045  0.22813295
#> 63   4.426423e-01  0.01086962 -0.09844410 -1.503062786  1.985961013 -0.72609235
#> 67   9.033405e-01  0.36818113  0.43905910 -1.869835220 -1.532304959  0.19685991
#> 69   1.078220e+00 -0.67088509 -0.12897487 -0.006920486  0.132607072  1.51208523
#> 73  -2.775533e-01  0.08189777  1.15864880 -0.868513838  0.023324324 -0.26792175
#> 77  -1.534731e+00  0.60891834 -0.09251995 -0.393581293 -0.447829785  0.89303412
#> 79  -1.007185e+00 -1.09744141 -0.61278832 -0.366034235 -0.147892840 -0.02109460
#> 85   1.846665e+00 -1.11074408 -1.65527245 -1.181848286  0.655585719 -0.52451102
#> 87   4.598973e-01  1.47304659 -0.88896457 -1.362698426  1.025836345  0.70440278
#> 89   1.047147e+00 -1.47398368 -0.27563686 -0.120697773  1.196974085  0.35248102
#> 92  -1.565506e+00  1.95646236 -0.24793620 -2.245493048 -1.106082051 -0.47945838
#> 94   8.108519e-01  0.27630069  0.21370914 -0.696030686  2.093449812  1.46044862
#> 96  -1.288092e+00  0.98405886  1.05733544  0.106396714 -0.271302666  0.11304593
#> 98   4.005176e-01 -0.50224904  0.35559885 -1.089233970  0.798498614  0.26608982
#> 103  1.409657e+00 -0.27534111  0.05581697  2.273330238  0.302310949 -0.59051842
#> 106 -6.914289e-01  0.49480022 -1.40953044 -1.737196764 -1.223382139 -0.36506531
#> 108 -4.268607e-01  0.02869604 -0.50726182 -0.556630446 -0.939569571 -0.64358033
#> 111  1.053535e+00 -0.61855619 -1.17979643  0.541345461  1.031508328 -0.05774120
#> 113  6.529466e-01 -0.74028834 -2.09551183  0.257085838 -0.651081804 -0.83321601
#> 117 -2.321983e-01 -1.81405117  1.33175661  0.751739693  1.397538174 -0.63360401
#> 119 -1.956481e-01 -1.07289362 -0.27998834 -0.191093270  1.231368954  0.31316979
#> 121  5.202874e-01 -1.60213507 -0.51785413 -0.612603028  0.755970295  0.83531660
#> 124 -1.613885e+00 -0.56846776  0.31398873 -1.153054415 -0.641089430  0.28192761
#> 127 -1.698405e+00 -0.24527429 -0.26430974 -0.225587854  0.067132554  0.20733841
#> 129 -6.310242e-01  0.05338848  0.57817177  0.966220678 -0.314000659 -0.26447198
#> 132 -4.954504e-01  0.42884707  1.88640008 -0.596901045 -0.044608396  0.42461622
#> 135  8.744545e-01  1.26052689 -0.98127818  0.070839262  0.960703149  0.97537189
#> 137 -1.193721e+00 -2.50521660 -0.02357319 -0.615773388 -0.265721687 -0.55928078
#> 140 -1.020092e+00 -1.76801703  0.78282172 -1.217225559  0.107327544  0.09957890
#> 142  5.444176e-01  0.55377052  0.32319482  0.924510980 -0.245063426  1.14730601
#> 146  2.947221e-02 -0.60253574  0.48501054 -0.167359953 -0.526583214  0.88017122
#> 148  2.656186e-01  0.44102805  0.54922692  0.942624343 -0.363949929  0.81028734
#> 150 -2.825134e-01 -0.55827044  0.17902661 -0.491501081 -1.173213129  0.27847646
#> 155  1.493181e+00 -0.69875194  0.27074685 -0.934335717  0.087360501  0.27279129
#> 157  1.926288e+00  0.46856174  0.42214179  1.391090527  1.142210580  1.06091622
#> 162  4.987549e-01  0.09568110 -1.38500768  0.239703251  0.629895595  1.48365664
#> 164  2.837287e-02  0.07582120 -0.15338277 -0.698020644 -0.140126350 -1.65212581
#> 167  3.619744e-01 -0.60838823  0.45243631  0.135119101  0.304676190  0.88955045
#> 169 -4.598486e-01  1.36451803 -0.97968242 -0.867472291  0.062579348 -0.82993961
#> 172  1.607247e+00 -0.02606599 -1.35100202  2.264203041 -0.595492227  0.42693460
#> 175 -1.304674e+00 -0.61203017  1.08210205 -0.052713487 -0.479425569 -0.89346127
#> 178 -1.935180e+00 -0.77297812 -1.70946013  1.118375668  0.594878117  0.26009633
#> 180  5.628838e-01  2.24561761  0.25696684 -0.415633936  1.726124100 -0.06549335
#> 182 -2.706323e-01  0.50918164 -0.39373778 -2.770232145  0.833799542 -1.22423158
#> 184 -6.364970e-01  1.14090853 -1.30333307 -0.707034259  0.593210372  0.84335247
#> 186  7.205944e-01  1.56695678 -0.05906856 -0.333092356  1.039755943  0.06030896
#> 188 -6.666134e-01 -0.02960781 -1.60440837  1.083720329 -1.106023997  0.31143406
#> 190 -7.513125e-01 -1.73470457  0.78364743  1.191743889 -0.167456674  1.87764034
#> 193  1.867603e-01 -1.92080154 -2.81047495 -0.003744799  1.125636824 -1.63022348
#> 197 -8.087900e-01  0.52553790  0.73204999 -1.108041238  0.039521644  0.75211715
#> 199 -1.095117e+00 -0.24153164  0.84117649  0.279373635  0.184359531  1.11229562
#> 202  2.599273e+00  0.74072561 -0.43659298  0.884351215 -0.002089814 -0.40887097
#> 204  6.002466e-01 -0.88728470  1.94844279  1.965451946  1.355040490  0.74278901
#> 206  4.404425e-01 -0.61009243  0.08980867  0.365724118  1.943336988  1.32602194
#> 208  3.161941e-01  1.13887239  0.79136928  0.316551332 -0.418649317  1.20401617
#> 210  8.763725e-01 -0.76683207 -0.17310842 -0.753742968 -0.472037233  0.45474206
#> 213 -1.128438e+00 -0.08263854 -0.17606023 -0.049954406 -1.300070987 -1.45052135
#> 216  2.354278e+00  1.77210142 -1.50154119 -0.706987179 -0.727620088 -0.34968155
#> 222  4.506096e-01 -1.58802023  0.52663906 -0.293597144  0.276905464 -0.88370751
#> 224 -7.804845e-01  0.86896084 -0.20686490  0.355343762  0.012660397 -1.29225591
#> 226  2.703433e+00 -0.79467732  0.05425567  0.668959977  0.395660393  0.41526737
#> 228 -8.736823e-01 -0.12189276  0.09460728 -1.138706691  0.499187607 -0.72418126
#> 230  1.328297e+00 -1.44770280  0.46221047 -0.454916356 -0.892797134  0.18926976
#> 232 -1.003012e-01  0.77145057  0.09380287  0.425618710  0.118794351  1.47361822
#> 234  7.330261e-01 -0.95933506 -1.02513695 -1.742519272 -1.420976163  1.09987872
#> 236  6.961334e-01  0.39703205  2.35600620 -0.559123927 -0.732301669 -0.35050976
#> 238 -8.991687e-01 -0.33690821 -0.78181004  0.091668756  3.196952116  0.74731135
#> 240 -8.268887e-01 -0.59529070 -1.20257610 -2.396204159 -0.631206700  0.72504905
#> 242  3.038499e-01  1.33386485 -0.87551810  0.946040620 -1.528897845  0.80817144
#> 248 -4.428912e-01 -0.21068533 -1.93234671  0.656954377 -2.104165495 -0.09326675
#> 251  4.185894e-01 -0.67593397  0.27995945 -2.062621548 -0.576601114 -0.93494775
#> 253 -4.517134e-01 -0.36729358  0.50153037 -1.622614037 -1.401692708  0.17445710
#> 255  7.137715e-01  0.90032733  0.75038207  0.298587778  1.111638879  0.33545597
#> 257  3.836691e-01  1.45207593 -0.79487127  0.391921155 -0.127419361 -0.68385640
#> 259  3.876727e-01  0.70027936  2.21854740 -0.272017084  1.167692712  0.61710084
#> 267  9.314116e-01  2.97103982  0.40545829  0.327050700 -1.832043905 -1.41933076
#> 269  1.858781e-01 -0.44314718 -0.85276176  0.051848981 -0.397819828 -0.79127116
#> 276 -1.722830e+00  0.48378275 -0.61878169  0.307370067 -2.633998165  0.37005741
#> 278  5.552711e-03  0.79256852 -0.22306786 -2.478719026 -0.010038188 -0.36028395
#> 280  1.003572e+00  1.02768246  1.59895252  0.401194926 -1.366980279 -0.06059989
#> 282 -1.838809e-01 -1.46965524 -1.24950676 -0.581053005 -0.815991968 -0.86100553
#> 284  9.415233e-01  0.24895714  0.09312425 -0.687775498  0.272550252  1.61968515
#> 286 -1.650481e+00 -0.07108644 -0.18860039 -0.990224695 -0.278545660  1.73607195
#> 289 -4.879961e-01 -0.50198777 -1.04536709  1.523550289 -0.089128908 -1.20785734
#> 291 -1.551705e-05 -1.05068459 -0.61499291  0.678350729 -1.166899884  0.08226290
#> 293  1.171601e+00 -0.20822981 -1.22851970 -0.996051597  0.650863519 -0.44313511
#> 300 -3.969945e-01 -0.83023508  0.97180206  0.261818876 -0.558358603 -0.53525476
#> 302  1.022019e+00  1.00243604 -0.36617068  1.083281948  0.469744684  0.72544750
#> 306 -2.023428e-01 -1.97029960  0.47758564  0.807183351  0.608511536 -0.22933907
#> 308 -8.242174e-01  0.78924002 -1.97851694  2.565159032  0.122246910  1.07204697
#> 310 -3.346106e-01 -2.60806885  0.09535175 -1.184304861 -1.133627251  0.98438483
#> 313  5.661046e-01 -0.55274348 -0.33172410 -0.751495239 -0.160055811 -0.59466457
#> 315  1.593058e+00 -0.71174154  0.63004337 -0.279672620 -1.227171843  0.14489905
#> 317  2.012181e+00 -0.86604925 -1.25947330  0.410259815  0.525532018  1.33663691
#> 319 -1.166718e+00  0.91423626  1.64362018  1.497329272  1.697953133 -1.30686353
#> 321 -1.388681e-01 -1.77769902  1.42498603  0.575638560 -1.812827055  1.45824573
#> 323  1.869876e+00 -0.41354511  1.55486338 -1.231366710 -2.884400585 -1.46995377
#> 325  9.915368e-01  0.67798169  0.38815305 -0.639118264 -1.995565330  2.04144753
#> 328  1.333841e-01 -0.48270373  1.78835649 -0.612605750  0.680086748  1.18283963
#> 330 -5.567841e-01 -0.10905729 -1.05085280  1.708498792  1.973774405  0.15117469
#> 332  9.924921e-01  0.55655578  0.86198326  0.633746508  0.565293021  1.68806302
#> 334 -5.774523e-01  0.67251268  0.59735911 -0.750712823 -0.001553540  0.89261995
#> 336 -4.870979e-01  0.95710856  1.15616884 -0.349870582 -0.532770810 -0.22564815
#> 339 -1.253393e+00  0.09458979 -1.09649031  1.113708295 -1.044990323 -1.34992175
#> 341  1.649494e+00  1.51685111 -0.33187866 -0.006957850 -2.014698256  0.83014021
#> 344 -9.804044e-01 -0.90294782 -0.55954290 -0.271268106 -0.510556771 -0.18602417
#> 350  3.678457e-01  0.50589434  0.47972320 -0.376764535  0.251541872 -1.03079166
#> 352  4.697681e-01 -0.88268429 -0.93688846 -1.060075407 -1.644939947  0.98400393
#> 355  1.511214e+00  1.24087708 -0.49772077  0.035662358 -0.038205462  0.91473200
#> 357  2.301863e-02 -1.42635396 -2.25546124  1.291236213 -0.747644984 -2.69112908
#> 360 -1.334339e+00  0.09894491  0.98420403  1.540746060 -1.861810715 -0.34183038
#> 362 -1.060968e+00  1.09190754  0.41451864 -1.504016431  0.885651920 -0.32334815
#> 366  2.600626e-01 -0.67910512  0.17728317  1.756636680 -0.717377703 -2.22005666
#> 368  8.545524e-01 -0.28725994  0.22675177 -0.560440234 -0.486049272 -0.14093343
#> 371  1.774017e+00 -1.17692652 -0.85590872  1.420148719 -1.347136070  0.06219019
#>          noise10      noise11      noise12      noise13     noise14
#> 1    0.110836941  0.069794765  1.141868017 -0.628950231  1.59549225
#> 3    1.546244999  1.369734352  1.043564287  0.712438298  1.34198727
#> 6   -0.266836595  0.071530294 -0.002871695  0.011248481  0.23311297
#> 8    0.908769413 -0.066696202  0.103219017  1.306491128 -1.42788864
#> 11  -0.291521773  0.967207988 -1.756918370  0.519734284 -0.87211896
#> 14   0.523988192  1.677614419 -0.245058433 -1.981957276  0.58446194
#> 16   0.091112980  0.751333853 -1.292763585  0.240766137  0.80417641
#> 18   0.308028017  0.458894713  1.171088501  1.803097432 -0.66287516
#> 20  -0.338961778 -0.779511319  0.176291599  1.144960351  0.22807951
#> 23   1.400302478  0.523869924 -1.267753547  0.236213371  0.14250649
#> 26   0.091998309  0.701910569 -0.350556952  0.888737579  0.23883594
#> 28   0.365747717 -2.506178454  0.083126924  0.251632073 -0.60299581
#> 30   0.301098963  1.332048763 -0.629836176 -0.937862631  1.21162174
#> 40   0.892942051 -0.055617330  1.047946173 -0.395511413  0.77050384
#> 42   0.534336476  0.427220860 -1.196936680  0.158314634 -0.96260183
#> 44   0.764313712  0.966229986  0.294688358  1.872121606  0.01083961
#> 46   0.451787082 -1.347777040 -2.598610201 -1.178017479 -0.74641105
#> 48  -0.814884145 -0.333664604 -1.717617294 -0.600453445 -0.81791430
#> 52  -0.675740147  0.223509238 -1.644301458 -1.121835542 -0.24880743
#> 54   0.402241772  0.616617435  0.598503160 -0.120278103  1.00445856
#> 56   1.285441477  0.304144104 -0.103973589  1.160073708  0.62512879
#> 61  -0.502661325 -0.045570009 -0.338952078  1.992538590 -0.57069164
#> 63  -0.473624669  0.898979261 -0.312514441  0.237787107 -1.98540563
#> 67  -0.490846066 -1.114349190  2.426673256 -1.054857833 -1.31683488
#> 69   0.015992089  0.427349861  0.133730940 -0.007546664  0.29202159
#> 73   0.449308018  0.578116418  0.051829463 -1.353609195 -0.87944016
#> 77   0.651785678  1.132731481 -0.949794219  1.910832047  0.85592262
#> 79   0.430735128 -0.558183805  0.743987850  0.862818241  0.15511766
#> 85   0.380075803 -0.993274914  1.444477435 -0.072674304  1.03523524
#> 87   1.115973263 -0.904530585 -0.557344174 -0.668841920 -0.32722114
#> 89  -0.570189125 -0.690328296 -0.071939320 -0.023989497 -0.12280744
#> 92   0.395635765  0.601494416  0.620199966  0.668102166  0.78975596
#> 94   0.424286187 -1.233839094 -1.973504335  0.688594697  0.82400068
#> 96   0.160357965  0.156336264  1.440475948  0.534884721  1.42683561
#> 98   0.325252307  0.243697443 -0.301891407  0.852903150 -0.71701370
#> 103  0.553360212  1.600645574  0.736902568 -0.999189818  0.64706010
#> 106  0.245407720 -0.005222829 -1.774234417  1.241628915 -0.81079156
#> 108  2.148895997  1.355702309  1.017709415  0.076287830 -0.39514818
#> 111  1.027857647  1.912958466  0.015055886  0.192593478 -0.13294297
#> 113  2.182252679 -0.599416358  0.400938358 -0.866747842  0.29237974
#> 117  0.485375380  0.599423107  0.908552286  1.090575811  0.04817175
#> 119 -0.480286570 -0.449630112  1.069435076  0.102800457 -0.73471004
#> 121  0.870484350  3.116710044  0.299981613 -1.549964996 -1.04160571
#> 124 -0.046481359 -0.320991809 -1.798836359 -0.461553672 -0.64341578
#> 127 -0.500178648 -0.347430401  0.065190711 -0.185413042  1.27284745
#> 129  0.934246715  2.100849169  1.467144799 -0.046398738 -0.34227709
#> 132  0.517865655  0.172745003  1.127560226 -0.578345585  0.27479357
#> 135 -0.050873539 -0.692236630  0.518626385 -1.350886445  0.40169253
#> 137  2.329928670 -0.508426186  0.281646209 -0.546511932  0.49543495
#> 140 -0.037425518 -0.756724506  1.231925752  0.995589123 -0.30733371
#> 142  0.303811696  0.486953301  0.982719773 -0.504253347 -0.62030091
#> 146 -0.068794325  0.620234856 -0.239806744 -0.735075472 -0.15158112
#> 148 -0.661345654  0.699025430 -1.434981210 -0.139971023 -1.60405882
#> 150 -0.748271599  0.853751107  0.642628320  2.626177707 -0.43606380
#> 155  0.016005153 -0.339872251  1.166164835  0.683823576 -0.49969212
#> 157 -0.510498715  1.192487818  1.351609388 -1.747848707 -1.76652775
#> 162  0.348126967 -0.144385871  0.411051380  0.822623676  1.75218423
#> 164 -0.114636746  0.012615351 -1.179126933 -0.381083523  0.16218558
#> 167 -1.079332282  1.195824502 -0.181019807  0.466913734  0.12714295
#> 169  0.608142629 -0.759885196  0.067768874 -0.025513954  0.66266364
#> 172  1.562328757  0.952996273 -0.918161515 -0.238765245  0.86126680
#> 175 -0.593584368  0.327206582  0.213066157 -0.262638993 -0.95312927
#> 178  0.942504503 -0.290893112 -0.443801946 -0.485538525  0.92353769
#> 180 -0.571087852  0.110211494  1.414032264 -0.689283187  1.73226715
#> 182  0.327811430  0.400580082  0.620262822  0.702122992  0.45826163
#> 184 -0.543184186 -1.775044870 -1.138857801  0.793933123 -1.79080031
#> 186  0.775401754  1.357749182  1.668260135 -0.109065534  2.05999399
#> 188  0.375110850 -0.164173466 -0.337542923 -1.208499345  1.50037417
#> 190 -0.497622261  1.160026118 -0.689638658 -0.855239348  1.17021152
#> 193  2.168558845  1.001045853  2.021516425 -1.333618915  0.12316967
#> 197  0.808635090  0.453220211 -0.160863456  0.459927254  0.51205049
#> 199 -1.035023432  1.026454814 -0.069986230  0.009734550  0.05485665
#> 202  1.258069012  2.100013396  0.227155381  0.121563209 -1.99153667
#> 204  1.844691097  1.115298965  0.487872494  0.417860683 -0.32439998
#> 206 -1.640094288  0.243356776 -0.707080870  1.083957654  0.19347351
#> 208 -0.945864601  1.681599469  1.860996228  0.029665147  1.70760986
#> 210 -1.438241248  0.334733897  1.180730844  1.407037293 -0.15784067
#> 213  0.873127892  1.377973809 -0.217048571 -1.329437060 -0.35366686
#> 216  0.319339748 -0.943823853  1.196072934  1.260583229  1.35165170
#> 222  0.685336251 -0.555009477  2.267081889 -2.096450039  0.04408136
#> 224  0.532968192 -1.357929831 -0.136231901  0.105529331  0.73987037
#> 226 -0.034731046  0.991950223  0.224963546  2.924925661  1.30436026
#> 228  1.604079549 -0.240093507 -0.145758593  0.398527450  0.91791516
#> 230  0.043710826 -1.641947910  0.416448114  2.145553887 -0.03304092
#> 232 -0.977568696  0.168534280 -1.131346367 -0.366471944  1.23054779
#> 234 -0.453537735 -1.335051789 -0.510277330 -0.444192559  0.73463220
#> 236 -0.927517633  1.040546507 -0.795326705 -1.526502150 -1.49551879
#> 238 -0.191836698 -0.705565384 -0.695436218  0.407328727 -0.26133638
#> 240  0.686822255  0.780427330 -0.620830309  0.352767497 -0.43917425
#> 242 -3.488035955  1.779702685 -1.651364741 -0.304449687  0.49699858
#> 248 -0.688966516 -1.418043760  0.464102279 -1.548059056  1.16540886
#> 251 -0.076015648  0.902837176 -0.014169812  0.029335230  2.11384840
#> 253 -1.345419730  1.366046847 -0.502474143 -0.271431172  0.40975682
#> 255 -1.245134002 -0.583476562 -0.085029828  1.070450638 -0.88818590
#> 257  0.435440467  2.137724610  1.556983867 -0.538547004 -0.22615502
#> 259  0.348445416 -1.697097809 -0.237407508 -1.061493734 -0.90726062
#> 267  1.114097026  1.920999701  0.021765699 -0.865149452 -0.81618363
#> 269  0.965876573  0.019081576  1.326476672  0.305279623 -0.24106573
#> 276  1.433375078 -0.078057679  2.357138949 -1.412724128  0.68684243
#> 278 -0.188002201  0.529913778  0.692675491 -0.318002576  0.28972815
#> 280 -0.124821487  0.728495813  1.789053262  0.047083960  1.21001463
#> 282  0.664202672 -0.307237155  0.958360695 -0.928592643 -1.44639216
#> 284 -1.070127918 -0.011477513  0.604708215  0.566836790 -1.02066350
#> 286 -0.093759937  0.586367500  0.481288941 -0.855690455 -1.39140669
#> 289  1.311258597  0.399234769 -1.194878242 -0.300449114  1.89655340
#> 291  2.225716656  0.951930551  0.953254125 -1.068712858  1.83791466
#> 293 -0.300906172  1.526739246  0.240047205  0.519490077 -0.37604586
#> 300  1.507438882  0.734689853  0.761733437 -1.261177615  1.82205563
#> 302 -0.110625663  0.596117629 -0.552640822 -1.262973850  0.57677954
#> 306  0.652596566  0.871869717 -0.053720480  1.024840095 -0.82145135
#> 308  0.194983341 -0.414887354  0.912567070 -0.700218827 -1.39987632
#> 310 -0.155820396 -0.444874710 -0.047672825  0.424219648  1.37993287
#> 313 -0.355953163 -1.296128018 -1.302242178 -0.583715569 -1.90585556
#> 315  1.975656544 -0.735215332 -1.267743325  1.507233837  0.75836274
#> 317  1.357855621 -0.820601885  0.339894107 -0.486113899  1.35730304
#> 319  0.914106559 -0.107817070  1.975744989  0.647912437 -1.33805954
#> 321  2.397051876  1.244979481 -0.134008295  0.972706043  2.03669907
#> 323  0.580232472 -0.066503037 -1.330504897  0.562172038  0.38552344
#> 325  0.627147996  0.151420687 -1.117100995 -0.169130643  0.05634288
#> 328  0.926701811  0.458266239  0.787730561 -2.411982042  0.95222305
#> 330 -1.496148179 -1.131240216 -1.417247309 -0.251699964  0.55459633
#> 332 -1.204255811 -0.062149675  1.059477090 -1.866060350 -1.87153333
#> 334 -1.948964267  0.854440818 -0.860542786 -0.298333295 -0.24390109
#> 336 -0.261830822  1.024739012  1.005247799  0.109666401 -1.34755587
#> 339  1.182144748 -0.655926971  1.215565826 -0.691320681  0.34864125
#> 341 -0.001017215  0.428460696  1.672710705  0.620459241 -1.11788615
#> 344 -0.794851299 -0.842161461  0.689838172  0.487983106 -0.08193441
#> 350  0.108760153  2.570837162  0.232157674 -1.303769050 -0.82955317
#> 352 -1.015271790 -1.126671361  0.887676055  1.525919909  0.64896533
#> 355  0.316631055  0.935623157  0.186759600  0.118584490 -1.19450154
#> 357  0.719274848 -0.573454338  0.166050566 -0.870877815 -0.98502481
#> 360 -0.784690924 -0.014738613  0.863786366  0.123313683  0.31830108
#> 362 -0.278817242 -1.889376991 -1.609168061  0.509024643 -0.08868684
#> 366  2.654572744  0.558332863 -0.168361602  1.395151316 -0.32708768
#> 368 -1.819164286  2.114972759  1.366278615 -1.103711655  0.45650339
#> 371 -1.095014436 -0.947021083  0.283281139  0.021704358 -0.15103363
#>          noise15     noise16      noise17      noise18     noise19      noise20
#> 1   -0.570703615 -1.17056381  1.129461470  0.682116609  0.31032572  1.429982096
#> 3    0.553086573  0.99954557 -1.351246110  1.101574366  0.07188301 -1.049804293
#> 6   -0.645209695  0.65746872 -1.270510949 -1.097235674 -0.31987215  0.378252279
#> 8   -0.808761611 -1.26661457 -1.359607779  0.887061177 -0.06124014  0.329707674
#> 11  -0.812110066  0.24850455  1.057725975 -1.005959850  0.75767564 -0.767050180
#> 14   1.215473190  0.56108570 -0.917608448 -3.108079392 -0.61364929 -1.873793341
#> 16   1.796541921  0.19290155 -1.301371505  0.012554394  1.52426622  2.071637835
#> 18   1.148022183  0.20359625  1.605637557 -1.065045540  0.46802555  0.987454683
#> 20  -0.441435689 -0.05586193  0.645039697 -0.536420549  2.29702540 -1.069218203
#> 23   0.838842102  0.06782346  0.416627622 -1.283790488  0.22787234  0.169165305
#> 26  -0.868935106  0.79097726  0.389199085 -0.583207541  0.01319329  0.141352962
#> 28   2.417692695 -0.42227824  0.061394475  1.154246802 -1.10778401 -1.190852988
#> 30  -1.259120417  0.63853498  0.473049165 -0.269664288 -0.14875025  1.334608566
#> 40   0.123051924 -1.65601873 -0.799722114  0.665232324  0.41784099 -0.653372978
#> 42  -0.444457540 -0.91270481  0.424385705 -2.014794718 -0.43251053 -0.031555238
#> 44  -0.836689789 -1.02297944  1.790372621  0.253932038 -0.04180797 -1.538291348
#> 46   0.768438046 -0.19625546 -0.253394856 -0.641196445  0.68692932  0.389470711
#> 48  -0.618475645  1.17868515  0.222226727  0.079340913  0.10497143  1.914111283
#> 52   1.437916116  0.21619775 -0.860561344 -0.928754207 -0.52800564  0.667501920
#> 54   0.131843112  0.30747754 -0.485895232  0.600266237  0.85394982 -0.809557921
#> 56   0.284812763  0.29781518  0.389305330  0.486861121  1.24889472 -2.278069307
#> 61   0.116968874 -1.10485933  0.534289839  0.685394404 -0.54926605 -1.133651562
#> 63  -1.190794602 -1.34842537  0.279540460 -0.357576047 -0.62976028 -0.527334574
#> 67  -0.069898972  1.29588785 -1.846975875  0.555804063  0.64907020  0.665664694
#> 69  -0.734681730 -0.28301643  1.113171061 -0.050200148  0.33230482 -0.692925640
#> 73  -1.215704994 -0.77045139 -1.242712265 -0.984781605  0.62458527 -2.296949897
#> 77   0.394139487  1.31153162 -0.637352057  0.846168744 -0.75568766 -0.131830335
#> 79  -0.252234166 -0.48652470  0.831181918  1.464865890  0.43236601 -0.773169689
#> 85   1.114841461  1.12599272  0.771224199  0.298811864 -0.52862473  0.638365243
#> 87   0.575861328 -0.13623860 -2.151834070 -0.850988125 -0.20857139  0.950931230
#> 89   0.718138309  0.52891741 -0.795314573 -0.161416404 -1.48025409  0.722499848
#> 92   0.313424194  1.54658024 -1.217562250 -0.169408364 -1.80934636 -1.548219173
#> 94  -0.365883000 -0.34816350 -2.077013458 -0.210683404  0.45626645 -1.182562642
#> 96  -0.964938376 -0.23664166  1.360785416  1.635500759  0.07633169 -0.399633101
#> 98  -0.286535120 -0.13968757  0.311774483  0.026942028 -0.25822783  0.009417941
#> 103  0.067063462 -0.22146916 -1.430162826 -1.018010327 -1.62290922  0.984722991
#> 106  0.370110214  0.23611871  0.494348986  1.043777327  1.49888582  0.281635013
#> 108  0.605400318 -0.88657966 -0.467763434  0.432180431  0.95970011  0.357525877
#> 111  2.864725647  1.22183945  0.578765408  0.854682610  0.58584175  0.012417109
#> 113 -0.016667988 -0.70061903 -0.210563706 -0.970104188  0.77697679  0.757431864
#> 117  0.119089843 -0.40294638 -0.855537948 -0.004407222  0.98429556 -0.244163873
#> 119 -1.300808462 -0.05258768  0.514573992  1.303514478  1.11424694  0.491383226
#> 121  1.156638818 -1.35851904 -1.735774079  1.483455199 -0.23819599  1.030298862
#> 124 -0.069778861  0.75875241  1.132197581 -1.544591882  0.55909954 -0.120396778
#> 127  0.214476892 -0.47763685  0.159109871 -0.019786985  0.75460057 -0.982021244
#> 129 -0.356134521 -0.02809061 -1.763907554  0.737135158  0.59053272 -0.527530337
#> 132  0.203853243 -1.35797742 -0.504899301 -1.722296859 -0.30473524 -0.465064844
#> 135 -1.903974076 -1.14390001 -2.753923558 -0.980666951 -0.04055831 -0.080073550
#> 137  0.928807114 -0.99959634 -0.742931301  1.688845465 -0.37448768  1.422603199
#> 140 -1.254407413  0.89189277 -1.985388462  0.901883330 -0.33311206  0.030560770
#> 142 -0.246177527 -0.14267026  0.177694951 -0.670243296 -0.38896724  1.218736035
#> 146 -2.995830254  1.12957756 -0.269600827 -0.966343153  2.91567464 -2.163193796
#> 148  0.292239398 -0.86445913  1.865346054  2.827335213  1.08896048  2.453851953
#> 150 -1.973078333  0.57077139  0.660705440 -0.387346216  0.51907277  1.032439300
#> 155 -0.987670327 -1.69062155  0.555231736  0.006087521  0.11587279  0.369609880
#> 157  0.923074177 -0.03209771 -1.451414274 -0.909222447  0.24030203 -0.163905236
#> 162  1.216101168 -1.45877570 -1.029356790 -1.098832540  0.57255537 -0.692005400
#> 164 -0.183199506  0.54662510 -1.438264122 -0.069962237 -0.99091631  0.666661533
#> 167 -0.579519222  1.15187357 -1.035769781  0.229589282 -0.35106173 -1.273463748
#> 169 -0.009245817 -0.77844062  1.615908336 -0.583897211 -1.14487089  0.872084588
#> 172  0.805066418 -0.71326681  1.046726328 -0.304363669  0.10818408  0.021432986
#> 175  0.137969498 -0.27895192  0.838434741 -0.925095374 -1.70471259  1.297933713
#> 178  0.521644984 -0.61665003 -2.115534412  0.903249858  1.35088069  0.099087354
#> 180 -1.413053044  0.59837716 -1.515134084 -0.920169823  0.51261200 -0.504389484
#> 182 -0.395000061  0.16418149 -0.029840868  0.810458143 -0.40529393  1.025816522
#> 184 -1.351931646 -1.22073937  0.178949100  0.753181414 -1.44533076  0.616020328
#> 186  0.824952213 -0.24264384  0.495084503 -0.391709772  1.27597949  0.392900955
#> 188 -0.172747763 -0.23363318  1.625934617  0.067352021  0.67307807 -0.137390203
#> 190  1.430618446 -0.96983995 -0.918062336  0.501965979  0.15474602 -2.503692937
#> 193  0.772017161  0.13350462 -0.241843810  1.330766801 -0.54681065  1.195563799
#> 197  0.007934963  2.23756620  0.366910055  0.554947722  1.03130524 -0.898343333
#> 199 -0.413181059  0.08037882 -0.807038893  1.110028738  1.22145365 -1.059564116
#> 202  0.001557187  1.22820276  0.013434844  0.289644059  1.37941185 -0.777552254
#> 204 -0.100829256 -1.57558736 -0.286900982  2.217678301 -0.83058436 -0.499396978
#> 206 -1.783272674 -0.53817853  0.799207220  0.685657762 -0.46758473  1.532683646
#> 208  0.095200551  0.47804745  0.237665194  0.132179801  0.34533891  1.349202210
#> 210 -1.101535016 -0.30288691 -0.004703028 -1.233061332 -0.84453641 -2.275540495
#> 213 -0.143946814  1.29346275  1.309922172  0.155333237  0.77880748  0.519124733
#> 216  0.297554423  1.22395687  0.341419837  0.461207348 -1.92306756  1.464413473
#> 222  0.153758778 -0.23263791 -1.715515918 -1.164782728 -0.60918016 -0.871545839
#> 224 -2.483856029 -1.67583551 -1.472488641 -0.291749037 -0.60736201 -1.228330216
#> 226  0.908475665 -0.64776120  1.863428114 -0.181088924 -1.42366614 -0.051475687
#> 228  0.089904986  0.45777656  0.640830124 -0.225733866  1.24242872  0.206958637
#> 230  0.519779634  2.03660259 -0.027743529 -0.171931978 -0.38619740  0.572824007
#> 232  1.492364242  0.97522042  0.674311548 -0.108305943 -1.12217445  1.797679258
#> 234  1.315276192  0.26430489 -1.999211932  0.149325551 -0.89190549  0.652603382
#> 236  0.297474241  0.45067067 -0.987248475  1.664234730  0.08826268 -0.375384480
#> 238  0.467920190 -0.05233699  1.066611527  0.152676418  0.82110757  0.814774019
#> 240 -0.918640918  1.04118903 -0.068320783 -1.558656045  0.31908522  0.083978663
#> 242  1.576778177  0.54608638  0.380666991  0.712524892 -0.60145028  1.827424187
#> 248 -0.617459457 -1.42622117 -0.100540674  0.664940347 -1.09554723  0.672099625
#> 251 -0.868544736  0.75706538  1.885687733  0.758985989  2.29053569  0.115793592
#> 253 -0.147306055  0.70536844 -1.299975362 -1.948230765 -1.17277692 -1.069488075
#> 255 -0.992318098 -0.18603321  0.095305774 -0.294752966 -0.39430854 -0.833944215
#> 257 -0.178883039  0.08997331  1.017219998 -0.945570091  0.48491509 -0.979557548
#> 259 -0.962578135 -0.08254726 -0.982227830  1.947458481  1.22789130  0.607921591
#> 267  0.505647228  0.33823008 -1.122202198  0.145986469  0.44267367 -0.256355116
#> 269 -0.334458603 -1.86386102 -0.247025869 -0.703618633 -0.84313772 -1.578925321
#> 276 -1.244107974  0.91307848  0.481131649 -1.077006150  0.92557533 -0.596380076
#> 278 -0.041384330 -0.77793221 -1.131198868  0.721256194  1.21782761 -0.892774178
#> 280 -0.232736687 -0.89896768  0.004451573  1.729537323 -0.22357543 -0.893280470
#> 282 -0.567465542  2.25161966 -1.792515220 -0.265930401 -3.57299903 -1.249902140
#> 284  0.404297976 -0.74553975 -0.560441033  0.399991453 -0.27348060 -0.614822553
#> 286 -1.227803943 -1.40765035 -0.886639858  0.341210171 -0.85146127  0.885042219
#> 289  0.138889715 -1.90194141 -0.219344275  0.290855476  0.50866082  0.349107383
#> 291  0.198878609  0.57450971  0.885287305 -0.563793298  0.79099055 -0.895684580
#> 293 -0.101775851 -0.01376196  0.576228762  0.923308352  0.85264474  1.105692830
#> 300 -2.149986783  0.80178585 -0.141702804  1.951924962  0.13656409 -0.110370658
#> 302  0.152626567 -0.84881619  1.372502994  0.575413676  1.18416208  1.603517813
#> 306 -0.772056880  0.68347183  0.316203243 -0.120124156  1.94668905  0.326184604
#> 308  0.205646054  0.25311410 -0.539473710  0.192645402 -0.44623010 -0.737371003
#> 310 -0.077026568 -0.29501119 -0.462226438  3.057498745 -0.55929102  0.148044999
#> 313  0.775144627  0.17259885  1.105333994  1.008626319  0.35310136  0.042662350
#> 315 -0.469012851  0.08822521  1.691021353 -1.214606730 -1.68481763  0.292377973
#> 317 -1.524606099  0.69967882  1.069366883  0.769932378 -0.68652035  0.511211734
#> 319 -0.325449052 -0.76186912 -0.423250698 -0.735166163  0.39175507 -1.815132521
#> 321  0.410811908  1.24511609  0.982699204  0.644248368  0.64726691 -0.662414056
#> 323 -0.053652310 -1.00542892 -1.118777048 -0.341557080  0.66657625  0.678606559
#> 325  0.611387239 -1.29194524  0.018597699  0.568851002 -0.36220451  0.254105977
#> 328  0.207518062 -0.68219343 -0.710592872 -0.714357169 -0.30914299  0.286656054
#> 330 -0.644450193 -0.28969699  0.327667870 -0.541233402  2.26376939 -0.410803559
#> 332  1.054947200 -0.95027694  0.287990565 -0.177520758  2.62855946 -2.090080302
#> 334  1.645769057  0.45716005 -0.748800546  1.752691632 -1.17246457  1.908916331
#> 336 -0.854977310 -0.21344730 -0.059388671  0.726129651 -0.30345443 -0.863907176
#> 339 -1.683707078  2.21278709 -1.648961848  0.355806322  0.19296950  2.398912393
#> 341 -0.174793639  0.73475161  0.222508262  1.585456318 -0.55947109  1.520980507
#> 344  1.191196530 -1.00416326  0.727781535 -1.738055113 -0.32896240 -0.064019967
#> 350 -0.190253286 -0.86614633 -0.085466239 -0.424949368 -0.98125935 -0.480778378
#> 352 -0.100297443 -0.94179765  1.805031255  1.216180766  0.26253001  0.249998291
#> 355 -0.717184164 -0.42313091 -0.449275715 -0.932926192  0.45606989  0.623591224
#> 357  0.822085117 -0.26225824 -0.482729893 -0.266094683 -1.77589709 -0.092443101
#> 360  3.189423027  0.17564236  0.410269509  0.042775897  1.44674698 -0.824271936
#> 362  0.246991224  1.08961592 -0.529776351  0.030207446  1.02919393  0.426220654
#> 366 -0.640031817  0.47620997  1.119289665 -0.766734104  0.10338863  0.270846704
#> 368 -0.862724941 -1.70176076 -0.073560150  1.491905251 -0.01572247  1.047893704
#> 371  1.417893872  0.80108724 -2.613769146  0.623165244  0.09937594  1.109710485
#>          noise21      noise22     noise23     noise24
#> 1   -0.671651890  0.494057245  1.24207389  1.99527361
#> 3   -0.738097925  0.129259922 -0.79877492 -0.49303206
#> 6   -0.238506693  0.044139279 -0.31183205  0.54580510
#> 8    0.859752906  1.415246963  0.51701961 -1.03235289
#> 11   0.714827734  0.433269203 -2.08773586  1.59735424
#> 14   0.438555959  0.308892284 -0.79131692 -0.40996481
#> 16  -0.431420861 -0.361493084  1.78237898 -1.33010515
#> 18  -0.644453202 -0.991180287  0.13268446  0.29092970
#> 20  -0.744880291  1.295647789 -0.78142739 -0.59896410
#> 23  -2.807382126 -0.487490810 -1.89973483 -2.23885436
#> 26   1.313328787  0.082391676 -0.85663046 -1.42941234
#> 28  -0.467317050 -0.565662219  0.84847096 -0.18025489
#> 30   0.871810093  0.677482342 -0.67590533  0.52590750
#> 40  -1.736872575 -0.938735042 -0.89772034  0.28365671
#> 42   0.048032523  0.929491786  0.61171892 -1.07207121
#> 44  -1.374057751 -0.335316329  0.86996320  0.62369692
#> 46  -2.159456238 -1.613533398  0.61129822  0.90107236
#> 48   1.123301972 -0.753998390 -1.34101055 -0.77664559
#> 52   0.836225842 -0.086515355 -2.19214762 -0.89374521
#> 54   0.745488729 -0.503820436  0.32565191  0.53519995
#> 56   0.823333763  1.787385329 -1.86519373  0.08004127
#> 61  -0.387097060 -0.119924815 -2.48654213 -1.02308589
#> 63   0.001666858  0.413756068  0.70605926  0.83226464
#> 67  -0.352921312  1.219064987  0.16404369  0.16613672
#> 69  -2.248476175  0.619594265  0.90859350  0.74698119
#> 73  -0.972549787 -0.325790543 -2.60990508  1.04627985
#> 77   0.178489104 -0.652157781  1.20300767 -1.17685528
#> 79  -1.404799822  1.037034512 -1.27788048  1.11718483
#> 85   0.090049641  0.828993737 -0.14331353 -0.16648553
#> 87   0.228188884  0.427864960 -0.14995991 -1.15003232
#> 89   0.870308610  0.271534717  1.41305347 -1.03716255
#> 92  -1.516309000 -0.863942269 -1.55432206 -0.30744689
#> 94   1.786752676 -1.104073144 -0.36405356 -1.44489902
#> 96   0.136575739  0.559554821 -0.50711195  0.93971967
#> 98   0.637427293 -0.621477025  0.92946868  0.30938742
#> 103 -0.058309364  0.797661743  0.57666460  0.64087110
#> 106  0.684773820 -0.248445533 -0.13643852  0.48504667
#> 108  0.862320249 -1.258664985 -0.06903368  1.25133497
#> 111  0.477621410 -3.265552712  1.65746670 -0.88869076
#> 113 -1.315049991  0.270130552 -0.56207815  0.53651475
#> 117 -1.084925234  0.453747371  0.27655387 -1.73005208
#> 119 -2.078598565  0.111408684  0.09473646  1.75601939
#> 121  0.554120285  0.470043254  0.12257213 -0.15488170
#> 124 -0.114265171  1.387772650  0.44672343  0.53003848
#> 127  0.810919443  0.790406956  0.02170920  0.76265531
#> 129  0.837066447 -1.269094535 -0.82645992 -1.29440727
#> 132  0.059647005  0.688617266  1.69720976  0.20572604
#> 135 -0.812729711 -0.493153880  0.89154125  2.47895798
#> 137  0.038583304 -0.519862171  1.19172288  1.77712668
#> 140 -0.041695489  0.034672712  0.59923846  0.06276456
#> 142  0.657359175  0.398006476 -0.13814684 -0.78551805
#> 146 -2.795572840 -1.238169701 -0.67194190 -0.23147931
#> 148 -0.834636429 -0.206529894  0.57528750 -0.05525098
#> 150  1.037799632 -0.381590085 -1.08411234  0.60959532
#> 155  0.063085704 -0.235113615  0.70523876  2.25875401
#> 157 -0.642367254  0.857697233 -0.84433446  0.87832071
#> 162 -0.370887866  1.598616913  0.43037316 -0.73626380
#> 164 -1.752013366  0.030425970  0.33099907  1.05848155
#> 167  0.353762856  0.702998272 -0.88024054  0.06450893
#> 169 -0.316286679 -1.236903551  1.85709803 -0.31971024
#> 172 -0.274175145  0.490367188 -0.24476231  1.92285908
#> 175 -0.305604966 -1.032461572 -0.68632900  0.73832944
#> 178  0.443293610 -1.308589057  0.99172330 -0.37774465
#> 180 -0.246766391 -1.410926920 -0.81853132 -1.04832318
#> 182 -1.209316958 -0.812098878  0.18157478 -0.12295405
#> 184 -0.602939197  0.038942648 -0.41332550 -0.30601628
#> 186 -0.675872050 -0.348323981 -0.87014010 -0.11760477
#> 188 -0.939744476 -1.167027064 -0.87751707  0.18232103
#> 190 -1.809072262 -0.222409013 -0.30426424 -0.76556253
#> 193 -0.926835215  1.441370588 -0.31619338 -2.13256630
#> 197 -1.247611499  1.142515167 -0.44476029 -0.26637196
#> 199  1.222880578 -1.328604215  0.45522111  0.46444681
#> 202  0.296190347  0.092302506  0.51322318  0.29052130
#> 204 -1.509979389 -0.029084998  0.17612142 -0.06999767
#> 206 -0.863047562 -1.325551799 -0.29133548 -0.03516987
#> 208  0.852536053  0.337870255  1.22180583  0.52544881
#> 210  1.770342445 -1.830146333 -1.37811909  0.49264563
#> 213  0.894919467 -0.240767970 -0.67229061 -0.52936243
#> 216  0.180336473 -0.562768364 -0.90170614 -0.54323064
#> 222 -0.116920072 -0.429251470  0.13986680  1.87588061
#> 224 -0.933741356 -0.714014258 -0.46866335 -0.17332795
#> 226 -1.375430771  0.402606996  1.43677002 -0.67365287
#> 228  0.884041666  2.100005591 -1.50845341 -0.68659873
#> 230 -0.045131760  1.090667211 -0.01519736  1.01137282
#> 232 -0.121121531 -0.452421699  0.80968150 -0.57231139
#> 234  0.624341436  0.299638017 -0.33192935  0.41083903
#> 236  1.278607407  1.281218047 -1.00475474  0.23349311
#> 238  0.824995466 -1.328690592  0.26885267 -0.81115019
#> 240 -0.490572391  0.536800159  1.78861364  0.53606612
#> 242  0.354128369 -1.766239910 -0.31366349  0.21990529
#> 248  0.635101733  0.760119594 -1.09034249  0.40422112
#> 251 -0.189440797  0.043187858  0.70612745  0.46797830
#> 253 -1.506657024  1.503481134 -1.36248859 -0.89561962
#> 255 -1.384672051  0.973647167 -0.25862110  1.01672535
#> 257  1.838891796  0.441749641 -0.59200685  0.86889501
#> 259 -0.280379145 -0.330019764  0.65359772  1.32583652
#> 267 -1.465064915 -1.234941917 -0.35054343  0.25168788
#> 269 -0.219295002  1.007399740  0.50832098  0.07404756
#> 276  0.235805845 -0.005006708 -1.06923412 -1.06417755
#> 278  0.412587666 -0.431523485  1.12292644 -0.43055376
#> 280  3.644521410  0.315627419  0.36908514  0.27853205
#> 282 -0.774617146 -0.391956106 -0.56001074  0.35905320
#> 284 -1.263623359  0.348917156  0.91545327  0.02753967
#> 286  1.149383183 -0.333247717 -0.47742688 -0.15445112
#> 289  0.178076011 -0.401532232  0.51429641 -1.48887128
#> 291  0.336535095  1.629752231 -0.46499900 -1.88199569
#> 293  0.529536646 -1.005442211  0.81999814 -0.26954706
#> 300  1.521900092  0.577220345 -1.14660239 -0.25721002
#> 302 -0.492881896  0.285238340 -0.55028892 -1.09507053
#> 306  1.551583874  0.113892716 -1.28683531 -0.86172089
#> 308 -0.817994040 -1.599823042 -0.49098675 -0.14229086
#> 310 -1.293396142 -0.061922296 -1.26132236 -0.54548731
#> 313  0.961878616  1.162297581 -0.45705249 -0.39300015
#> 315  1.649144066 -0.019615956  0.85398374  0.49710735
#> 317  0.397754776 -0.585285517  1.42536489  0.49043796
#> 319  0.405790839 -0.660006875  0.29093043  0.05597377
#> 321 -0.621573294  1.414799546  0.03000344  0.68960632
#> 323  0.514355152 -0.649572357 -1.48727847  1.67897574
#> 325 -0.403984501  0.566795762 -0.71790301 -0.85011062
#> 328 -0.277354047 -1.850680302 -0.21569250 -1.30351762
#> 330 -0.650341045  0.506803648 -0.69127595 -1.48728733
#> 332  0.067149419  0.989591834  1.30445174  1.92709216
#> 334 -0.167195810 -0.151591654  0.44853226  1.25358688
#> 336  2.033556963 -0.966066111  0.11012089 -2.25677160
#> 339 -0.418988163  0.491527362  0.20274344 -2.13623658
#> 341  1.206100784 -0.387982120 -1.02922430 -0.93663069
#> 344  1.447947792  1.707926758  0.93893812 -0.15156470
#> 350  2.313811374  0.818355764 -0.32150542 -0.14947514
#> 352 -0.260832804  0.708262275 -0.76940171 -0.37499654
#> 355  0.055682520 -0.240688972  2.18911005 -0.09954716
#> 357 -0.089907753  1.677147372  0.40612827  0.14451757
#> 360  0.456593701 -0.857928430 -0.43063269 -2.10297868
#> 362 -0.106524212  0.714539973  0.88248636 -1.27138234
#> 366  0.283027966 -1.317993505 -0.26754059 -0.81043443
#> 368  0.349513061 -1.138511159  0.04189374  1.26203746
#> 371 -1.561308244 -0.884222843 -0.53919840 -0.08659003
#> 
#> $y
#>   [1] 3022.5 2786.2 2746.0 3662.9 2273.5 2623.6 3608.7 3626.5 3175.9 2750.2
#>  [11] 2618.1 2668.3 3661.4 2947.4 3418.9 1811.3 2783.6 3539.7 3714.5 3623.4
#>  [21] 3641.8 2624.7 2718.4 3305.0 3072.8 3223.1 1985.7 2376.2 4590.9 2538.8
#>  [31] 1701.1 3069.0 3602.7 3168.0 4123.0 3842.5 3349.6 2371.1 3020.4 2817.2
#>  [41] 3151.9 3142.3 3291.1 2169.9 3860.6 3883.6 3275.5 2645.1 2883.2 4879.5
#>  [51] 2907.3 2778.3 3070.9 3828.8 2770.4 3547.1 2315.3 4085.2 4061.8 3017.1
#>  [61] 2948.6 3355.7 2897.2 3115.7 4446.8 2282.9 2134.7 1618.2 3850.9 3153.4
#>  [71] 3245.9 3904.1 3236.9 3080.2 4256.5 2424.9 3925.5 3519.6 2648.8 1541.7
#>  [81] 3011.2 1710.2 3035.1 2978.0 1958.9 2228.6 2876.1 2483.4 2947.8 4420.0
#>  [91] 3252.1 1690.0 3231.7 3188.4 3181.2 3471.4 3401.4 2990.4 3470.7 2421.0
#> [101] 3662.8 3299.8 3324.9 3033.8 3751.1 3488.6 3277.1 2153.5 3924.9 2208.6
#> [111] 2304.7 3151.8 3113.4 3013.7 2760.8 3063.1 3163.4 2267.1 3803.9 2042.6
#> [121] 3580.1 3257.6 2854.4 4150.1 2821.7 3788.7 3891.3 4143.1 3810.5 1357.8
#> [131] 1995.2 2504.0 3520.0 2993.0 3326.3 3616.6
#> 
#> $e
#>   [1] 1 0 1 1 0 1 0 0 1 1 0 0 1 0 1 0 0 1 0 1 1 0 1 1 1 0 0 0 1 0 0 1 0 1 0 1 0
#>  [38] 1 1 1 1 1 1 0 0 1 1 0 0 0 1 1 1 1 0 1 0 1 1 1 0 0 0 1 1 0 0 0 1 1 1 1 1 1
#>  [75] 0 0 1 1 0 0 1 0 0 1 0 0 0 0 0 1 1 0 1 1 0 1 1 1 0 1 0 1 0 0 1 1 1 0 1 1 0
#> [112] 1 1 1 0 1 0 0 1 0 0 0 1 1 0 1 0 1 1 0 0 0 1 0 0 1
#>