Make predictions from a sail object
Source: R/methods.R

Similar to other predict methods, this functions predicts fitted values, logits, coefficients and more from a fitted sail object.

# S3 method for sail
predict(object, newx, newe, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), ...)

# S3 method for sail
coef(object, s = NULL, ...)

Arguments

object
Fitted sail model object

newx
matrix of new values for x at which predictions are to be made. Do not include the intercept (this function takes care of that). Must be a matrix. This argument is not used for type=c("coefficients","nonzero"). This matrix will be passed to design_sail to create the design matrix necessary for predictions. This matrix must have the same number of columns originally supplied to the sail fitting function.

newe
vector of new values for the exposure variable e. This is passed to the design_sail function.

s
Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire sequence used to create the model.

type
Type of prediction required. Type "link" gives the linear predictors for "binomial" (not implemented yet); for "gaussian" models it gives the fitted values. Type "response" gives the fitted probabilities for "binomial" (not implemented yet), for "gaussian" type "response" is equivalent to type "link". Type "coefficients" computes the coefficients at the requested values for s. Note that for "binomial" models, results are returned only for the class corresponding to the second level of the factor response (not implemented yet). Type "class" applies only to "binomial" models, and produces the class label corresponding to the maximum probability (not implemented yet). Type "nonzero" returns a list of the the nonzero coefficients for each value of s. Default: "link"

...
currently ignored

Value

The object returned depends on type.

Details

R Source code file for predict, coef, plot and print methods for the sail package Author: Sahir Bhatnagar Created: 2016 Updated: April 6, 2018

s is the new vector at which predictions are requested. If s is not in the lambda sequence used for fitting the model, the predict function will use linear interpolation to make predictions. The new values are interpolated using a fraction of predicted values from both left and right lambda indices. coef(...) is equivalent to predict(sail.object, type="coefficients",...)

See also

Examples

f.basis <- function(i) splines::bs(i, degree = 3)
fit <- sail(x = sailsim$x, y = sailsim$y, e = sailsim$e,
            basis = f.basis, dfmax = 10)
predict(fit) # predicted response for whole solution path

#>              s1       s2       s3       s4       s5       s6       s7       s8
#>   [1,] 5.328505 5.312344 5.296375 5.281383 5.267072 5.253411 5.257267 5.336341
#>   [2,] 5.328505 5.317729 5.307080 5.297083 5.287540 5.278431 5.240673 5.201687
#>   [3,] 5.328505 5.352655 5.376516 5.398919 5.420303 5.440715 5.433341 5.257203
#>   [4,] 5.328505 5.363960 5.398991 5.431881 5.463275 5.493243 5.519670 5.627959
#>   [5,] 5.328505 5.317676 5.306976 5.296930 5.287340 5.278186 5.239282 5.048142
#>   [6,] 5.328505 5.333953 5.339336 5.344390 5.349214 5.353819 5.346847 5.241367
#>   [7,] 5.328505 5.358872 5.388875 5.417045 5.443934 5.469601 5.495679 5.594291
#>   [8,] 5.328505 5.319900 5.311397 5.303414 5.295793 5.288519 5.303351 5.361400
#>   [9,] 5.328505 5.357078 5.385310 5.411816 5.437117 5.461269 5.499097 5.606405
#>  [10,] 5.328505 5.318884 5.309378 5.300453 5.291933 5.283801 5.296838 5.379314
#>  [11,] 5.328505 5.332412 5.336273 5.339897 5.343356 5.346659 5.368181 5.402347
#>  [12,] 5.328505 5.338132 5.347644 5.356575 5.365099 5.373237 5.397153 5.553607
#>  [13,] 5.328505 5.340542 5.352436 5.363602 5.374261 5.384435 5.414028 5.568086
#>  [14,] 5.328505 5.314200 5.300065 5.286794 5.274126 5.262034 5.272316 5.287086
#>  [15,] 5.328505 5.355949 5.383066 5.408525 5.432826 5.456023 5.499289 5.688577
#>  [16,] 5.328505 5.313074 5.297826 5.283511 5.269846 5.256802 5.263811 5.240051
#>  [17,] 5.328505 5.361154 5.393413 5.423700 5.452610 5.480206 5.501163 5.462529
#>  [18,] 5.328505 5.308092 5.287923 5.268986 5.250911 5.233657 5.238227 5.233760
#>  [19,] 5.328505 5.326306 5.324132 5.322091 5.320143 5.318284 5.319029 5.332211
#>  [20,] 5.328505 5.320967 5.313519 5.306527 5.299852 5.293480 5.274577 5.278485
#>  [21,] 5.328505 5.341195 5.353734 5.365506 5.376743 5.387469 5.396371 5.479212
#>  [22,] 5.328505 5.318708 5.309028 5.299939 5.291264 5.282983 5.263635 5.297851
#>  [23,] 5.328505 5.319227 5.310059 5.301452 5.293236 5.285393 5.272310 5.156377
#>  [24,] 5.328505 5.328805 5.329101 5.329379 5.329645 5.329898 5.349519 5.489646
#>  [25,] 5.328505 5.348542 5.368339 5.386926 5.404668 5.421604 5.419804 5.348038
#>  [26,] 5.328505 5.331709 5.334874 5.337845 5.340682 5.343389 5.360307 5.499095
#>  [27,] 5.328505 5.320889 5.313363 5.306297 5.299552 5.293114 5.308511 5.465073
#>  [28,] 5.328505 5.348121 5.367502 5.385698 5.403067 5.419647 5.413928 5.286905
#>  [29,] 5.328505 5.353525 5.378246 5.401456 5.423611 5.444759 5.473971 5.594711
#>              s9      s10      s11      s12      s13      s14      s15      s16
#>   [1,] 5.426858 5.510309 5.586865 5.656783 5.723161 5.787565 5.896755 6.030564
#>   [2,] 5.225322 5.257114 5.296325 5.342278 5.379507 5.403159 5.405705 5.396896
#>   [3,] 5.029636 4.813023 4.607290 4.412178 4.197994 3.955665 3.710442 3.467693
#>   [4,] 5.795454 5.961905 6.126954 6.290339 6.467090 6.661562 6.822368 6.960888
#>   [5,] 4.817274 4.590861 4.368912 4.151380 3.956261 3.788944 3.607057 3.417402
#>   [6,] 5.092418 4.943645 4.795218 4.647246 4.485305 4.305568 4.126551 3.950017
#>   [7,] 5.734586 5.869184 5.998021 6.121156 6.252669 6.397271 6.563601 6.741063
#>   [8,] 5.399757 5.430888 5.455157 5.472945 5.450957 5.379868 5.361580 5.374808
#>   [9,] 5.724615 5.833774 5.934130 6.026023 6.116604 6.208794 6.268718 6.306999
#>  [10,] 5.460952 5.534527 5.600329 5.658712 5.685666 5.674811 5.711208 5.775207
#>  [11,] 5.404355 5.400543 5.391284 5.376929 5.375509 5.392223 5.383781 5.358261
#>  [12,] 5.764458 5.971381 6.174206 6.372871 6.583692 6.811294 7.018353 7.210084
#>  [13,] 5.761345 5.947694 6.127178 6.299955 6.481300 6.675733 6.877225 7.079911
#>  [14,] 5.258966 5.227319 5.192521 5.154878 5.120180 5.090424 5.038645 4.973087
#>  [15,] 5.930445 6.166746 6.397407 6.622464 6.857376 7.106739 7.318375 7.503856
#>  [16,] 5.162559 5.091669 5.027549 4.970150 4.933927 4.922115 4.917844 4.919907
#>  [17,] 5.391552 5.316131 5.236536 5.153025 5.081185 5.026112 4.946990 4.852780
#>  [18,] 5.180779 5.126767 5.072054 5.016862 4.954344 4.882613 4.805992 4.727050
#>  [19,] 5.343717 5.348772 5.347558 5.340325 5.326583 5.307051 5.271642 5.226145
#>  [20,] 5.320878 5.360783 5.398012 5.432506 5.484589 5.560492 5.604095 5.626995
#>  [21,] 5.611929 5.741182 5.866773 5.988624 6.120657 6.266973 6.380726 6.472813
#>  [22,] 5.393256 5.492011 5.593560 5.697426 5.822660 5.974528 6.094815 6.195166
#>  [23,] 4.995773 4.851775 4.724179 4.612493 4.477930 4.308216 4.139350 3.976861
#>  [24,] 5.667390 5.838558 6.003187 6.161427 6.280173 6.350162 6.481753 6.648914
#>  [25,] 5.262191 5.172923 5.080355 4.984650 4.905332 4.848427 4.855849 4.901037
#>  [26,] 5.687800 5.871594 6.050369 6.224130 6.402522 6.588228 6.760732 6.923322
#>  [27,] 5.669723 5.869004 6.062862 6.251354 6.436717 6.620087 6.831093 7.056554
#>  [28,] 5.121449 4.956467 4.792089 4.628400 4.429551 4.185298 3.954905 3.735676
#>  [29,] 5.754328 5.907229 6.053418 6.193022 6.340951 6.501807 6.647567 6.782083
#>             s17      s18       s19       s20       s21       s22       s23
#>   [1,] 6.180978 6.336712  5.682393  5.810664  5.934827  6.023231  6.094509
#>   [2,] 5.416122 5.450096  3.410125  3.264560  3.192416  3.166728  3.028263
#>   [3,] 3.264075 3.085070  2.878374  2.700081  2.514198  2.261798  2.005866
#>   [4,] 7.037222 7.078686 13.532976 13.772641 13.880648 13.925343 14.227380
#>   [5,] 3.238589 3.067803  3.480978  3.410103  3.319438  3.241102  3.233211
#>   [6,] 3.778644 3.612835  3.463331  3.339944  3.214023  3.084117  2.959596
#>   [7,] 6.930098 7.121076 10.795490 11.063548 11.267033 11.421284 11.710636
#>   [8,] 5.388365 5.401987  6.083227  6.150374  6.204346  6.245384  6.297172
#>   [9,] 6.318365 6.315192  7.044633  7.049728  7.059763  7.116584  7.179371
#>  [10,] 5.781190 5.756676  6.058373  6.066853  6.058756  6.048564  6.027446
#>  [11,] 5.326418 5.291218  4.609673  4.608532  4.611159  4.630940  4.643597
#>  [12,] 7.417363 7.628524  8.784838  8.975410  9.161557  9.351848  9.570922
#>  [13,] 7.302056 7.530217  8.608919  8.826211  9.043384  9.273405  9.523323
#>  [14,] 4.881801 4.778892  5.312811  5.326561  5.294180  5.230511  5.240894
#>  [15,] 7.707742 7.917637 11.690603 11.980297 12.239320 12.533167 12.951958
#>  [16,] 4.956556 5.012115  5.395259  5.507381  5.592962  5.633004  5.724357
#>  [17,] 4.782591 4.725275  2.721022  2.584317  2.472589  2.339155  2.140099
#>  [18,] 4.615748 4.489999  5.773226  5.796773  5.759465  5.679135  5.694885
#>  [19,] 5.132434 5.014742  4.791157  4.702476  4.602082  4.511164  4.426196
#>  [20,] 5.622447 5.602803  4.299769  4.217781  4.146475  4.083984  3.992197
#>  [21,] 6.585509 6.706756  8.045350  8.156108  8.237832  8.282253  8.398712
#>  [22,] 6.315138 6.442405  4.105825  4.059911  4.063654  4.076888  3.999947
#>  [23,] 3.817486 3.664311  4.148318  3.993236  3.851619  3.728216  3.587247
#>  [24,] 6.842428 7.046747  7.570764  7.712982  7.874090  8.037422  8.162348
#>  [25,] 4.909483 4.897228  4.875214  4.862111  4.815599  4.710841  4.605641
#>  [26,] 7.029391 7.103807  7.379728  7.424033  7.461441  7.510003  7.549627
#>  [27,] 7.262483 7.454608  6.368230  6.452076  6.554535  6.650556  6.675429
#>  [28,] 3.549933 3.384846  2.717611  2.533411  2.359457  2.149419  1.907391
#>  [29,] 6.916896 7.048930  9.569381  9.755404  9.904738 10.045318 10.277901
#>              s24       s25       s26       s27       s28       s29        s30
#>   [1,]  6.196556  6.288191  6.344576  6.437038  6.498834  6.581521  6.6322837
#>   [2,]  3.100708  3.145608  3.036686  3.112850  3.082053  3.147180  3.1093585
#>   [3,]  1.865524  1.720544  1.499575  1.393606  1.223801  1.124354  0.9579073
#>   [4,] 14.061578 13.981180 14.286961 14.130487 14.263197 14.134710 14.2877834
#>   [5,]  3.105347  2.993688  2.981090  2.848855  2.790463  2.665357  2.6167808
#>   [6,]  2.856334  2.751513  2.633174  2.540017  2.433896  2.345905  2.2452974
#>   [7,] 11.743344 11.814420 12.085663 12.107386 12.275783 12.303224 12.4735629
#>   [8,]  6.317960  6.336371  6.366046  6.379939  6.397581  6.411256  6.4279734
#>   [9,]  7.173025  7.172027  7.219758  7.200599  7.222467  7.205052  7.2306949
#>  [10,]  6.020711  6.009513  5.980615  5.975588  5.954856  5.951246  5.9310347
#>  [11,]  4.638505  4.630341  4.626364  4.612553  4.601456  4.586512  4.5747992
#>  [12,]  9.686458  9.813955 10.015874 10.112709 10.268051 10.359214 10.5076753
#>  [13,]  9.674855  9.831672 10.051786 10.177064 10.351047 10.468371 10.6346610
#>  [14,]  5.129775  5.042208  5.051704  4.954353  4.925968  4.843710  4.8265789
#>  [15,] 13.039966 13.168917 13.548110 13.601403 13.847689 13.904992 14.1560847
#>  [16,]  5.736696  5.763637  5.848286  5.868109  5.925660  5.952490  6.0121591
#>  [17,]  2.098189  2.035292  1.849499  1.819894  1.695246  1.661141  1.5307267
#>  [18,]  5.560558  5.456421  5.472631  5.357490  5.326948  5.230694  5.2135745
#>  [19,]  4.338487  4.255015  4.179737  4.101711  4.030615  3.959249  3.8949090
#>  [20,]  3.966898  3.937276  3.867301  3.848751  3.804266  3.783953  3.7389647
#>  [21,]  8.391082  8.408884  8.533159  8.528228  8.603941  8.603869  8.6814437
#>  [22,]  4.094463  4.167737  4.116064  4.212218  4.221003  4.304394  4.3021470
#>  [23,]  3.487795  3.386658  3.257381  3.168888  3.061195  2.977380  2.8741759
#>  [24,]  8.319410  8.461060  8.560882  8.697374  8.801480  8.925014  9.0167139
#>  [25,]  4.563377  4.514259  4.417181  4.386001  4.311557  4.277938  4.1984422
#>  [26,]  7.583744  7.622184  7.668725  7.704060  7.749501  7.785251  7.8302956
#>  [27,]  6.813717  6.933311  6.957990  7.091619  7.155618  7.277841  7.3296776
#>  [28,]  1.801458  1.682313  1.462429  1.381891  1.219710  1.142383  0.9811978
#>  [29,] 10.307984 10.367418 10.584595 10.602733 10.739991 10.763262 10.9041025
#>               s31        s32        s33        s34
#>   [1,]  6.6773694  6.6276377  6.6042698  6.5519353
#>   [2,]  3.1837350  3.1809741  3.2629330  3.2691556
#>   [3,]  0.8891931  0.7874287  0.7536297  0.6514772
#>   [4,] 14.1648351 14.2713112 14.1402589 14.2316192
#>   [5,]  2.4823082  2.3986513  2.2535454  2.1741410
#>   [6,]  2.1712562  2.1101186  2.0631627  2.0050580
#>   [7,] 12.4762303 12.5955157 12.5854374 12.6988559
#>   [8,]  6.4694430  6.5423684  6.6095394  6.6691361
#>   [9,]  7.2369295  7.2877723  7.2976564  7.3405050
#>  [10,]  5.9223407  5.9029580  5.8990097  5.8789133
#>  [11,]  4.5452148  4.5126811  4.4775952  4.4474295
#>  [12,] 10.5752683 10.6893618 10.7447451 10.8503343
#>  [13,] 10.7330517 10.8832299 10.9787424 11.1172620
#>  [14,]  4.7336812  4.6657494  4.5525855  4.4915035
#>  [15,] 14.1819537 14.3208327 14.2886758 14.4156095
#>  [16,]  6.0659235  6.1628885  6.2321741  6.3181074
#>  [17,]  1.4889958  1.3283362  1.2567846  1.1063897
#>  [18,]  5.1416934  5.1449637  5.0864456  5.0813680
#>  [19,]  3.8168950  3.7314660  3.6456845  3.5697779
#>  [20,]  3.7358606  3.6950706  3.6758764  3.6384234
#>  [21,]  8.6618665  8.6689769  8.6109987  8.6174673
#>  [22,]  4.3833252  4.3801826  4.4451934  4.4474455
#>  [23,]  2.7799765  2.6589003  2.5631392  2.4547659
#>  [24,]  9.1509110  9.2913636  9.4471679  9.5736759
#>  [25,]  4.1515269  4.0449605  3.9788386  3.8803785
#>  [26,]  7.8840722  7.9852281  8.0736086  8.1728372
#>  [27,]  7.4212726  7.4354697  7.5058441  7.5246885
#>  [28,]  0.8964787  0.7194643  0.6242270  0.4598978
#>  [29,] 10.9166880 11.0278854 11.0333274 11.1391914
#>  [ reached getOption("max.print") -- omitted 71 rows ]

predict(fit, s = 0.45) # predicted response for a single lambda value

#>                1
#>   [1,]  6.326807
#>   [2,]  3.071012
#>   [3,]  1.569213
#>   [4,] 14.190595
#>   [5,]  2.985060
#>   [6,]  2.670469
#>   [7,] 12.000181
#>   [8,]  6.356694
#>   [9,]  7.204716
#>  [10,]  5.989722
#>  [11,]  4.627617
#>  [12,]  9.952239
#>  [13,]  9.982418
#>  [14,]  5.048711
#>  [15,] 13.428608
#>  [16,]  5.821609
#>  [17,]  1.908051
#>  [18,]  5.467523
#>  [19,]  4.203460
#>  [20,]  3.889353
#>  [21,]  8.493994
#>  [22,]  4.132349
#>  [23,]  3.298123
#>  [24,]  8.529423
#>  [25,]  4.447775
#>  [26,]  7.654058
#>  [27,]  6.950212
#>  [28,]  1.531725
#>  [29,] 10.516152
#>  [30,]  2.714016
#>  [31,]  5.308843
#>  [32,]  7.673281
#>  [33,]  7.426926
#>  [34,]  2.734772
#>  [35,] 14.371333
#>  [36,]  5.051290
#>  [37,]  5.497576
#>  [38,]  4.116825
#>  [39,]  4.798834
#>  [40,]  5.627460
#>  [41,]  2.787194
#>  [42,]  2.665108
#>  [43,]  5.724695
#>  [44,] 12.829355
#>  [45,]  2.628145
#>  [46,]  4.810065
#>  [47,]  6.029279
#>  [48,] 13.580470
#>  [49,]  3.750771
#>  [50,]  7.296754
#>  [51,]  5.098049
#>  [52,]  2.966294
#>  [53,]  6.301266
#>  [54,]  6.810312
#>  [55,]  2.440902
#>  [56,]  3.173088
#>  [57,]  4.856200
#>  [58,]  6.594824
#>  [59,]  7.214932
#>  [60,]  3.982319
#>  [61,]  2.319107
#>  [62,]  1.651061
#>  [63,]  9.302222
#>  [64,]  3.046891
#>  [65,]  4.660780
#>  [66,]  4.311627
#>  [67,]  4.125699
#>  [68,]  5.711979
#>  [69,]  7.380162
#>  [70,]  7.179483
#>  [71,]  4.118144
#>  [72,]  2.390327
#>  [73,]  7.242306
#>  [74,]  4.231757
#>  [75,]  3.678513
#>  [76,]  3.134130
#>  [77,]  3.325025
#>  [78,] 14.179218
#>  [79,]  6.517843
#>  [80,]  1.457899
#>  [81,]  5.152173
#>  [82,]  4.433921
#>  [83,]  4.532760
#>  [84,]  5.426843
#>  [85,]  4.402068
#>  [86,]  2.849786
#>  [87,]  4.366844
#>  [88,]  4.226206
#>  [89,]  4.419905
#>  [90,]  4.641473
#>  [91,]  5.436960
#>  [92,]  8.437918
#>  [93,]  5.192923
#>  [94,]  6.646552
#>  [95,]  5.812423
#>  [96,]  3.467148
#>  [97,]  6.404750
#>  [98,]  3.590069
#>  [99,]  5.025222
#> [100,]  4.869444

predict(fit, s = c(2.15, 0.32, 0.40), type="nonzero") # nonzero coefficients

#> [[1]]
#>                    1
#> (Intercept) 5.328505
#> 
#> [[2]]
#>                         2
#> (Intercept)  5.567903e+00
#> X1_1        -1.785340e-01
#> X1_2         9.057618e-01
#> X1_3         2.638720e+00
#> X2_1        -4.265388e-01
#> X2_2        -3.280625e-01
#> X2_3        -1.054229e-01
#> X3_1         4.110255e+00
#> X3_2        -3.066381e+00
#> X3_3        -2.400295e+00
#> X4_1        -3.206033e-01
#> X4_2        -9.986955e+00
#> X4_3        -2.032217e+00
#> X8_1         6.479998e-01
#> X8_2        -1.163753e-01
#> X8_3        -9.716133e-01
#> X11_1        1.815156e-01
#> X11_2       -6.387968e-01
#> X11_3       -1.184713e+00
#> X14_1        1.740984e-02
#> X14_2       -4.367701e-02
#> X14_3       -1.069606e-01
#> E            2.055010e+00
#> X1_1:E      -5.230371e-06
#> X1_2:E       3.289542e-05
#> X1_3:E       9.392354e-05
#> X3_1:E       1.905874e+00
#> X3_2:E      -1.421839e+00
#> X3_3:E      -1.112969e+00
#> X4_1:E      -6.196339e-01
#> X4_2:E      -1.930188e+01
#> X4_3:E      -3.927686e+00
#> 
#> [[3]]
#>                        3
#> (Intercept)   5.56680738
#> X1_1         -0.27015072
#> X1_2          0.69315972
#> X1_3          2.26877573
#> X3_1          3.25343103
#> X3_2         -2.48078286
#> X3_3         -2.40746822
#> X4_1         -0.35564789
#> X4_2         -9.01034518
#> X4_3         -1.96408443
#> X8_1          0.46818495
#> X8_2         -0.08871758
#> X8_3         -0.74277147
#> X11_1         0.17070499
#> X11_2        -0.47761895
#> X11_3        -0.88108041
#> E             2.03170825
#> X3_1:E        1.33382715
#> X3_2:E       -1.01705456
#> X3_3:E       -0.98683438
#> X4_1:E       -0.74899277
#> X4_2:E      -18.97431236
#> X4_3:E       -4.13642287
#>