Make predictions from a sail object
Source: R/methods.R

Similar to other predict methods, this functions predicts fitted values, logits, coefficients and more from a fitted sail object.

# S3 method for sail
predict(object, newx, newe, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), ...)

# S3 method for sail
coef(object, s = NULL, ...)

Arguments

object
Fitted sail model object

newx
matrix of new values for x at which predictions are to be made. Do not include the intercept (this function takes care of that). Must be a matrix. This argument is not used for type=c("coefficients","nonzero"). This matrix will be passed to design_sail to create the design matrix necessary for predictions. This matrix must have the same number of columns originally supplied to the sail fitting function.

newe
vector of new values for the exposure variable e. This is passed to the design_sail function.

s
Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire sequence used to create the model.

type
Type of prediction required. Type "link" gives the linear predictors for "binomial" (not implemented yet); for "gaussian" models it gives the fitted values. Type "response" gives the fitted probabilities for "binomial" (not implemented yet), for "gaussian" type "response" is equivalent to type "link". Type "coefficients" computes the coefficients at the requested values for s. Note that for "binomial" models, results are returned only for the class corresponding to the second level of the factor response (not implemented yet). Type "class" applies only to "binomial" models, and produces the class label corresponding to the maximum probability (not implemented yet). Type "nonzero" returns a list of the the nonzero coefficients for each value of s. Default: "link"

...
currently ignored

Value

The object returned depends on type.

Details

R Source code file for predict, coef, plot and print methods for the sail package Author: Sahir Bhatnagar Created: 2016 Updated: April 6, 2018

s is the new vector at which predictions are requested. If s is not in the lambda sequence used for fitting the model, the predict function will use linear interpolation to make predictions. The new values are interpolated using a fraction of predicted values from both left and right lambda indices. coef(...) is equivalent to predict(sail.object, type="coefficients",...)

See also

Examples

f.basis <- function(i) splines::bs(i, degree = 5)
fit <- sail(x = sailsim$x, y = sailsim$y, e = sailsim$e,
            basis = f.basis, dfmax = 10)
predict(fit) # predicted response for whole solution path

#>              s1       s2       s3       s4       s5       s6       s7       s8
#>   [1,] 5.218785 5.201827 5.185073 5.169343 5.154328 5.149904 5.204258 5.282486
#>   [2,] 5.218785 5.207477 5.196305 5.185816 5.175803 5.190410 5.194024 5.144528
#>   [3,] 5.218785 5.244124 5.269159 5.292664 5.315100 5.295813 5.120594 4.907344
#>   [4,] 5.218785 5.255986 5.292741 5.327248 5.360187 5.440547 5.619727 5.780001
#>   [5,] 5.218785 5.207422 5.196195 5.185655 5.175594 5.120524 4.888444 4.608860
#>   [6,] 5.218785 5.224501 5.230149 5.235451 5.240513 5.199369 5.013307 4.803999
#>   [7,] 5.218785 5.250647 5.282127 5.311683 5.339894 5.414983 5.594250 5.758418
#>   [8,] 5.218785 5.209755 5.200834 5.192458 5.184463 5.150988 5.110716 5.113172
#>   [9,] 5.218785 5.248766 5.278386 5.306196 5.332742 5.368775 5.457901 5.564134
#>  [10,] 5.218785 5.208690 5.198716 5.189351 5.180413 5.167315 5.192590 5.253689
#>  [11,] 5.218785 5.222884 5.226934 5.230737 5.234367 5.200207 5.120638 5.077488
#>  [12,] 5.218785 5.228886 5.238866 5.248236 5.257180 5.322657 5.557330 5.811795
#>  [13,] 5.218785 5.231415 5.243894 5.255609 5.266792 5.322744 5.532833 5.772514
#>  [14,] 5.218785 5.203774 5.188944 5.175020 5.161729 5.103896 4.997172 4.939712
#>  [15,] 5.218785 5.247581 5.276032 5.302743 5.328240 5.409339 5.677229 5.978504
#>  [16,] 5.218785 5.202593 5.186595 5.171575 5.157238 5.105188 5.018392 4.981941
#>  [17,] 5.218785 5.253042 5.286888 5.318665 5.348997 5.340081 5.216106 5.074634
#>  [18,] 5.218785 5.197366 5.176204 5.156336 5.137371 5.071958 4.951951 4.881991
#>  [19,] 5.218785 5.216477 5.214196 5.212055 5.210011 5.202122 5.173016 5.136042
#>  [20,] 5.218785 5.210876 5.203061 5.195724 5.188721 5.224659 5.317920 5.371147
#>  [21,] 5.218785 5.232100 5.245256 5.257607 5.269396 5.335806 5.516517 5.677124
#>  [22,] 5.218785 5.208505 5.198349 5.188813 5.179711 5.209575 5.293752 5.345758
#>  [23,] 5.218785 5.209049 5.199430 5.190400 5.181779 5.152197 5.059803 4.964747
#>  [24,] 5.218785 5.219099 5.219410 5.219702 5.219980 5.263461 5.454394 5.673762
#>  [25,] 5.218785 5.239808 5.260580 5.280081 5.298696 5.294832 5.185599 5.034808
#>  [26,] 5.218785 5.222146 5.225467 5.228584 5.231560 5.289930 5.508245 5.741080
#>  [27,] 5.218785 5.210793 5.202897 5.195483 5.188407 5.235724 5.464248 5.726777
#>              s9      s10      s11      s12      s13      s14      s15      s16
#>   [1,] 5.353093 5.416656 5.473756 5.524935 5.570722 5.611567 5.647910 5.680144
#>   [2,] 5.100839 5.062304 5.028300 4.998283 4.971701 4.948210 4.927359 4.908798
#>   [3,] 4.708167 4.522230 4.348671 4.186657 4.035372 3.894018 3.761865 3.638268
#>   [4,] 5.935958 6.087335 6.233922 6.375574 6.512130 6.643694 6.770197 6.891632
#>   [5,] 4.338843 4.078447 3.827666 3.586466 3.354790 3.132419 2.919221 2.715066
#>   [6,] 4.599947 4.401426 4.208646 4.021774 3.840952 3.666135 3.497374 3.334705
#>   [7,] 5.914368 6.062240 6.202222 6.334545 6.459417 6.577234 6.688237 6.792696
#>   [8,] 5.110062 5.102251 5.090563 5.075726 5.058467 5.039236 5.018607 4.997032
#>   [9,] 5.660627 5.748112 5.827320 5.898945 5.963670 6.022086 6.074787 6.122300
#>  [10,] 5.306803 5.352737 5.392275 5.426141 5.455052 5.479546 5.500221 5.517578
#>  [11,] 5.031480 4.983371 4.933861 4.883553 4.833038 4.782621 4.732735 4.683713
#>  [12,] 6.058262 6.296644 6.526907 6.749063 6.963110 7.169284 7.367645 7.558276
#>  [13,] 6.001784 6.220933 6.430295 6.630209 6.821010 7.003142 7.176939 7.342710
#>  [14,] 4.881986 4.824598 4.768093 4.712915 4.659511 4.608038 4.558812 4.512037
#>  [15,] 6.269562 6.550481 6.821398 7.082476 7.333872 7.575946 7.808907 8.032951
#>  [16,] 4.953115 4.931328 4.915979 4.906446 4.902165 4.902512 4.907003 4.915146
#>  [17,] 4.932939 4.791657 4.651361 4.512567 4.375760 4.241199 4.109236 3.980179
#>  [18,] 4.814369 4.749434 4.687476 4.628697 4.573308 4.521267 4.472689 4.427597
#>  [19,] 5.094035 5.047643 4.997492 4.944177 4.888269 4.830214 4.770483 4.709506
#>  [20,] 5.420341 5.465486 5.506603 5.543776 5.577058 5.606699 5.632790 5.655483
#>  [21,] 5.831382 5.979203 6.120552 6.255462 6.383942 6.506261 6.622498 6.732787
#>  [22,] 5.398443 5.451323 5.503958 5.555984 5.607027 5.656964 5.705524 5.752523
#>  [23,] 4.884256 4.816558 4.759922 4.712718 4.673408 4.640673 4.613265 4.590105
#>  [24,] 5.883303 6.083302 6.274081 6.455974 6.629303 6.794496 6.951874 7.101734
#>  [25,] 4.883750 4.732959 4.582926 4.434114 4.286949 4.141710 3.998708 3.858247
#>  [26,] 5.965241 6.180768 6.387762 6.586364 6.776706 6.959139 7.133852 7.301044
#>  [27,] 5.979714 6.223176 6.457340 6.682405 6.898571 7.106212 7.305572 7.496872
#>              s17       s18       s19       s20       s21        s22       s23
#>   [1,]  5.198560  5.322553  5.439881  5.531408  5.578070  5.6383803  5.690112
#>   [2,]  3.501793  3.397213  3.303233  3.254078  3.216763  3.2080637  3.175126
#>   [3,]  3.557176  3.448515  3.343997  3.140207  2.871249  2.6838893  2.514866
#>   [4,] 12.835062 12.924430 13.002644 12.949153 12.965302 12.9081235 12.965649
#>   [5,]  3.596626  3.450630  3.304109  3.163373  3.089778  2.9817557  2.899404
#>   [6,]  3.085674  2.977397  2.871778  2.745599  2.616196  2.5085156  2.408738
#>   [7,] 11.314672 11.461915 11.597280 11.638487 11.723471 11.7565034 11.863628
#>   [8,]  5.668327  5.797384  5.915465  6.008382  6.087040  6.1433426  6.207731
#>   [9,]  6.836952  6.817776  6.797795  6.836809  6.923481  6.9584146  6.988089
#>  [10,]  5.644084  5.773884  5.895283  5.993442  6.064736  6.1293941  6.194314
#>  [11,]  4.244846  4.206191  4.166625  4.170604  4.189285  4.1799454  4.158655
#>  [12,]  8.998459  9.112558  9.223938  9.355098  9.520210  9.6417451  9.772396
#>  [13,]  8.830885  8.981202  9.125008  9.308739  9.535111  9.7027154  9.870218
#>  [14,]  5.388124  5.384076  5.374590  5.318791  5.262474  5.1780275  5.129489
#>  [15,] 12.565951 12.720817 12.864581 13.058900 13.392996 13.5783879 13.810890
#>  [16,]  5.125072  5.154461  5.182033  5.178699  5.153001  5.1291735  5.117979
#>  [17,]  1.656148  1.480060  1.313533  1.135310  0.880943  0.7083192  0.504275
#>  [18,]  5.782503  5.810691  5.831607  5.781966  5.725907  5.6456482  5.611666
#>  [19,]  4.477010  4.423365  4.371011  4.312667  4.261095  4.2096062  4.165947
#>  [20,]  4.222510  4.138380  4.062932  4.004639  3.939847  3.9066460  3.858254
#>  [21,]  8.513635  8.531895  8.548985  8.509310  8.482699  8.4453540  8.447941
#>  [22,]  3.940702  3.860456  3.790727  3.751739  3.698082  3.6855504  3.645773
#>  [23,]  4.142892  4.066561  3.994433  3.929334  3.864687  3.8150962  3.759368
#>  [24,]  7.316691  7.526322  7.728157  7.941499  8.131126  8.3123029  8.471819
#>  [25,]  3.618044  3.608843  3.597108  3.514591  3.368285  3.2887649  3.205594
#>  [26,]  7.495920  7.583553  7.672158  7.767468  7.860227  7.9476432  8.033621
#>  [27,]  6.063182  6.214459  6.365625  6.528808  6.630443  6.7643872  6.861944
#>               s24        s25        s26         s27          s28        s29
#>   [1,]  5.7502681  5.8191722  5.9072353  5.99106019  6.070710155  6.1059582
#>   [2,]  3.1670234  3.1351133  3.1166686  3.10025840  3.086289467  3.0645267
#>   [3,]  2.3760596  2.2306010  2.1174141  2.01027076  1.908605318  1.7735360
#>   [4,] 12.9339142 12.9596875 12.9506193 12.94384918 12.937777348 12.9338711
#>   [5,]  2.7875047  2.7034196  2.6203563  2.53742971  2.454396555  2.4046436
#>   [6,]  2.3162455  2.2118509  2.1111666  2.01543720  1.924523383  1.8345511
#>   [7,] 11.9076611 12.0081331 12.1268264 12.24153880 12.351473524 12.5634305
#>   [8,]  6.2526634  6.2606164  6.2711254  6.28039570  6.288387733  6.2850710
#>   [9,]  6.9961824  6.9947829  6.9724764  6.94944458  6.925961386  6.8698827
#>  [10,]  6.2517708  6.2665655  6.2476798  6.22950097  6.211973794  6.1764302
#>  [11,]  4.1335804  4.1209727  4.1024201  4.08313582  4.063326295  4.0325662
#>  [12,]  9.8785380 10.0129038 10.1600709 10.30137144 10.436884189 10.5839343
#>  [13,] 10.0079054 10.1649702 10.3341861 10.49532785 10.648680014 10.8332819
#>  [14,]  5.0550491  4.9937110  4.9190518  4.84862015  4.781742101  4.7444749
#>  [15,] 13.9466434 14.1112831 14.2811737 14.44210911 14.594008598 14.7515703
#>  [16,]  5.1069439  5.1290875  5.1792304  5.22990084  5.280778182  5.3354497
#>  [17,]  0.3608408  0.2450644  0.1613365  0.08129381  0.005058774 -0.1506930
#>  [18,]  5.5481547  5.4863952  5.4082901  5.33554821  5.267310496  5.2235777
#>  [19,]  4.1195990  4.0617089  3.9727403  3.88810218  3.807631026  3.7331049
#>  [20,]  3.8343874  3.8235133  3.7884866  3.75526035  3.724019237  3.6872899
#>  [21,]  8.4266807  8.4472929  8.4829738  8.51796710  8.551732088  8.5879297
#>  [22,]  3.6416225  3.6629209  3.7028135  3.74186246  3.780420690  3.8089802
#>  [23,]  3.7106795  3.6096948  3.5010769  3.39537020  3.292634452  3.1961095
#>  [24,]  8.6296342  8.7440190  8.8758760  9.00178768  9.122358315  9.2420734
#>  [25,]  3.1545185  3.1290023  3.0842527  3.04068784  2.998282754  3.0132423
#>  [26,]  8.1191236  8.1993230  8.2399750  8.28128087  8.323250459  8.3703917
#>  [27,]  6.9919644  7.1063722  7.2071278  7.30518691  7.400920291  7.4650043
#>               s30        s31        s32        s33         s34        s35
#>   [1,]  6.1238268  6.1301291  6.1166002  6.0733863  6.04657307  6.0030962
#>   [2,]  3.0777546  3.0674803  3.0980307  3.1120432  3.14977960  3.1668520
#>   [3,]  1.6555333  1.5222352  1.4623857  1.3974413  1.36190910  1.3030212
#>   [4,] 12.7843425 12.7695531 12.6494092 12.6537559 12.55159755 12.5607045
#>   [5,]  2.3223365  2.2809515  2.1643953  2.0778244  1.95372298  1.8733214
#>   [6,]  1.7538153  1.6712490  1.6109592  1.5542313  1.50764179  1.4599417
#>   [7,] 12.7225256 12.9365803 13.0148859 13.1474938 13.19479778 13.3140491
#>   [8,]  6.2564853  6.2321602  6.2391405  6.2664994  6.28787330  6.3128277
#>   [9,]  6.7767250  6.7151788  6.6920338  6.7082409  6.70644303  6.7192495
#>  [10,]  6.1298326  6.0786482  6.0296546  5.9734657  5.92773463  5.8764102
#>  [11,]  3.9951457  3.9648713  3.9061719  3.8399231  3.77683656  3.7180126
#>  [12,] 10.7020490 10.8452699 10.9385402 11.0485005 11.12618106 11.2260532
#>  [13,] 10.9945515 11.1706930 11.3066917 11.4608219 11.57794958 11.7144352
#>  [14,]  4.6883807  4.6639308  4.5739334  4.4992276  4.40136708  4.3341332
#>  [15,] 14.7782439 14.9182039 14.9111246 14.9925091 14.97331110 15.0521528
#>  [16,]  5.3873611  5.4402204  5.5227657  5.6183375  5.71036926  5.8006276
#>  [17,] -0.2702262 -0.4309370 -0.5249314 -0.6737035 -0.75354765 -0.8872935
#>  [18,]  5.1528255  5.1148122  5.0853572  5.0982235  5.07878652  5.0838916
#>  [19,]  3.6592280  3.5939146  3.5067304  3.4177425  3.32928636  3.2480666
#>  [20,]  3.6761757  3.6516545  3.6861907  3.7221298  3.76796800  3.7983396
#>  [21,]  8.5804766  8.6163795  8.5884907  8.5890740  8.55481238  8.5564423
#>  [22,]  3.8758663  3.9140646  3.9993501  4.0622067  4.14577449  4.2004487
#>  [23,]  3.1271602  3.0451289  2.9475082  2.8250322  2.72049518  2.6063810
#>  [24,]  9.3698916  9.4675933  9.6013605  9.7245572  9.85579157  9.9665939
#>  [25,]  3.0849813  3.0996737  3.1124963  3.0803709  3.06899164  3.0276838
#>  [26,]  8.4224599  8.4753070  8.5490685  8.6342427  8.71720234  8.8015019
#>  [27,]  7.5632393  7.6148685  7.6745046  7.6738877  7.71689924  7.7153246
#>               s36
#>   [1,]  5.9499450
#>   [2,]  3.1971976
#>   [3,]  1.2512183
#>   [4,] 12.5585521
#>   [5,]  1.7954494
#>   [6,]  1.4302618
#>   [7,] 13.3882282
#>   [8,]  6.3555209
#>   [9,]  6.6938839
#>  [10,]  5.8089725
#>  [11,]  3.6813056
#>  [12,] 11.3278683
#>  [13,] 11.8382137
#>  [14,]  4.2377759
#>  [15,] 15.0723372
#>  [16,]  5.8855458
#>  [17,] -0.9764130
#>  [18,]  5.0557461
#>  [19,]  3.1962446
#>  [20,]  3.8171421
#>  [21,]  8.5357325
#>  [22,]  4.2885632
#>  [23,]  2.4884749
#>  [24,] 10.0864326
#>  [25,]  2.9801621
#>  [26,]  8.8465933
#>  [27,]  7.7603623
#>  [ reached getOption("max.print") -- omitted 73 rows ]

predict(fit, s = 0.45) # predicted response for a single lambda value

#>                 1
#>   [1,]  5.9673302
#>   [2,]  3.1049040
#>   [3,]  2.0406020
#>   [4,] 12.9457657
#>   [5,]  2.5609054
#>   [6,]  2.0425373
#>   [7,] 12.2090648
#>   [8,]  6.2777714
#>   [9,]  6.9559647
#>  [10,]  6.2346472
#>  [11,]  4.0885950
#>  [12,] 10.2613706
#>  [13,] 10.4497102
#>  [14,]  4.8685587
#>  [15,] 14.3965499
#>  [16,]  5.2155565
#>  [17,]  0.1039531
#>  [18,]  5.3561407
#>  [19,]  3.9120624
#>  [20,]  3.7646664
#>  [21,]  8.5080609
#>  [22,]  3.7308081
#>  [23,]  3.4252948
#>  [24,]  8.9661433
#>  [25,]  3.0530206
#>  [26,]  8.2695876
#>  [27,]  7.2774273
#>  [28,]  0.6069349
#>  [29,] 10.9422711
#>  [30,]  3.1215912
#>  [31,]  5.5472516
#>  [32,]  7.2761855
#>  [33,]  7.2225974
#>  [34,]  2.8208377
#>  [35,] 15.3952405
#>  [36,]  3.8305990
#>  [37,]  5.5692762
#>  [38,]  3.2742106
#>  [39,]  4.7834290
#>  [40,]  5.9606372
#>  [41,]  3.6930737
#>  [42,]  1.9975338
#>  [43,]  5.1962678
#>  [44,] 14.1181516
#>  [45,]  2.9703602
#>  [46,]  4.9564295
#>  [47,]  5.5903901
#>  [48,] 14.3625990
#>  [49,]  3.2300620
#>  [50,]  6.6860504
#>  [51,]  5.2866827
#>  [52,]  2.8915055
#>  [53,]  7.2923740
#>  [54,]  7.0052334
#>  [55,]  2.4800241
#>  [56,]  1.6810682
#>  [57,]  4.3981578
#>  [58,]  7.2815998
#>  [59,]  7.2839412
#>  [60,]  3.8163757
#>  [61,]  2.4338554
#>  [62,]  1.7739611
#>  [63,]  8.8410993
#>  [64,]  3.0398599
#>  [65,]  4.1841543
#>  [66,]  4.7443962
#>  [67,]  4.6778625
#>  [68,]  5.8730607
#>  [69,]  7.9891444
#>  [70,]  7.0381016
#>  [71,]  3.9652579
#>  [72,]  2.3176202
#>  [73,]  6.5322176
#>  [74,]  5.0524293
#>  [75,]  3.1110196
#>  [76,]  3.3361725
#>  [77,]  2.9802147
#>  [78,] 10.3236908
#>  [79,]  6.1715216
#>  [80,]  2.3401505
#>  [81,]  4.8713425
#>  [82,]  4.1001736
#>  [83,]  3.5676141
#>  [84,]  5.5466083
#>  [85,]  5.4093080
#>  [86,]  1.9181951
#>  [87,]  3.7973994
#>  [88,]  3.1304788
#>  [89,]  4.3321610
#>  [90,]  4.9917768
#>  [91,]  4.5832952
#>  [92,]  8.9661252
#>  [93,]  5.0035549
#>  [94,]  6.4896425
#>  [95,]  5.6725366
#>  [96,]  3.3864414
#>  [97,]  6.7593506
#>  [98,]  3.7087113
#>  [99,]  4.4293731
#> [100,]  5.2384184

predict(fit, s = c(2.15, 0.32, 0.40), type="nonzero") # nonzero coefficients

#> [[1]]
#>                    1
#> (Intercept) 5.218785
#> 
#> [[2]]
#>                        2
#> (Intercept)   5.43273436
#> X1_1         -1.13416427
#> X1_2          0.35963402
#> X1_3          0.75400639
#> X1_4          1.03258622
#> X1_5          2.48105815
#> X2_1         -0.11829376
#> X2_2         -0.48895824
#> X2_3         -0.38891641
#> X2_4         -0.08634648
#> X2_5         -0.02015260
#> X3_1          4.15082632
#> X3_2          2.06155613
#> X3_3         -1.35859812
#> X3_4         -2.51996947
#> X3_5         -1.12096397
#> X4_1          5.08504419
#> X4_2         -3.23311409
#> X4_3         -7.63199308
#> X4_4         -3.53921362
#> X4_5         -0.51629061
#> X8_1          0.38999497
#> X8_2          0.26635389
#> X8_3          0.05017593
#> X8_4         -0.22972941
#> X8_5         -0.54499349
#> X11_1         0.30715164
#> X11_2         0.05511638
#> X11_3        -0.17773018
#> X11_4        -0.55016367
#> X11_5        -0.56668826
#> E             2.04388079
#> X1_1:E       -0.40554050
#> X1_2:E        0.12863559
#> X1_3:E        0.26963743
#> X1_4:E        0.36921214
#> X1_5:E        0.88713279
#> X3_1:E        2.24128302
#> X3_2:E        1.11316550
#> X3_3:E       -0.73358326
#> X3_4:E       -1.36067658
#> X3_5:E       -0.60526384
#> X4_1:E        8.66157069
#> X4_2:E       -5.50710626
#> X4_3:E      -12.99989747
#> X4_4:E       -6.02851789
#> X4_5:E       -0.87942094
#> 
#> [[3]]
#>                        3
#> (Intercept)   5.42476209
#> X1_1         -0.86999759
#> X1_2          0.17590128
#> X1_3          0.50737552
#> X1_4          0.85765989
#> X1_5          2.06926791
#> X3_1          3.61153587
#> X3_2          1.67474172
#> X3_3         -1.29749034
#> X3_4         -2.34271650
#> X3_5         -1.22518668
#> X4_1          4.58763038
#> X4_2         -2.95668713
#> X4_3         -6.85307630
#> X4_4         -3.47331082
#> X4_5         -0.48292544
#> X8_1          0.22355987
#> X8_2          0.14896198
#> X8_3          0.02374242
#> X8_4         -0.13301848
#> X8_5         -0.31359383
#> X11_1         0.18381461
#> X11_2         0.03622276
#> X11_3        -0.10082936
#> X11_4        -0.28964119
#> X11_5        -0.30994252
#> E             2.01540165
#> X1_1:E       -0.08210993
#> X1_2:E        0.01692025
#> X1_3:E        0.04811573
#> X1_4:E        0.08042242
#> X1_5:E        0.19392618
#> X3_1:E        1.92973919
#> X3_2:E        0.89485842
#> X3_3:E       -0.69328380
#> X3_4:E       -1.25177593
#> X3_5:E       -0.65465049
#> X4_1:E        8.44328543
#> X4_2:E       -5.44162867
#> X4_3:E      -12.61257956
#> X4_4:E       -6.39285649
#> X4_5:E       -0.88883177
#>