Print Method for sail object
Source: R/methods.R

Print a summary of the sail path at each step along the path.

# S3 method for sail
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

x
fitted sail object

digits
significant digits in printout. Default: max(3, getOption("digits") - 3)

...
additional print arguments

Value

OUTPUT_DESCRIPTION

Details

The call that produced the object x is printed, followed by a five-column matrix with columns df_main, df_interaction, df_environment, %Dev and Lambda. The df_ columns are the corresponding number of nonzero coefficients for main effects, interactions and exposure, respectively. %dev is the percent deviance explained (relative to the null deviance). For type="gaussian" this is the r-squared.

See also

Examples

if(interactive()){
data("sailsim")
f.basis <- function(i) splines::bs(i, degree = 3)
fit <- sail(x = sailsim$x, y = sailsim$y, e = sailsim$e,
            basis = f.basis)
fit
 }

#> 
#> Call:  sail(x = sailsim$x, y = sailsim$y, e = sailsim$e, basis = f.basis) 
#> 
#>      df_main df_interaction df_environment     %Dev  Lambda
#> s1         0              0              0 0.000000 1.48800
#> s2         0              0              1 0.001698 1.42100
#> s3         0              0              1 0.003357 1.35600
#> s4         0              0              1 0.004897 1.29400
#> s5         0              0              1 0.006352 1.23500
#> s6         0              0              1 0.007726 1.17900
#> s7         0              0              1 0.009026 1.12600
#> s8         0              0              1 0.010250 1.07500
#> s9         1              0              1 0.018960 1.02600
#> s10        2              0              1 0.042450 0.97910
#> s11        2              0              1 0.068960 0.93460
#> s12        2              0              1 0.094080 0.89210
#> s13        2              0              1 0.117900 0.85160
#> s14        2              0              1 0.140400 0.81290
#> s15        3              0              1 0.163800 0.77590
#> s16        4              0              1 0.189000 0.74070
#> s17        4              0              1 0.214000 0.70700
#> s18        5              0              1 0.238100 0.67490
#> s19        5              0              1 0.262200 0.64420
#> s20        3              1              1 0.439000 0.61490
#> s21        4              2              1 0.457000 0.58700
#> s22        5              2              1 0.475400 0.56030
#> s23        5              2              1 0.491600 0.53480
#> s24        5              2              1 0.505300 0.51050
#> s25        5              2              1 0.518200 0.48730
#> s26        5              2              1 0.530100 0.46520
#> s27        5              2              1 0.541100 0.44400
#> s28        5              2              1 0.551400 0.42380
#> s29        5              2              1 0.561300 0.40460
#> s30        5              2              1 0.569300 0.38620
#> s31        5              2              1 0.577300 0.36860
#> s32        6              2              1 0.585500 0.35190
#> s33        6              2              1 0.594900 0.33590
#> s34        6              2              1 0.603700 0.32060
#> s35        7              3              1 0.612300 0.30600
#> s36        7              3              1 0.620400 0.29210
#> s37        8              4              1 0.629200 0.27890
#> s38        8              4              1 0.641900 0.26620
#> s39        9              4              1 0.650000 0.25410
#> s40        9              4              1 0.658500 0.24250
#> s41        8              4              1 0.689000 0.23150
#> s42        8              4              1 0.695300 0.22100
#> s43        8              4              1 0.701400 0.21090
#> s44        9              4              1 0.707200 0.20140
#> s45        8              6              1 0.728900 0.19220
#> s46        8              6              1 0.733700 0.18350
#> s47        8              6              1 0.738200 0.17510
#> s48        8              7              1 0.742400 0.16720
#> s49        9              7              1 0.747600 0.15960
#> s50       10              7              1 0.752100 0.15230
#> s51       10              7              1 0.756500 0.14540
#> s52       10              7              1 0.760900 0.13880
#> s53       11              7              1 0.765200 0.13250
#> s54       11              7              1 0.769900 0.12650
#> s55       11              7              1 0.773900 0.12070
#> s56       11              7              1 0.778100 0.11520
#> s57       12              9              1 0.794000 0.11000
#> s58       12              9              1 0.798500 0.10500
#> s59       12              9              1 0.802700 0.10020
#> s60       12              9              1 0.806600 0.09566
#> s61       13              9              1 0.810400 0.09131
#> s62       13              9              1 0.814100 0.08716
#> s63       13              9              1 0.817200 0.08320
#> s64       16              9              1 0.820000 0.07942
#> s65       16              9              1 0.822900 0.07581
#> s66       16              9              1 0.825600 0.07236
#> s67       16              9              1 0.828200 0.06907
#> s68       16              9              1 0.830700 0.06593
#> s69       17             10              1 0.833500 0.06294
#> s70       17             10              1 0.840000 0.06008
#> s71       16             12              1 0.849000 0.05735
#> s72       16             12              1 0.851400 0.05474
#> s73       18             12              1 0.853900 0.05225
#> s74       19             12              1 0.856700 0.04988
#> s75       19             12              1 0.859600 0.04761
#> s76       19             12              1 0.862600 0.04545
#> s77       19             15              1 0.897700 0.04338
#> s78       20             18              1 0.907600 0.04141
#> s79       20             18              1 0.909800 0.03953
#> s80       20             19              1 0.925800 0.03773
#> s81       20             20              1 0.927700 0.03602
#> s82       20             20              1 0.929700 0.03438
#> s83       20             20              1 0.931800 0.03282
#> s84       20             20              1 0.934500 0.03132
#> s85       20             20              1 0.936100 0.02990
#> s86       20             20              1 0.937700 0.02854
#> s87       20             20              1 0.939300 0.02724
#> s88       20             20              1 0.943900 0.02601
#> s89       20             20              1 0.944700 0.02482
#> s90       20             20              1 0.945600 0.02370
#> s91       20             20              1 0.946500 0.02262
#> s92       20             20              1 0.947100 0.02159
#> s93       20             20              1 0.948500 0.02061
#> s94       20             20              1 0.949200 0.01967
#> s95       20             20              1 0.950100 0.01878
#> s96       20             20              1 0.950700 0.01792
#> s97       20             20              1 0.951400 0.01711
#> s98       20             20              1 0.952000 0.01633
#> s99       20             20              1 0.952600 0.01559
#> s100      20             20              1 0.954900 0.01488