Internal sail functions
Source: R/utils.R

Internal sail helper functions

cbbPalette gives a Color Blind Palette

nonzero is to determine which coefficients are non-zero

check_col_0 is to check how many columns are 0 and is used in the fitting functions lspath

SoftThreshold(x, lambda)

l2norm(x)

cbbPalette

nonzero(beta, bystep = FALSE)

check_col_0(M)

Arguments

x
numeric value of a coefficient

lambda
tuning parameter value

beta
vector or 1 column matrix of regression coefficients

bystep
bystep = FALSE means which variables were ever nonzero. bystep = TRUE means which variables are nonzero for each step

M
is a matrix

Format

An object of class character of length 8.

Details

These functions are not intended for use by users.