Simulated Data Used in Bhatnagar et al. (2018+) Paper
Source: R/data.R

A dataset containing simulated data used in the accompanying paper to this package

sailsim

Format

A list with 7 elements:

x

a matrix of dimension 100 x 20 where rows are observations and columns are predictors

y

a numeric response vector of length 100

e

a numeric exposure vector of length 100

f1,f2,f3,f4

the true functions

Details

The code used to simulate the data is available at https://github.com/sahirbhatnagar/sail/blob/master/data-raw/SIMULATED_data.R. See gendata for more details. The true model is given by $$Y = f1(X1) + f2(X2) + f3(X3) + f4(X4) + E * (2 + f3(X3) + f4(X4))$$ where

f2(t)=3(2t - 1)^2

f3(t)= 4sin(2pi*t) / (2-sin(2pi*t)

f4(t)=6(0.1sin(2pi*t) + 0.2cos(2pi*t) + 0.3sin(2pi*t)^2 + 0.4cos(2pi*t)^3 + 0.5sin(2pi*t)^3)

References

Lin, Y., & Zhang, H. H. (2006). Component selection and smoothing in multivariate nonparametric regression. The Annals of Statistics, 34(5), 2272-2297.

Huang J, Horowitz JL, Wei F. Variable selection in nonparametric additive models (2010). Annals of statistics. Aug 1;38(4):2282.

Bhatnagar SR, Yang Y, Greenwood CMT. Sparse additive interaction models with the strong heredity property (2018+). Preprint.

Examples

sailsim

#> $x
#>                 X1          X2          X3          X4          X5          X6
#>   [1,] 0.918269079 0.558447814 0.093077163 0.402654149 0.513329960 0.108696877
#>   [2,] 0.410693559 0.141769885 0.787817827 0.006500182 0.026438628 0.977739160
#>   [3,] 0.031466635 0.078718231 0.768249805 0.850392241 0.767466367 0.051661229
#>   [4,] 0.139289037 0.235001856 0.571775141 0.075654456 0.690776185 0.376860922
#>   [5,] 0.043225337 0.473602177 0.796429249 0.686278469 0.277182099 0.825344038
#>   [6,] 0.541450880 0.149062061 0.649435041 0.693991740 0.483847484 0.701358107
#>   [7,] 0.804297888 0.364338255 0.539137510 0.243328579 0.144458811 0.316207919
#>   [8,] 0.919484263 0.005697682 0.223786583 0.546938408 0.670017587 0.718059951
#>   [9,] 0.250728327 0.993247639 0.397984891 0.399510205 0.618896582 0.025442871
#>  [10,] 0.907571222 0.220336673 0.290059658 0.456858369 0.904414072 0.009690499
#>  [11,] 0.349457752 0.333680904 0.116905655 0.613235970 0.942974040 0.349506491
#>  [12,] 0.537413983 0.725525791 0.086629451 0.150518012 0.157165192 0.477145310
#>  [13,] 0.737063563 0.819859477 0.146866767 0.259164873 0.246693158 0.627589751
#>  [14,] 0.290884012 0.417723272 0.231230645 0.681784781 0.288159012 0.203849256
#>  [15,] 0.160384690 0.541995476 0.191287199 0.162567490 0.005485648 0.389565845
#>  [16,] 0.577349542 0.027588344 0.326394256 0.867198404 0.032309966 0.418046734
#>  [17,] 0.259170652 0.278588474 0.599024196 0.614619324 0.111773070 0.433453151
#>  [18,] 0.511335388 0.972919439 0.170091753 0.731547860 0.590604357 0.891433475
#>  [19,] 0.482714009 0.638631678 0.533267865 0.462121789 0.056429218 0.329247452
#>  [20,] 0.289148305 0.912464221 0.661843865 0.272505555 0.745270712 0.307622275
#>  [21,] 0.092544849 0.404640990 0.566313418 0.190793676 0.923130517 0.311147721
#>  [22,] 0.295560639 0.845001844 0.651135013 0.039814066 0.637823154 0.424831189
#>  [23,] 0.055727404 0.504693865 0.603942817 0.955560014 0.431130010 0.239310450
#>  [24,] 0.995092802 0.028319959 0.134100311 0.269587249 0.243254295 0.344532467
#>  [25,] 0.980426212 0.668513181 0.699965235 0.527050356 0.541846638 0.355681637
#>  [26,] 0.579623078 0.111907076 0.407037264 0.186452534 0.075655303 0.791223413
#>  [27,] 0.842065119 0.547666735 0.256610427 0.202209257 0.099172842 0.935521836
#>  [28,] 0.643489873 0.284608032 0.727608676 0.663979909 0.199200688 0.544091445
#>  [29,] 0.559368738 0.260679788 0.467046066 0.258613658 0.028652255 0.926350324
#>  [30,] 0.385571765 0.220144983 0.142050270 0.308194516 0.361999650 0.449998267
#>  [31,] 0.387070375 0.760740802 0.272220036 0.874574474 0.787015017 0.810402758
#>  [32,] 0.235415933 0.272880020 0.555234163 0.989568571 0.819821968 0.828716257
#>  [33,] 0.712535433 0.077349979 0.366052995 0.072075998 0.449988666 0.819770810
#>  [34,] 0.615262844 0.022543827 0.976080917 0.462774031 0.340412029 0.189685963
#>  [35,] 0.403406707 0.859161389 0.328070697 0.127863202 0.482436555 0.799225194
#>  [36,] 0.258537860 0.478214060 0.512818700 0.465370459 0.098102531 0.585084590
#>  [37,] 0.046793783 0.609890997 0.259384452 0.896023224 0.181922966 0.081666230
#>  [38,] 0.474801444 0.510414007 0.523892158 0.572144479 0.266449734 0.304159877
#>  [39,] 0.025472944 0.395811494 0.322014237 0.800299945 0.366658207 0.346785525
#>  [40,] 0.255168821 0.537315284 0.169593521 0.922191335 0.537305671 0.581217286
#>  [41,] 0.854805695 0.438797273 0.973417659 0.846566127 0.337148180 0.619988889
#>  [42,] 0.191822105 0.720291457 0.670376878 0.775789222 0.713104505 0.016332273
#>  [43,] 0.353357093 0.031101877 0.266653169 0.490891914 0.001002802 0.519009865
#>  [44,] 0.016159547 0.810021117 0.277189713 0.219021406 0.296395551 0.189877041
#>  [45,] 0.595484344 0.143388595 0.636277534 0.838474641 0.442690343 0.020126043
#>  [46,] 0.519982670 0.399730126 0.678278664 0.942553181 0.062797634 0.626396426
#>  [47,] 0.774889001 0.503619910 0.291910523 0.520133443 0.313984717 0.859582365
#>  [48,] 0.951050141 0.074852764 0.139031575 0.131735145 0.494498092 0.316031892
#>  [49,] 0.223979242 0.300682204 0.042987005 0.764386855 0.799322373 0.566930006
#>  [50,] 0.506514526 0.304581977 0.788620976 0.347267838 0.717229215 0.197590932
#>                 X7           X8         X9        X10         X11         X12
#>   [1,] 0.051250849 0.5957425488 0.26509781 0.52452295 0.349708573 0.369411233
#>   [2,] 0.300626530 0.7314908972 0.42865124 0.59275715 0.038592480 0.015125651
#>   [3,] 0.484540881 0.8368777081 0.50204370 0.68557182 0.291898028 0.160063417
#>   [4,] 0.033936431 0.2916696800 0.57977329 0.38991155 0.899261829 0.628670894
#>   [5,] 0.914297574 0.4375823478 0.48816427 0.28332536 0.527337488 0.670555037
#>   [6,] 0.862538750 0.7453307656 0.82317723 0.48634809 0.589780780 0.263520854
#>   [7,] 0.039838086 0.3891682967 0.56132528 0.25058347 0.319265400 0.191405967
#>   [8,] 0.632053483 0.8715615608 0.35344362 0.36664116 0.588647229 0.732180534
#>   [9,] 0.098899216 0.5133802902 0.53093717 0.32341713 0.721693232 0.351439335
#>  [10,] 0.843670109 0.8088676850 0.46304995 0.57522039 0.974604385 0.662468216
#>  [11,] 0.401227074 0.1465984324 0.36181369 0.04591417 0.605644482 0.728334280
#>  [12,] 0.413576920 0.2002980963 0.11702495 0.50803857 0.248692958 0.900331389
#>  [13,] 0.936378039 0.1515782760 0.03293680 0.20396032 0.165225696 0.158129339
#>  [14,] 0.742613880 0.5797804033 0.02511214 0.70524525 0.780200699 0.084834663
#>  [15,] 0.473772413 0.2834316902 0.13378354 0.70642343 0.115716574 0.067480715
#>  [16,] 0.769874048 0.0651748427 0.07687078 0.57244564 0.082224241 0.302947630
#>  [17,] 0.219289307 0.4066697375 0.76824991 0.90305939 0.328320395 0.155874907
#>  [18,] 0.954580692 0.6957603698 0.52033888 0.35510641 0.844091072 0.830850207
#>  [19,] 0.628902683 0.6296993257 0.45004974 0.74340204 0.910610619 0.340222195
#>  [20,] 0.825846039 0.2123513843 0.30363760 0.27271448 0.719536948 0.753632589
#>  [21,] 0.109928537 0.1169954059 0.74527695 0.08327421 0.040895763 0.399558151
#>  [22,] 0.542555873 0.2849802880 0.13738126 0.86776080 0.210936374 0.492976930
#>  [23,] 0.579532464 0.9074938260 0.39839066 0.28092556 0.636384573 0.253023881
#>  [24,] 0.580564987 0.8527146454 0.05161284 0.34899637 0.085034045 0.707789566
#>  [25,] 0.165809945 0.2054445802 0.38243044 0.48415766 0.782511678 0.476574886
#>  [26,] 0.905160652 0.0539713274 0.43349983 0.12481553 0.886873778 0.546778013
#>  [27,] 0.011966763 0.5814372306 0.83648525 0.89540412 0.627567684 0.507070857
#>  [28,] 0.906282802 0.8817938215 0.04599801 0.38385115 0.354849321 0.589575083
#>  [29,] 0.850448683 0.1529690197 0.53372766 0.33406307 0.466775220 0.499598382
#>  [30,] 0.028329807 0.9648464934 0.89060369 0.58887559 0.810480417 0.774576466
#>  [31,] 0.567252356 0.3970254660 0.24473394 0.20135943 0.658813146 0.575818556
#>  [32,] 0.386724862 0.0562249057 0.79948629 0.37569514 0.276053655 0.335148320
#>  [33,] 0.412717713 0.2698397739 0.98854733 0.62488803 0.323073623 0.844989780
#>  [34,] 0.577926494 0.0001215118 0.02389191 0.08748498 0.868277377 0.487440545
#>  [35,] 0.332108641 0.8786905240 0.01174771 0.32240835 0.490113497 0.376095920
#>  [36,] 0.527235373 0.4877088461 0.62995197 0.87579879 0.830073687 0.551435495
#>  [37,] 0.312400921 0.5261569594 0.45413302 0.28363945 0.107767334 0.078346761
#>  [38,] 0.804247237 0.6915988124 0.62992304 0.96041071 0.373100822 0.813407619
#>  [39,] 0.544797445 0.3376643339 0.47115453 0.49639428 0.448245066 0.120225705
#>  [40,] 0.968749521 0.0453832506 0.07148714 0.90779420 0.605233564 0.847833287
#>  [41,] 0.381697190 0.5008840268 0.21558394 0.88000774 0.667164954 0.857044800
#>  [42,] 0.437371198 0.3106835377 0.24776380 0.17796975 0.476180397 0.851258274
#>  [43,] 0.023380000 0.1705267651 0.09986972 0.51540259 0.309054144 0.948425551
#>  [44,] 0.938584714 0.1972948874 0.37033483 0.70721566 0.044888581 0.118115195
#>  [45,] 0.305506194 0.8346836441 0.69301682 0.42291535 0.536501576 0.923144253
#>  [46,] 0.871997487 0.7092315503 0.54217170 0.01331532 0.512236837 0.545280965
#>  [47,] 0.034045131 0.8871208834 0.65955701 0.49274382 0.276408578 0.638219755
#>  [48,] 0.165968426 0.6720145545 0.36518163 0.04846610 0.592745907 0.575564258
#>  [49,] 0.234500438 0.0995969628 0.08918600 0.43738711 0.084261131 0.395674204
#>  [50,] 0.495495090 0.3151181436 0.79227361 0.37390139 0.028167106 0.774721188
#>                X13         X14         X15         X16        X17        X18
#>   [1,] 0.024199279 0.939985754 0.050952427 0.920357616 0.81047731 0.35781537
#>   [2,] 0.739010666 0.551293217 0.166065088 0.481766801 0.22701586 0.25051520
#>   [3,] 0.476295705 0.065616481 0.793197343 0.306548994 0.71214655 0.24431060
#>   [4,] 0.568885727 0.826624048 0.666762468 0.158245282 0.42217234 0.18979738
#>   [5,] 0.247936813 0.602093439 0.140022141 0.026358857 0.61339972 0.48098631
#>   [6,] 0.528323212 0.239066765 0.690651701 0.035480533 0.45451557 0.43469831
#>   [7,] 0.511807900 0.468903573 0.022106569 0.524446410 0.79884815 0.26636901
#>   [8,] 0.760540257 0.512505026 0.670483330 0.479230838 0.46031622 0.43892228
#>   [9,] 0.277064109 0.824339381 0.657686269 0.560292587 0.90337852 0.08969454
#>  [10,] 0.214449165 0.328801775 0.211869506 0.008375495 0.57925904 0.16764799
#>  [11,] 0.283369033 0.440126390 0.764198587 0.826453380 0.46664201 0.69151006
#>  [12,] 0.102769196 0.402092807 0.787946705 0.644694323 0.27332759 0.45536543
#>  [13,] 0.098681611 0.106277970 0.872010108 0.445905291 0.71214762 0.07568710
#>  [14,] 0.102066732 0.707817795 0.161470991 0.178842996 0.15398451 0.29399577
#>  [15,] 0.570834562 0.961833559 0.300647212 0.055869623 0.14117071 0.74263045
#>  [16,] 0.274367876 0.657155989 0.454352858 0.081850790 0.72342938 0.65598308
#>  [17,] 0.344667414 0.877605306 0.383663955 0.154449057 0.72127114 0.59188395
#>  [18,] 0.052849809 0.626645339 0.293650825 0.073078520 0.55915568 0.38276311
#>  [19,] 0.502455293 0.605641056 0.080272312 0.735606654 0.09150990 0.08193191
#>  [20,] 0.012098661 0.969547066 0.765080196 0.497920701 0.40246831 0.43979384
#>  [21,] 0.267072157 0.847277067 0.075724479 0.294634596 0.41353206 0.17239624
#>  [22,] 0.274574167 0.737894434 0.700850532 0.768800075 0.36888853 0.42196598
#>  [23,] 0.741514030 0.372651792 0.839743035 0.100719568 0.92021685 0.78835271
#>  [24,] 0.621631034 0.003212775 0.693736281 0.599445805 0.29902050 0.36322365
#>  [25,] 0.462566932 0.680447986 0.009596745 0.132010981 0.23483914 0.96928034
#>  [26,] 0.252903595 0.032017939 0.580335994 0.924670432 0.43028939 0.44052671
#>  [27,] 0.782272017 0.417841700 0.337980897 0.156340748 0.83763796 0.77396504
#>  [28,] 0.096650909 0.561834418 0.677265212 0.165489307 0.86777754 0.76302474
#>  [29,] 0.869775673 0.595059473 0.709411990 0.332962039 0.13912051 0.25653965
#>  [30,] 0.696626219 0.388694002 0.118053270 0.682848772 0.77918472 0.83008716
#>  [31,] 0.386769633 0.083397886 0.883182328 0.158581917 0.75581160 0.66585469
#>  [32,] 0.452431357 0.143306390 0.855776970 0.736478803 0.22612597 0.81994216
#>  [33,] 0.580057597 0.961758373 0.937834050 0.676479256 0.26668475 0.76274462
#>  [34,] 0.437164479 0.315962560 0.932182220 0.034093186 0.83035259 0.58890245
#>  [35,] 0.320951395 0.586159811 0.326229965 0.258337923 0.27878669 0.93958369
#>  [36,] 0.607485648 0.908382888 0.690128237 0.022666451 0.36323780 0.04028590
#>  [37,] 0.073646204 0.314921908 0.160594769 0.155148273 0.45054320 0.70859337
#>  [38,] 0.798823850 0.819127199 0.511837869 0.296870015 0.03296280 0.03064708
#>  [39,] 0.974653284 0.273585572 0.151887508 0.172459827 0.03774186 0.45828322
#>  [40,] 0.989218190 0.413194772 0.349677839 0.321708640 0.19747541 0.19312492
#>  [41,] 0.138367370 0.709884819 0.423532716 0.911335027 0.97736241 0.84889880
#>  [42,] 0.335369900 0.754377073 0.780900159 0.738997408 0.72504024 0.09510751
#>  [43,] 0.357174248 0.609885308 0.355405470 0.045447003 0.43255322 0.89959562
#>  [44,] 0.813316548 0.252174418 0.080471691 0.523033540 0.27999045 0.65905891
#>  [45,] 0.217859757 0.066413892 0.249207186 0.820063872 0.24259218 0.64700904
#>  [46,] 0.324243617 0.479474704 0.025694568 0.878537262 0.81005983 0.71183388
#>  [47,] 0.134407841 0.334764203 0.371766501 0.558413671 0.12478627 0.66852502
#>  [48,] 0.156441789 0.057055550 0.365378057 0.410529766 0.31489179 0.72721408
#>  [49,] 0.134528509 0.643832332 0.384766091 0.043921584 0.34503892 0.06186440
#>  [50,] 0.023168908 0.515543546 0.708108566 0.869379519 0.86772011 0.21702012
#>               X19        X20
#>   [1,] 0.24140422 0.53508413
#>   [2,] 0.89406624 0.03375841
#>   [3,] 0.81890864 0.41248497
#>   [4,] 0.13613675 0.23577136
#>   [5,] 0.81310768 0.90595767
#>   [6,] 0.20197644 0.31753686
#>   [7,] 0.12728306 0.60115974
#>   [8,] 0.02500543 0.02539099
#>   [9,] 0.50583613 0.67545368
#>  [10,] 0.42238214 0.75117191
#>  [11,] 0.42531677 0.08183596
#>  [12,] 0.14836002 0.10786269
#>  [13,] 0.37882117 0.33051960
#>  [14,] 0.32298506 0.70729565
#>  [15,] 0.50596994 0.53955629
#>  [16,] 0.10602984 0.36939439
#>  [17,] 0.38537332 0.03845316
#>  [18,] 0.14617422 0.57676590
#>  [19,] 0.38904630 0.95277371
#>  [20,] 0.61764377 0.87168801
#>  [21,] 0.30391222 0.45166008
#>  [22,] 0.46369767 0.98355831
#>  [23,] 0.58995048 0.84712579
#>  [24,] 0.35267555 0.90788451
#>  [25,] 0.02913312 0.72115625
#>  [26,] 0.93257014 0.60812449
#>  [27,] 0.60271061 0.97367934
#>  [28,] 0.11767273 0.74099449
#>  [29,] 0.03549462 0.04948158
#>  [30,] 0.90268298 0.95058775
#>  [31,] 0.71057068 0.19273520
#>  [32,] 0.17420392 0.14196982
#>  [33,] 0.81003014 0.90945326
#>  [34,] 0.34139851 0.05540302
#>  [35,] 0.52808135 0.08698340
#>  [36,] 0.20118982 0.52030824
#>  [37,] 0.78352154 0.26771381
#>  [38,] 0.11162081 0.36257094
#>  [39,] 0.55803956 0.63267987
#>  [40,] 0.58148543 0.62640101
#>  [41,] 0.25102662 0.05927219
#>  [42,] 0.99443282 0.81840348
#>  [43,] 0.50212810 0.87570941
#>  [44,] 0.36755249 0.49213823
#>  [45,] 0.92125209 0.68708039
#>  [46,] 0.69740044 0.01465982
#>  [47,] 0.14216640 0.09797399
#>  [48,] 0.14057286 0.71894777
#>  [49,] 0.80188385 0.28952820
#>  [50,] 0.42337938 0.32651857
#>  [ reached getOption("max.print") -- omitted 50 rows ]
#> 
#> $y
#>   [1]  2.64141457  3.34198672 -0.99708005 11.57701632  2.05980232  0.87725068
#>   [7] 13.56115922  6.75263454  2.89234437  6.34970822 -1.22583843  5.51990846
#>  [13] 15.27338881 -0.10695644 14.52034121  5.27746011 -3.25336570 -0.08164529
#>  [19]  4.81438702  5.64072574  6.68143531  4.64535810 -4.61737702 11.95052404
#>  [25] -0.81581696 12.41343609 12.12549113 -4.51566961 15.83846154  3.69111986
#>  [31] 10.18844098  5.92020936  8.98576653  2.99608490 18.42487710 -4.35441745
#>  [37]  7.99801164  4.67816626  3.02564342  1.39517373  5.73465410  0.86785469
#>  [43]  6.13171992 17.18997401  2.94863113  2.01555631  1.55037684 16.77056844
#>  [49] -0.11702998 13.88029075  6.42392020  0.08355090  8.88906145 11.96028432
#>  [55] -0.38008844 -0.26299205  3.88787808 10.98340494 16.33013288 -0.22818879
#>  [61]  0.98704051 -3.44673097  8.20566834 -2.74137768  0.96055157  6.82822925
#>  [67]  2.72674730  6.78568034 10.50415134 10.24596515  3.54129355 -2.16198714
#>  [73]  5.08556058  5.88395022  9.01467603 -1.98489007 -1.32222991 15.73021556
#>  [79]  3.61229961  1.84877582  3.93934834  6.81528703  2.63548134  6.39349046
#>  [85]  3.42975024  1.45098754  4.43356537 -4.65730630  6.26751915  5.59972200
#>  [91]  8.87100591 10.41157676  3.73894043  6.10533276  3.38444628  3.59204636
#>  [97]  8.26104031 13.13957549  7.06421131  3.94976416
#> 
#> $e
#>   [1] -0.75707185 -0.58976189  0.49544908  0.84670712 -0.59139575 -0.08563390
#>   [7]  0.68861145 -0.52229842  0.63289130 -0.55385135 -0.13351657  0.04421288
#>  [13]  0.11910019 -0.69940746  0.59781626 -0.73439497  0.75952534 -0.88917111
#>  [19] -0.32326132 -0.48912581  0.13938597 -0.55932054 -0.54320582 -0.24559684
#>  [25]  0.36764859 -0.15537878 -0.49157144  0.35456267  0.52248952 -0.98525765
#>  [31] -0.27553513  0.35918773 -0.41589947  0.47015301  0.74891693  0.95889122
#>  [37] -0.35452898 -0.20560735 -0.21122506 -0.22574737  0.13435908 -0.09540532
#>  [43] -0.20195152  0.68823367  0.55917770  0.17996861  0.17241897  0.37793950
#>  [49]  0.19430209  0.93415348 -0.54292847 -0.20287084  0.58591979 -0.20967543
#>  [55] -0.74667100  0.44337950 -0.12011884  0.12018609 -0.15115786 -0.12499030
#>  [61] -0.77088480 -0.38944971  0.11352771 -0.40960795 -0.55034136 -0.54824575
#>  [67] -0.17980185 -0.44602451  0.30829878 -0.25990996 -0.73977685 -0.08474036
#>  [73] -0.14943082 -0.80337628 -0.49754434 -0.61223951 -0.51105481  0.68400578
#>  [79] -0.25709837 -0.90320971 -0.24561208 -0.94365159  0.11303964 -0.62108243
#>  [85] -0.09250825 -0.05721819 -0.06182550  0.04827889 -0.53050289 -0.97899863
#>  [91]  0.88219455  0.01178292 -0.03335050 -0.56237866 -0.63036433 -0.47674469
#>  [97]  0.35020749  0.55624489 -0.93339467 -0.59468344
#> 
#> $f1
#> function (x) 
#> 5 * x
#> <bytecode: 0x11dfc298>
#> <environment: 0x11dafe18>
#> 
#> $f2
#> function (x) 
#> 3 * (2 * x - 1)^2
#> <bytecode: 0x11dfc570>
#> <environment: 0x11dafe18>
#> 
#> $f3
#> function (x) 
#> 4 * sin(2 * pi * x)/(2 - sin(2 * pi * x))
#> <bytecode: 0x11dfbb48>
#> <environment: 0x11dafe18>
#> 
#> $f4
#> function (x) 
#> 6 * (0.1 * sin(2 * pi * x) + 0.2 * cos(2 * pi * x) + 0.3 * sin(2 * 
#>     pi * x)^2 + 0.4 * cos(2 * pi * x)^3 + 0.5 * sin(2 * pi * 
#>     x)^3)
#> <bytecode: 0x11dfb3f8>
#> <environment: 0x11dafe18>
#> 
#> $f3.inter
#> function (x, e) 
#> e * f3(x)
#> <bytecode: 0xf9c38d8>
#> <environment: 0x11dafe18>
#> 
#> $f4.inter
#> function (x, e) 
#> e * f4(x)
#> <bytecode: 0xf9c3d70>
#> <environment: 0x11dafe18>
#>